### 2020-07-10 #2464 @kyleniemeyer
Hi @fverdugo, it looks like you have placed your paper in a separate repository from the software itself. We strongly encourage authors to put the paper in the same repository as the software, and in fact a lot of the checks that our bot @whedon does relies on this. (Right now we have no connection to the software repository.)

Do you have a particular reason for putting the paper in a separate repo? We do support the whole publication process with the paper being in a non-default branch, so if you prefer to not include the paper in your main/master branch, you could do that.

cc @arfon
