### 2020-07-11 f9fd41a
Added JOSS paper [skip ci]
