+++ paper.bib
@@ -1,0 +1,234 @@
+@misc{flux_web,
+howpublished = {https://fluxml.ai/},
+title = {{The Flux Project's Webpage}},
+url = {https://fluxml.ai/},
+urldate = {2019-08-25}
+}
+@article{Kirby2006,
+abstract = {As a key step towards a complete automation of the finite element method, we present a new algorithm for automatic and efficient evaluation of multilinear variational forms. The algorithm has been implemented in the form of a compiler, the FEniCS Form Compiler (FFC). We present benchmark results for a series of standard variational forms, including the incompressible Navier-Stokes equations and linear elasticity. The speedup compared to the standard quadrature-based approach is impressive; in some cases the speedup is as large as a factor of 1000. {\textcopyright} 2006 ACM.},
+archivePrefix = {arXiv},
+arxivId = {1112.0402},
+author = {Kirby, Robert C and Logg, Anders},
+doi = {10.1145/1163641.1163644},
+eprint = {1112.0402},
+file = {:home/fverdugo/.local/share/data/Mendeley Ltd./Mendeley Desktop/Downloaded/Kirby - Unknown - A Compiler for Variational Forms.pdf:pdf},
+issn = {00983500},
+journal = {ACM Transactions on Mathematical Software},
+keywords = {Automation,Compiler,Finite element,Variational form},
+number = {3},
+pages = {417--444},
+title = {{A compiler for variational forms}},
+volume = {32},
+year = {2006}
+}
+@misc{Carlsson,
+author = {Carlsson, Kristoffer},
+title = {{JuaFEM git repository}},
+url = {https://github.com/KristofferC/JuAFEM.jl}
+}
+@article{Bangerth2007,
+abstract = {An overview of the software design and data abstraction decisions chosen for deal.II, a general purpose finite element library written in C++, is given. The library uses advanced object-oriented and data encapsulation techniques to break finite element implementations into smaller blocks that can be arranged to fit users requirements. Through this approach, deal.II supports a large number of different applications covering a wide range of scientific areas, programming methodologies, and application-specific algorithms, without imposing a rigid framework into which they have to fit. A judicious use of programming techniques allows us to avoid the computational costs frequently associated with abstract object-oriented class libraries. The paper presents a detailed description of the abstractions chosen for defining geometric information of meshes and the handling of degrees of freedom associated with finite element spaces, as well as of linear algebra, input/output capabilities and of interfaces to other software, such as visualization tools. Finally, some results obtained with applications built atop deal.II are shown to demonstrate the powerful capabilities of this toolbox. {\textcopyright} 2007 ACM.},
+author = {Bangerth, W. and Hartmann, R. and Kanschat, G.},
+doi = {10.1145/1268776.1268779},
+isbn = {0098-3500},
+issn = {00983500},
+journal = {ACM Transactions on Mathematical Software},
+keywords = {Object-orientation,Software design},
+month = {aug},
+number = {4},
+publisher = {ACM},
+title = {{Deal.II --A general-purpose object-oriented finite element library}},
+url = {http://portal.acm.org/citation.cfm?doid=1268776.1268779},
+volume = {33},
+year = {2007}
+}
+@book{Johnson2009,
+author = {Johnson, Claes},
+isbn = {048646900X},
+pages = {288},
+publisher = {Dover Publications},
+title = {{Numerical Solution of Partial Differential Equations by the Finite Element Method}},
+year = {2009}
+}
+@article{Carlsson_NLsolve_2019,
[…]
