### 2020-07-11 #2464 @fverdugo
Hi @kyleniemeyer,

I have moved the paper to the repo https://github.com/gridap/Gridap.jl branch `joss`. Hope it is now OK!

cc @santiagobadia
