### 2020-07-12 #2464 @Kevin-Mattheus-Moerman
@kyleniemeyer this looks very interesting to me!! If others (e.g. @labarba or @meg-simula ? ) cannot take this to edit I'd be interested. Alternatively I can join as a reviewer.
Also folks from the Julia FEM community might be good candidates to serve as reviewers: https://github.com/JuliaFEM

I just wanted to point out their pre-print as well https://arxiv.org/abs/1910.01412
