+++ src/ReferenceFEs/ReferenceFEInterfaces.jl
@@ -106,7 +106,12 @@ end
     get_face_own_dofs_permutations(reffe::ReferenceFE,conf::Conformity) -> Vector{Vector{Vector{Int}}}
 """
 function get_face_own_dofs_permutations(reffe::ReferenceFE,conf::Conformity)
-  @abstractmethod
+  face_own_dofs = get_face_own_dofs(reffe)
+  _trivial_face_own_dofs_permutations(face_own_dofs)
+end
+
+function _trivial_face_own_dofs_permutations(face_own_dofs)
+  [ [collect(Int,1:length(dofs)),]  for dofs in face_own_dofs  ]
 end
 
 """
@@ -325,16 +330,8 @@ end
 # Concrete implementation
 
 """
-    struct GenericRefFE{D,C<:Conformity} <: ReferenceFE{D}
-      ndofs::Int
-      polytope::Polytope{D}
-      prebasis::Field
-      dofs::Dof
-      conformity::C
-      face_own_dofs::Vector{Vector{Int}}
-      face_own_dofs_permutations::Vector{Vector{Vector{Int}}}
-      face_dofs::Vector{Vector{Int}}
-      shapefuns::Field
+    struct GenericRefFE{C<:Conformity,D} <: ReferenceFE{D}
+      # private fields
     end
 
 This type is a *materialization* of the `ReferenceFE` interface. That is, it is a 
@@ -346,14 +343,13 @@ Note that some fields in this `struct` are type unstable deliberately in order t
 type signature. Don't access them in computationally expensive functions,
 instead extract the required fields before and pass them to the computationally expensive function.
 """
-struct GenericRefFE{D,C<:Conformity} <: ReferenceFE{D}
+struct GenericRefFE{C<:Conformity,D} <: ReferenceFE{D}
   ndofs::Int
   polytope::Polytope{D}
   prebasis::Field
   dofs::Dof
   conformity::C
-  face_own_dofs::Vector{Vector{Int}}
-  face_own_dofs_permutations::Vector{Vector{Vector{Int}}}
+  metadata
   face_dofs::Vector{Vector{Int}}
   shapefuns::Field
   @doc """
@@ -363,8 +359,7 @@ struct GenericRefFE{D,C<:Conformity} <: ReferenceFE{D}
         prebasis::Field,
         dofs::Dof,
         conformity::Conformity,
-        face_own_dofs::Vector{Vector{Int}},
-        face_own_dofs_permutations::Vector{Vector{Vector{Int}}},
+        metadata,
         face_dofs::Vector{Vector{Int}},
         shapefuns::Field=compute_shapefuns(dofs,prebasis)) where D
 
@@ -376,19 +371,17 @@ struct GenericRefFE{D,C<:Conformity} <: ReferenceFE{D}
     prebasis::Field,
     dofs::Dof,
     conformity::Conformity,
-    face_own_dofs::Vector{Vector{Int}},
-    face_own_dofs_permutations::Vector{Vector{Vector{Int}}},
+    metadata,
     face_dofs::Vector{Vector{Int}},
     shapefuns::Field=compute_shapefuns(dofs,prebasis)) where D
 
-    new{D,typeof(conformity)}(
+    new{typeof(conformity),D}(
       ndofs,
       polytope,
       prebasis,
       dofs,
       conformity,
-      face_own_dofs,
-      face_own_dofs_permutations,
+      metadata,
       face_dofs,
       shapefuns)
   end
@@ -403,24 +396,6 @@ get_prebasis(reffe::GenericRefFE) = reffe.prebasis
 get_dof_basis(reffe::GenericRefFE) = reffe.dofs
 
[…]
