### 2020-07-13 fe07454
Fixing some issues related with previous commit
