### 2020-07-13 b898ede
Added typeof(metadata) as type parameter in GenericRefFE
