+++ src/Gridap.jl
@@ -48,14 +48,14 @@ include("Integration/Integration.jl")
 
 include("ReferenceFEs/ReferenceFEs.jl")
 
-include("Geometry/Geometry.jl")
-
-include("FESpaces/FESpaces.jl")
-
-include("MultiField/MultiField.jl")
-
-include("Visualization/Visualization.jl")
-
-include("Exports.jl")
+#include("Geometry/Geometry.jl")
+#
+#include("FESpaces/FESpaces.jl")
+#
+#include("MultiField/MultiField.jl")
+#
+#include("Visualization/Visualization.jl")
+#
+#include("Exports.jl")
 
 end # module
