+++ src/ReferenceFEs/GeometricLagrangianRefFEs.jl
@@ -1,0 +1,766 @@
+struct GradConformity <: Conformity end
+const H1Conformity = GradConformity
+
+
+function get_face_own_nodes(reffe::GenericLagrangianRefFE{GradConformity},conf::GradConformity)
+  p = get_polytope(reffe)
+  orders = get_orders(reffe)
+  nodes, face_own_nodes = compute_nodes(p,orders)
+  face_own_nodes
+end
+
+function get_own_nodes_permutations(reffe::GenericLagrangianRefFE{GradConformity},conf::GradConformity)
+  p = get_polytope(reffe)
+  face_own_nodes = get_face_own_nodes(reffe,conf)
+  dofs = get_dof_basis(reffe)
+  interior_nodes = dofs.nodes[face_own_nodes[end]]
+  compute_own_nodes_permutations(p,interior_nodes)
+end
+
+function get_face_own_nodes_permutations(reffe::GenericLagrangianRefFE{GradConformity},conf::GradConformity)
+  own_nodes_permutations = get_own_nodes_permutations(reffe)
+  reffaces = reffe.reffe.metadata
+  _reffaces = vcat(reffaces...)
+  face_own_nodes_permutations = map(get_own_nodes_permutations,_reffaces)
+
+  _compute_face_own_nodes_permutations(
+    num_nodes(reffe),
+    get_face_own_nodes(reffe),
+    face_own_nodes_permutations,
+    own_nodes_permutations)
+end
+
+function _compute_face_own_nodes_permutations(
+  nnodes,
+  face_own_nodes,
+  face_own_nodes_permutations,
+  own_nodes_permutations)
+
+  if length(face_own_nodes_permutations) == 0
+    # Vertex degenerated case
+    return [own_nodes_permutations,]
+  else
+    if nnodes == length(face_own_nodes[end])
+      # 0-order degenerated case
+      _face_own_nodes_permutations = map( (x) -> fill(Int[],length(x)) , face_own_nodes_permutations )
+    else
+      # Standard case
+      _face_own_nodes_permutations = copy(face_own_nodes_permutations)
+    end
+    push!(_face_own_nodes_permutations,own_nodes_permutations)
+    return _face_own_nodes_permutations
+  end
+end
+
+# API particular to LagrangianRefFE{GradConformity}
+
+"""
+    ReferenceFE{N}(reffe::GenericLagrangianRefFE{GradConformity},iface::Integer) where N
+"""
+function ReferenceFE{N}(reffe::GenericLagrangianRefFE{GradConformity},iface::Integer) where N
+  reffaces = reffe.reffe.metadata
+  reffaces[N+1][iface]
+end
+
+function ReferenceFE{D}(reffe::GenericLagrangianRefFE{GradConformity,M,D} where M,iface::Integer) where D
+  @assert iface==1 "Only one D-face"
+  reffe
+end
+
+"""
+    get_reffaces(
+      ::Type{ReferenceFE{d}},
+      reffe::GenericLagrangianRefFE{GradConformity}) where d -> Vector{GenericLagrangianRefFE{GradConformity,M,d}}
+"""
+function get_reffaces(::Type{ReferenceFE{d}},reffe::GenericLagrangianRefFE{GradConformity}) where d
+  ftype_to_reffe, _ = _compute_reffes_and_face_types(reffe,Val{d}())
+  ftype_to_reffe
+  [reffe for reffe in ftype_to_reffe]
+end
+
+"""
+    get_face_type(reffe::GenericLagrangianRefFE{GradConformity}, d::Integer) -> Vector{Int}
+"""
+function get_face_type(reffe::GenericLagrangianRefFE{GradConformity}, d::Integer)
[…]
