+++ test/ReferenceFEsTests/LagrangianRefFEsTests.jl
@@ -1,189 +1,58 @@
 module LagrangianRefFEsTests
 
 using Test
-using Gridap.Io
-using Gridap.TensorValues
-using Gridap.Fields
-using Gridap.Polynomials
 using Gridap.ReferenceFEs
-using JSON
+using Gridap.Polynomials
+using Gridap.Fields
 
-orders = (2,3)
-b = MonomialBasis(Float64,QUAD,orders)
-r = [(0,0), (1,0), (2,0), (0,1), (1,1), (2,1), (0,2), (1,2), (2,2), (0,3), (1,3), (2,3)]
-@test get_exponents(b) == r
+D = 2
+T = Float64
+order = 1
+prebasis = MonomialBasis{D}(T,order)
 
-orders = (1,1,2)
-b = MonomialBasis(Float64,WEDGE,orders)
-r = [(0,0,0), (1,0,0), (0,1,0), (0,0,1), (1,0,1), (0,1,1), (0,0,2), (1,0,2), (0,1,2)]
-@test get_exponents(b) == r
+polytope = QUAD
+x = get_vertex_coordinates(polytope)
+dofs = LagrangianDofBasis(T,x)
 
-orders = (1,1,1)
-b = MonomialBasis(Float64,PYRAMID,orders)
-r = [(0,0,0), (1,0,0), (0,1,0), (1,1, 0), (0,0,1)]
-@test get_exponents(b) == r
+ndofs = 4
 
-orders = (1,1,1)
-b = MonomialBasis(Float64,TET,orders)
-r = [(0,0,0), (1,0,0), (0,1,0), (0,0,1)]
-@test get_exponents(b) == r
+face_own_dofs = [[1],[2],[3],[4],Int[],Int[],Int[],Int[],Int[]]
+a = [1]
+b = Int[]
+face_own_dofs_permutations = [[a],[a],[a],[a],[b,b],[b,b],[b,b],[b,b],fill(b,8)]
+face_dofs = [[1],[2],[3],[4],[1,2],[3,4],[1,3],[2,4],[1,2,3,4]]
+metadata = nothing
 
-orders = (2,2)
-extrusion = Tuple(QUAD.extrusion)
+struct MockConformity <: Conformity end
 
-dofs = LagrangianDofBasis(VectorValue{3,Float64},TET,1)
-@test dofs.nodes == Point{3,Float64}[(0,0,0), (1,0,0), (0,1,0), (0,0,1)]
-@test dofs.node_and_comp_to_dof == VectorValue{3,Int}[(1,5,9), (2,6,10), (3,7,11), (4,8,12)]
+reffe = GenericRefFE(
+  ndofs, polytope, prebasis, dofs,
+  MockConformity(), metadata, face_dofs)
 
-dofs = LagrangianDofBasis(Float64,WEDGE,(2,2,2))
-r = Point{3,Float64}[
-  (0.0,0.0,0.0),(1.0,0.0,0.0),(0.0,1.0,0.0),
-  (0.0,0.0,1.0),(1.0,0.0,1.0),(0.0,1.0,1.0),
-  (0.5,0.0,0.0),(0.5,0.0,1.0),(0.0,0.5,0.0),
-  (0.5,0.5,0.0),(0.0,0.5,1.0),(0.5,0.5,1.0),
-  (0.0,0.0,0.5),(1.0,0.0,0.5),(0.0,1.0,0.5),
-  (0.5,0.0,0.5),(0.0,0.5,0.5),(0.5,0.5,0.5)]
-@test dofs.nodes == r
+node_coordinates = Point{2,Float64}[(0,0),(1,0),(0,1),(1,1)]
+node_and_comp_to_dof = [1,2,3,4]
+face_own_nodes = face_own_dofs
[…]
