+++ test/ReferenceFEsTests/runtests.jl
@@ -14,9 +14,9 @@ using Test
 
 @testset "ReferenceFEInterfaces" begin include("ReferenceFEInterfacesTests.jl") end
 
-@testset "NodalReferenceFEs" begin include("NodalReferenceFEsTests.jl") end
-
 @testset "LagrangianRefFEs" begin include("LagrangianRefFEsTests.jl") end
+
+@testset "GeometricLagrangianRefFEs" begin include("GeometricLagrangianRefFEsTests.jl") end
 
 @testset "SerendipityRefFEs" begin include("SerendipityRefFEsTests.jl") end
 
