+++ test/ReferenceFEsTests/CDLagrangianRefFEsTests.jl
@@ -7,11 +7,11 @@ using Test
 
 T = Float64
 reffe = CDLagrangianRefFE(T,SEGMENT,(2,),(DISC,))
-test_reference_fe(reffe)
+test_lagrangian_reference_fe(reffe)
 
 T = Float64
 reffe = CDLagrangianRefFE(T,QUAD,1,(CONT,DISC))
-test_reference_fe(reffe)
+test_lagrangian_reference_fe(reffe)
 
 conf = CDConformity( (CONT,DISC) )
 @test conf == get_default_conformity(reffe)
@@ -20,19 +20,22 @@ conf = CDConformity( (CONT,DISC) )
 
 T = VectorValue{2,Float64}
 reffe = CDLagrangianRefFE(T,QUAD,(2,0),(CONT,DISC))
-test_reference_fe(reffe)
+test_lagrangian_reference_fe(reffe)
+
+@test get_face_own_dofs(reffe,L2Conformity()) == [[], [], [], [], [], [], [], [], [1, 2, 3, 4, 5, 6]]
+@test get_face_own_nodes(reffe,L2Conformity()) == [[], [], [], [], [], [], [], [], [1, 2, 3]]
 
 T = VectorValue{3,Float64}
 reffe = CDLagrangianRefFE(T,HEX,(2,2,2),(CONT,CONT,DISC))
-test_reference_fe(reffe)
+test_lagrangian_reference_fe(reffe)
 
 T = VectorValue{3,Float64}
 reffe = CDLagrangianRefFE(T,HEX,2,(CONT,CONT,DISC))
-test_reference_fe(reffe)
+test_lagrangian_reference_fe(reffe)
 
 T = VectorValue{2,Float64}
 reffe = LagrangianRefFE(T,QUAD,(2,0))
-test_reference_fe(reffe)
+test_lagrangian_reference_fe(reffe)
 
 reffe = LagrangianRefFE(T,QUAD,(2,2))
 
