### 2020-07-13 082c40d
Misc fixes
