+++ test/FESpacesTests/FESpaceFactoriesTests.jl
@@ -13,11 +13,6 @@ partition = (3,3)
 model = CartesianDiscreteModel(domain,partition)
 order = 3
 
-T = VectorValue{2,Float64}
-reffe = QDiscRefFE(T,QUAD,2)
-V = FESpace(model=model,reffe=reffe)
-@test isa(V,UnconstrainedFESpace)
-
 V = FESpace(model=model,reffe=QUAD4,conformity=GradConformity())
 v1 = FEFunction(V,rand(num_free_dofs(V)))
 @test isa(V,UnconstrainedFESpace)
