### 2020-07-14 29d41bd
Update NEWS.md [skip ci]
