+++ NEWS.md
@@ -7,13 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.13.0] - Unreleased 
 
 ### Changed
-  - The meaning of `inward/outward` has slightly changed for `SkeletonCellBasis` objects. Now, by accessing to these properties a `ReducedSkeletonCellBasis` is returned, which allows to use the result in a more flexible way (in particular, the result can be used in a similar way than the result of `jump` or `mean`).Since PR [#317](https://github.com/gridap/Gridap.jl/pull/317).
+  - The meaning of `inward/outward` has slightly changed for `SkeletonCellBasis` objects. Now, by accessing to these properties a `ReducedSkeletonCellBasis` is returned, which allows to use the result in a more flexible way (in particular, the result can be used in a similar way than the result of `jump` or `mean`). Since PR [#317](https://github.com/gridap/Gridap.jl/pull/317).
   - Major refactoring in `ReferenceFEs` module. Since PR [#319](https://github.com/gridap/Gridap.jl/pull/319). In particular:
-    - `NodalReferenceFE` has been partially replaced with a new abstract type `LagrangianRefFE`.
-    - `GenericNodalCartesianRefFE` has been partially replaced by `GenericLagrangianRefFE`.
-    - `DiscRefFE` partially replaced by `PDiscRefFE`.
+    - `NodalReferenceFE` has been replaced by a new abstract type `LagrangianRefFE`.
+    - `GenericNodalCartesianRefFE` has been replaced by `GenericLagrangianRefFE`.
+    - `DiscRefFE` replaced by `PDiscRefFE`.
 
-## Removed
+### Removed
   - Removals associated with the `ReferenceFEs` refactoring in PR [#319](https://github.com/gridap/Gridap.jl/pull/319):
     - Removed `QDiscRefFE` constructor. Use a standard `LagrangianRefFE` and `L2Conformity` instead.
     - Removed fields `face_own_dofs` and `face_own_dof_permutations` from `GenericRefFE`.
