### 2020-07-14 f4c46b2
Update NEWS.md [skip ci]
