### 2020-07-14 e9321cc
Update NEWS.md [skip ci]
