### 2020-07-18 ccd7565
Minor fix to let Vtk.jl support changes in Vtk 1.7.X versus 1.6.X
