+++ Project.toml
@@ -38,7 +38,7 @@ LineSearches = "7.0.1"
 NLsolve = "4.3.0"
 QuadGK = "2.3.1"
 StaticArrays = "0.12.1"
-WriteVTK = "~1.6"
+WriteVTK = "1.7"
 julia = "1.0"
 
 [extras]
