### 2020-07-22 #2464 @Kevin-Mattheus-Moerman
@fverdugo can you comment on this pre-print https://arxiv.org/abs/1910.01412 ? We do not allow duplicate publication at JOSS. Can you please clarify the difference between this pre-print (has it been published/submitted anywhere else yet?) and the JOSS paper?
