### 2020-07-22 #2464 @fverdugo
The preprint in arxiv is a completely different and separate thing.  It is just a paper-like version of our tutorials in https://github.com/gridap/Tutorials. Not under review nor published anywhere.

If you open the two PDFs, you would clearly see the differences. E.g.,

 https://arxiv.org/abs/1910.01412 has 31 pages.

[Our JOSS submission](https://github.com/openjournals/joss-papers/blob/joss.02464/joss.02464/10.21105.joss.02464.pdf) has 3 pages.
