### 2020-07-22 0813bcf
Custom implementation of testitem for LocalToGlobalArray
