### 2020-07-22 #2464 @PetrKryslUCSD
Sure, I'll be happy to help.
Petr Krysl
Prof. and Vice chair for undergraduate education
Structural engineering department
University of California, San Diego

On Wed, Jul 22, 2020 at 1:40 AM Kevin Mattheus Moerman <
notifications@github.com> wrote:

> @PetrKryslUCSD
> <https://mailtrack.io/trace/link/f64d1255a92497d78300cc2d1815cf684a18186c?url=https%3A%2F%2Furldefense.com%2Fv3%2F__https%3A%2F%2Fgithub.com%2FPetrKryslUCSD__%3B!!Mih3wA!SoLQZEam3hZK0IMsDoJ2N1ce1leiSzg3sI78YnPyt9b7akEy2IgKhtIi4Wd6d2A%24&userId=1197778&signature=7ef2218e2a99b162>
> would you be interested in reviewing this submission for JOSS
> <https://mailtrack.io/trace/link/149944ff631daedea69f551dce18a0295e2b9d61?url=https%3A%2F%2Furldefense.com%2Fv3%2F__https%3A%2F%2Fjoss.theoj.org%2F__%3B!!Mih3wA!SoLQZEam3hZK0IMsDoJ2N1ce1leiSzg3sI78YnPyt9b7akEy2IgKhtIiJs1AZq4%24&userId=1197778&signature=fe0ab5ae01de3097>
> ?
>
> *Gridap: An extensible Finite Element toolbox in Julia*
> *Paper*:
> https://github.com/openjournals/joss-papers/blob/joss.02464/joss.02464/10.21105.joss.02464.pdf
> <https://mailtrack.io/trace/link/2985170d23800ce3654aeb6e5e10749a38dd2aae?url=https%3A%2F%2Furldefense.com%2Fv3%2F__https%3A%2F%2Fgithub.com%2Fopenjournals%2Fjoss-papers%2Fblob%2Fjoss.02464%2Fjoss.02464%2F10.21105.joss.02464.pdf__%3B!!Mih3wA!SoLQZEam3hZK0IMsDoJ2N1ce1leiSzg3sI78YnPyt9b7akEy2IgKhtIigOAvd4s%24&userId=1197778&signature=a58d89c3dff86a3a>
> *Software*: https://github.com/gridap/Gridap.jl
> <https://mailtrack.io/trace/link/53d107c0f316426aa31afe8ec22e0e6e64470020?url=https%3A%2F%2Furldefense.com%2Fv3%2F__https%3A%2F%2Fgithub.com%2Fgridap%2FGridap.jl__%3B!!Mih3wA!SoLQZEam3hZK0IMsDoJ2N1ce1leiSzg3sI78YnPyt9b7akEy2IgKhtIi_x6YOBc%24&userId=1197778&signature=22014b6d17429a24>
>
> The review process
> <https://mailtrack.io/trace/link/8062f875e35a541ab06892c34766a4935bdbe302?url=https%3A%2F%2Furldefense.com%2Fv3%2F__https%3A%2F%2Fjoss.readthedocs.io%2Fen%2Flatest%2Freviewer_guidelines.html__%3B!!Mih3wA!SoLQZEam3hZK0IMsDoJ2N1ce1leiSzg3sI78YnPyt9b7akEy2IgKhtIigm3_HRI%24&userId=1197778&signature=2753ec6c1a386b26>
> takes place on GitHub and focuses on the software and a short paper.
>
> Let me know if you are interested and I'll add you as reviewer.
>
> Thanks!
>
