### 2020-07-22 452fa2c
Supporting degenerated case in integration kernel (IntKernel)
