+++ src/Fields/Fields.jl
@@ -76,6 +76,7 @@ export function_field
 import Gridap.Arrays: kernel_cache
 import Gridap.Arrays: apply_kernel!
 import Gridap.Arrays: kernel_return_type
+import Gridap.Arrays: kernel_testitem!
 import Gridap.TensorValues: outer
 import Gridap.TensorValues: inner
 import Gridap.TensorValues: symmetric_part
