### 2020-07-22 117ee87
Stressing a FESource defined on an empty boundary triangulation
