### 2020-07-23 #2464 @Kevin-Mattheus-Moerman
@maikkirapo @ahojukka5 @KristofferC @burcugrbz would you be interested in reviewing this submission for [JOSS](https://joss.theoj.org/)?

_Gridap: An extensible Finite Element toolbox in Julia_
**Paper**: https://github.com/openjournals/joss-papers/blob/joss.02464/joss.02464/10.21105.joss.02464.pdf
**Software**: https://github.com/gridap/Gridap.jl

The [review process](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) takes place on GitHub and focuses on the software and a short paper.

Let me know if you are interested and I'll add you as reviewer.

Thanks!
