### 2020-07-24 3eaad76
Fixing some corener cases for 0-length arrays
