### 2020-07-24 97b173f
Ironing some more corner cases for 0-length arrays
