+++ src/Arrays/Interface.jl
@@ -194,6 +194,9 @@ function test_array(
   @test t
   @test IndexStyle(a) == IndexStyle(b)
   @test isa(testitem(a),eltype(a))
+  if length(a) > 0
+    @test testitem(a) == first(a)
+  end
   true
 end
 
