### 2020-07-25 #2464 @TeroFrondelius
I am on vacation, but I can do the review in the beginning of August (after a week or so). Also, this would be my first JOSS review, thus possible some learning due to this.
