### 2020-07-25 #2464 @PetrKryslUCSD
I have to admit I can't make sense of the instructions.
Do the reviewers submit issues in a repository specifically for this paper?
Where is it? Or are the issues to be filed in the repository of the package
that is
being reviewed? Is there an example of such a review process?

Thanks,
Petr Krysl

On Wed, Jul 22, 2020 at 1:40 AM Kevin Mattheus Moerman <
notifications@github.com> wrote:

> @PetrKryslUCSD
> <https://mailtrack.io/trace/link/eb864f86335eb6422e9ddd44432b063ab2effced?url=https%3A%2F%2Furldefense.com%2Fv3%2F__https%3A%2F%2Fgithub.com%2FPetrKryslUCSD__%3B!!Mih3wA!SoLQZEam3hZK0IMsDoJ2N1ce1leiSzg3sI78YnPyt9b7akEy2IgKhtIi4Wd6d2A%24&userId=1197778&signature=7a55f931dd777ec6>
> would you be interested in reviewing this submission for JOSS
> <https://mailtrack.io/trace/link/fc47183d55431888d89318dd432a27ea666640a8?url=https%3A%2F%2Furldefense.com%2Fv3%2F__https%3A%2F%2Fjoss.theoj.org%2F__%3B!!Mih3wA!SoLQZEam3hZK0IMsDoJ2N1ce1leiSzg3sI78YnPyt9b7akEy2IgKhtIiJs1AZq4%24&userId=1197778&signature=4cd565402ca1c5df>
> ?
>
> *Gridap: An extensible Finite Element toolbox in Julia*
> *Paper*:
> https://github.com/openjournals/joss-papers/blob/joss.02464/joss.02464/10.21105.joss.02464.pdf
> <https://mailtrack.io/trace/link/f76c1c6d1891c0ac75e34d0e32fdb1e9f0d52828?url=https%3A%2F%2Furldefense.com%2Fv3%2F__https%3A%2F%2Fgithub.com%2Fopenjournals%2Fjoss-papers%2Fblob%2Fjoss.02464%2Fjoss.02464%2F10.21105.joss.02464.pdf__%3B!!Mih3wA!SoLQZEam3hZK0IMsDoJ2N1ce1leiSzg3sI78YnPyt9b7akEy2IgKhtIigOAvd4s%24&userId=1197778&signature=14bcee207227e669>
> *Software*: https://github.com/gridap/Gridap.jl
> <https://mailtrack.io/trace/link/6a2854227bcf1aff45876868bd6a16f2a45606ad?url=https%3A%2F%2Furldefense.com%2Fv3%2F__https%3A%2F%2Fgithub.com%2Fgridap%2FGridap.jl__%3B!!Mih3wA!SoLQZEam3hZK0IMsDoJ2N1ce1leiSzg3sI78YnPyt9b7akEy2IgKhtIi_x6YOBc%24&userId=1197778&signature=171d43befaa7540f>
>
> The review process
> <https://mailtrack.io/trace/link/b84bd55646fa44bd7d56049ca76fd2bc3363eae6?url=https%3A%2F%2Furldefense.com%2Fv3%2F__https%3A%2F%2Fjoss.readthedocs.io%2Fen%2Flatest%2Freviewer_guidelines.html__%3B!!Mih3wA!SoLQZEam3hZK0IMsDoJ2N1ce1leiSzg3sI78YnPyt9b7akEy2IgKhtIigm3_HRI%24&userId=1197778&signature=6a5f559db409d763>
> takes place on GitHub and focuses on the software and a short paper.
>
> Let me know if you are interested and I'll add you as reviewer.
>
> Thanks!
>
