### 2020-07-25 #2464 @PetrKryslUCSD
I guess what confuses me is the variety of conversations here. This does not seem to be focused on the review itself.
But looking at the other issues, perhaps this is how it is supposed to look. Meaning reviewers submitting their comments here. Am I right?
