### 2020-07-25 #2464 @Kevin-Mattheus-Moerman
> I guess what confuses me is the variety of conversations here. This does not seem to be focused on the review itself.
> But looking at the other issues, perhaps this is how it is supposed to look. Meaning reviewers submitting their comments here. Am I right?

Apologies for any confusion, let me clarify. This is a prereview issue where we assign editors/reviewers. The actual review will take place in a dedicated review issue. I'm hoping to recruit one more reviewer.
In the review issue there will be a checkbox list to guide you through the process. You can leave comments to the authors there and also link to issues on their main repository for larger issues.
