### 2020-07-25 #2464 @Kevin-Mattheus-Moerman
@arfon @openjournals/dev it seems whedon is stuck. It won't assign a reviewer. Can you help?
