### 2020-07-26 #2464 @arfon
> @arfon @openjournals/dev it seems whedon is stuck. It won't assign a reviewer. Can you help?

Not sure what was going on here sorry. Minor hiccup...
