### 2020-07-26 #2520 @Kevin-Mattheus-Moerman
@PetrKryslUCSD @TeroFrondelius this is where the review of the [paper](https://github.com/openjournals/joss-papers/blob/joss.02520/joss.02520/10.21105.joss.02520.pdf) and the [software](https://github.com/gridap/Gridap.jl) takes place. You each have a set of checkboxes at the top of this issue to guide you through the process. You can leave small comments here but for larger items we encourage you to open dedicated issues on the software project's repository and to link to them here.

FYI See the above instructions :point_up: to turn off notifications from the other review issues.

Let me know if you have any questions.

Our review criteria: https://joss.readthedocs.io/en/latest/review_criteria.html
