### 2020-07-26 gridap/Gridap.jl#333 @TeroFrondelius
joss-paper: fix JUMP citation
Here https://github.com/jump-dev/JuMP.jl#citing-jump it says:
>Citing JuMP
If you find JuMP useful in your work, we kindly request that you cite the following paper (pdf):

>@article{DunningHuchetteLubin2017,
author = {Iain Dunning and Joey Huchette and Miles Lubin},
title = {JuMP: A Modeling Language for Mathematical Optimization},
journal = {SIAM Review},
volume = {59},
number = {2},
pages = {295-320},
year = {2017},
doi = {10.1137/15M1020575},
}

Xref https://github.com/openjournals/joss-reviews/issues/2520
