### 2020-07-26 #2464 @KristofferC
FWIW, if still needed, I can probably look at it as well.
