### 2020-07-26 gridap/Gridap.jl#334 @PetrKryslUCSD
joss-paper: Software design
Starting on page 1, in the sentence beginning with "In contrast...": The paper introduces some concepts which are in my opinion described at a level that does not/cannot provide any real insight.

For instance, when I hear "cell -wise" I can certainly form a mental picture, but the detail supplied in the paper is not sufficiently granular to allow me to judge the agreement or disagreement of that mental picture with what is designed and implemented in the repository.

What is a "cell -wise array" and a "cell -wise field"? Again, there is a hint that this might be interesting, but not enough detail to make sense of it really.

Similarly "laziness".  The concept is introduced, but the reader is not told enough to make sense of it.

Next, "lambda variational form" is mentioned. What is it, how is it used?

Finally, "state diagram" is mentioned. What is it? Why should the reader care? Not enough information is provided.

So, I think the authors need to decide whether or not this is something that can be described in a compact space, and if it cannot, is it worthwhile introducing it in the first place?
