### 2020-07-26 gridap/Gridap.jl#335 @PetrKryslUCSD
joss-paper: Missing operators
The weak form definition is incomplete: the gradient operators and the dot product operator are missing. Perhaps the font did not include these symbols?
