### 2020-07-26 gridap/Gridap.jl#336 @PetrKryslUCSD
joss-paper: Build tutorials not described
The readme does not describe how to get the tutorial notebooks opened from a cloned  repository. In particular, when the repository is cloned, there are no notebooks. They need to be built. The build step is not documented. It appears to be activate/instantiate/build.
Then one can open a notebook with `IJulia.notebook()`.
