### 2020-07-26 #2520 @PetrKryslUCSD
Filed three issues: https://github.com/gridap/Gridap.jl/issues/334, https://github.com/gridap/Gridap.jl/issues/335, https://github.com/gridap/Gridap.jl/issues/336.
