### 2020-07-27 gridap/Gridap.jl#336 @fverdugo
Related to  https://github.com/openjournals/joss-reviews/issues/2520
