### 2020-07-27 #2520 @fverdugo
I have created a new milestone

https://github.com/gridap/Gridap.jl/issues?q=is%3Aopen+is%3Aissue+milestone%3Ajoss_paper

to track the progress in addressing all the issues associated with this review.

You are welcome to use it, when opening a new issue in Gridap associated with this review.
