### 2020-07-27 593eea6
Added low level autodiff functions
