### 2020-07-27 52e5ac2
Added low-level autodiff for the FE layer
