### 2020-07-28 a5a56b2
Allowing an indirection array in autodiff_array_gradeint et al
