### 2020-07-28 7707152
Allowing indirection vector in FE autodiff layer
