### 2020-07-28 f886350
Added NonLinearFETermWithAutodiff
