### 2020-07-28 e9a1e1e
Added p-Laplacian with autodiff
