+++ test/FESpacesTests/FEAutodiffTests.jl
@@ -5,12 +5,12 @@ using Gridap.FESpaces
 using Gridap.Arrays
 using Gridap.Fields
 using Gridap.Geometry
-
+using Gridap.TensorValues 
 domain = (0,1,0,1)
 partition = (2,2)
 model = CartesianDiscreteModel(domain,partition)
 
-V = FESpace(model=model,valuetype=Float64,reffe=:Lagrangian,order=1,conformity=:H1)
+V = FESpace(model=model,valuetype=Float64,reffe=:Lagrangian,order=2,conformity=:H1)
 U = TrialFESpace(V)
 
 dv = get_cell_basis(V)
@@ -78,5 +78,27 @@ cell_j_Γ_auto = autodiff_cell_jacobian_from_residual(user_uh_to_cell_residual_�
 test_array(cell_r_Γ_auto,cell_r_Γ)
 test_array(cell_j_Γ_auto,cell_j_Γ)
 test_array(cell_h_Γ_auto,cell_h_Γ)
+
+const p = 3
+@law j(∇u) = norm(∇u)^(p-2) * ∇u
+@law dj(∇du,∇u) = (p-2)*norm(∇u)^(p-4)*inner(∇u,∇du)*∇u + norm(∇u)^(p-2)*∇du
+f(x) = 0
+
+res(u,v) = ∇(v)⋅j(∇(u)) - v*f
+jac(u,du,v) = ∇(v)⋅dj(∇(du),∇(u))
+
+function user_uh_to_cell_residual_2(uh)
+  cell_r = integrate(res(uh,dv),trian,quad)
+end
+
+function user_uh_to_cell_jacobian_2(uh)
+  cell_j = integrate(jac(uh,du,dv),trian,quad)
+end
+
+cell_j = user_uh_to_cell_jacobian_2(uh)
+
+cell_j_auto = autodiff_cell_jacobian_from_residual(user_uh_to_cell_residual_2,uh)
+
+test_array(cell_j_auto,cell_j)
 
 end # module
