### 2020-07-28 f03f892
Fixing JuMP citation [skip ci]
Cross-ref issue #333
