+++ paper.md
@@ -25,7 +25,7 @@ bibliography: paper.bib
 
 Gridap is a new Finite Element (FE) framework, exclusively written in the Julia programming language, for the numerical simulation of a wide range of mathematical models governed by partial differential equations (PDEs). The library provides a feature-rich set of discretization techniques, including continuous and discontinuous FE methods with Lagrangian, Raviart-Thomas, or Nédélec interpolations, and supports a wide range of problem types including linear, nonlinear, single-field, and multi-field PDEs. Gridap is designed to help application experts to easily simulate real-world problems, to help researchers improve productivity when developing new FE-related techniques, and also for its usage in numerical PDE courses.
 
-The main motivation behind Gridap is to find an improved balance between computational performance, user-experience and work-flow productivity, when working with FE libraries. Previous FE frameworks, e.g., FEniCS [@Alnaes2015] or Deal.II [@Bangerth2007] usually provide a high-level user front-end to facilitate the use of the library and a computational back-end to achieve performance. The user front-end is usually programmable in a interpreted language like Python, whereas the computational back-end is usually coded in a compiled language like C/C++ or Fortran. Users can benefit from the high-level front-end (i.e., for rapid prototyping) and simultaneously enjoy the performance of the compiled back-end. This approach reaches a compromise between performance and productivity when the back-end provides all the functionality required by the user. However, it does not satisfactorily address the needs of researchers on numerical methods willing to extend the library with new techniques or features. These extensions usually need to be done at the level of the computational back-end for performance reasons. Thus, the researcher is forced to develop new code in a compiled language like C/C++ instead of benefiting from the productivity of scripting languages like Python, incurring in serious productivity losses. In order to overcome this limitation, Gridap is fully implemented in the Julia programming language [@Bezanson2017]. Julia combines the performance of compiled languages with the productivity of interpreted ones by using type inference and just-in-time compilation to generate fast code. As a result, there is no need to use two different languages to write low-level performance code and high-level user interfaces. In addition, writing a FE library in Julia also allows one to leverage the feature-rich ecosystem of Julia libraries and exploit its excellent package manager. It permits a seamless coupling of Gridap with application-specific libraries, like optimization [@jump_web], approximation of ordinary differential equations [@Rackauckas2017], or data science [@flux_web], which can certainly boost the capabilities of a FE solver.
[…]
