### 2020-07-28 c460aae
Simplifying description on software design
