+++ paper.md
@@ -27,9 +27,7 @@ Gridap is a new Finite Element (FE) framework, exclusively written in the Julia 
 
 The main motivation behind Gridap is to find an improved balance between computational performance, user-experience and work-flow productivity, when working with FE libraries. Previous FE frameworks, e.g., FEniCS [@Alnaes2015] or Deal.II [@Bangerth2007] usually provide a high-level user front-end to facilitate the use of the library and a computational back-end to achieve performance. The user front-end is usually programmable in a interpreted language like Python, whereas the computational back-end is usually coded in a compiled language like C/C++ or Fortran. Users can benefit from the high-level front-end (i.e., for rapid prototyping) and simultaneously enjoy the performance of the compiled back-end. This approach reaches a compromise between performance and productivity when the back-end provides all the functionality required by the user. However, it does not satisfactorily address the needs of researchers on numerical methods willing to extend the library with new techniques or features. These extensions usually need to be done at the level of the computational back-end for performance reasons. Thus, the researcher is forced to develop new code in a compiled language like C/C++ instead of benefiting from the productivity of scripting languages like Python, incurring in serious productivity losses. In order to overcome this limitation, Gridap is fully implemented in the Julia programming language [@Bezanson2017]. Julia combines the performance of compiled languages with the productivity of interpreted ones by using type inference and just-in-time compilation to generate fast code. As a result, there is no need to use two different languages to write low-level performance code and high-level user interfaces. In addition, writing a FE library in Julia also allows one to leverage the feature-rich ecosystem of Julia libraries and exploit its excellent package manager. It permits a seamless coupling of Gridap with application-specific libraries, like optimization [@Dunning2017], approximation of ordinary differential equations [@Rackauckas2017], or data science [@flux_web], which can certainly boost the capabilities of a FE solver.
 
-Another major feature of Gridap is that it is not a simple Julia translation of a standard object-oriented FE code. There are other FE libraries written in Julia that have been inspired by standard FE frameworks, see e.g., JuAFEM [@Carlsson], whose interface resembles Deal.II.  In contrast,  Gridap adopts a novel software design based on cell-wise (scalar, vector, tensor) field and cell-wise array abstractions. Cell-wise fields can represent finite element functions, for which one defines unary operations (e.g., gradients) and binary operations (e.g., products) together with evaluations on cell-wise arrays of quadrature points to compute, e.g., elemental matrices and vectors that can finally be assembled for all cells of the computational domain. A key design concept of Gridap is _laziness_. Instead of explicitly storing the results of operations between cell-wise fields and cell-wise arrays, Gridap implements lazy objects that can compute the result at a given cell on-the-fly. This allows the library developers to hide assembly loops and other core computations from the user-code, leading to a very compact, user-friendly, syntax, while providing a high degree of flexibility for users to define their own weak forms using lambda functions. These lamba variational forms are internally called in Gridap with cell-wise fields arguments. This laziness also permits most Gridap objects to be immutable, eliminating the need of state diagrams in complex abstractions as the global finite element space.
-
-This design makes Gridap highly expressive. A Poisson or Stokes problem can be solved with Gridap in 10-20 lines of code by using the different abstractions provided by the library, as this example for the Poisson equation shows:
[…]
