+++ paper.bib
@@ -1,3 +1,125 @@
+@article{Innes2018,
+author = {Innes, Mike},
+doi = {10.21105/joss.00602},
+file = {:home/fverdugo/.local/share/data/Mendeley Ltd./Mendeley Desktop/Downloaded/Innes - 2018 - Flux Elegant machine learning with Julia.pdf:pdf},
+issn = {2475-9066},
+journal = {Journal of Open Source Software},
+month = {may},
+number = {25},
+pages = {602},
+publisher = {The Open Journal},
+title = {{Flux: Elegant machine learning with Julia}},
+url = {http://joss.theoj.org/papers/10.21105/joss.00602},
+volume = {3},
+year = {2018}
+}
+@article{Dunning2017,
+abstract = {JuMP is an open-source modeling language that allows users to express a wide range of optimization problems (linear, mixed-integer, quadratic, conic-quadratic, semidefinite, and nonlinear) in a high-level, algebraic syntax. JuMP takes advantage of advanced features of the Julia programming language to offer unique functionality while achieving performance on par with commercial modeling tools for standard tasks. In this work we will provide benchmarks, present the novel aspects of the implementation, and discuss how JuMP can be extended to new problem classes and composed with state-of-The-Art tools for visualization and interactivity.},
+archivePrefix = {arXiv},
+arxivId = {1508.01982},
+author = {Dunning, Iain and Huchette, Joey and Lubin, Miles},
+doi = {10.1137/15M1020575},
+eprint = {1508.01982},
+issn = {00361445},
+journal = {SIAM Review},
+keywords = {Algebraic modeling languages,Automatic differentiation,Scientific computing},
+month = {may},
+number = {2},
+pages = {295--320},
+publisher = {Society for Industrial and Applied Mathematics Publications},
+title = {{JuMP: A modeling language for mathematical optimization}},
+volume = {59},
+year = {2017}
+}
+@article{Carlsson_NLsolve_2019,
+author = {Carlsson, Kristoffer and Mogensen, Patrick Kofod and Villemot, S{\'{e}}bastien and Lyon, Spencer and Gomez, Matthieu and Rackauckas, Christopher and Holy, Tim and Kelman, Tony and Widmann, David and Macedo, Miguel Raz Guzm{\'{a}}n and Et al.},
+doi = {10.5281/zenodo.3357100},
+journal = {Zenodo},
+month = {jul},
+publisher = {Zenodo},
+title = {{JuliaNLSolvers/NLsolve.jl: v4.1.0}},
+year = {2019}
+}
+@misc{flux_web,
+howpublished = {https://fluxml.ai/},
+title = {{The Flux Project's Webpage}},
+url = {https://fluxml.ai/},
+urldate = {2019-08-25}
+}
+@article{Bezanson2017,
[…]
