### 2020-07-28 gridap/Gridap.jl#333 @fverdugo
@TeroFrondelius I have fixed the JuMP citation as you suggested.

If you are happy with the changes, please close the issue.
