### 2020-07-28 gridap/Gridap.jl#334 @fverdugo
@PetrKryslUCSD Following your suggestion, I have simplified the paragraph about software design. I have opted for not introducing concepts that cannot be properly explained in such compact document.

If you are happy with the changes, please close the issue.
