### 2020-07-28 gridap/Gridap.jl#335 @fverdugo
@PetrKryslUCSD The operators are in the paper.md source file, but they are not visualized correctly.

In standard .tex documents to be compiled with pdflatex, I am able to display these operators by including
```tex
\usepackage[utf8x]{inputenc}
\usepackage{textgreek}
```
in the preamble.  However, I do not know how to achieve this in this for the JOSS paper.

Dear editor  @Kevin-Mattheus-Moerman, could you provide any hint on how to fix this? Could you please forward the issue to the person in charge of the paper compilation system?
