### 2020-07-28 gridap/Gridap.jl#336 @fverdugo
@PetrKryslUCSD The steps to clone, build and open the notebooks of the Gridap tutorials are explained in the documentation pages of the tutorials repository https://gridap.github.io/Tutorials/stable/

In the read-me file of https://github.com/gridap/Tutorials there is a link to the documentation pages. Do you believe this is not enough?
