### 2020-07-28 gridap/Gridap.jl#336 @PetrKryslUCSD
My comment was based on the order in which the user would proceed. If I start with the readme, follow the link to the stable version of the tutorials, and click on the first tutorial I can see how to run it in the mybinder, but I don't know how to run it on my own machine.  So I go back to the readme.

Honestly, I missed the description that is below the list of tutorials. My bad. But I think it points to a possible reorganization where the documentation of the tutorials starts with instructions on how to run them, and then follows the list of tutorials.

Just a suggestion.
