### 2020-07-31 cb93c10
Updated NEWS.md
