### 2020-07-31 b682b9c
Update NEWS.md
