+++ test/ArraysTests/runtests.jl
@@ -28,6 +28,8 @@ using Test
 
 @testset "AppendedArrays" begin include("AppendedArraysTests.jl") end
 
+@testset "BlockArraysCoo" begin include("BlockArraysCooTests.jl") end
+
 @testset "AutodiffTests" begin include("AutodiffTests.jl") end
 
 end # module
