+++ test/ArraysTests/BlockArraysCooTests.jl
@@ -1,0 +1,236 @@
+module BlockArraysCooTests
+
+using BlockArrays
+using Gridap
+using Gridap.Arrays
+
+struct BlockArrayCoo{T,N,A,X} <: AbstractBlockArray{T,N}
+  blocks::Vector{A}
+  blockids::Vector{NTuple{N,Int}}
+  axes::X
+  ptrs::Array{Int,N}
+  function BlockArrayCoo(
+    blocks::Vector{<:AbstractArray{T,N}},
+    blockids::Vector{NTuple{N,Int}},
+    axes::NTuple{N},
+    ptrs::Array{Int,N}=_compute_ptrs(blockids,axes)) where {T,N}
+
+    A = eltype(blocks)
+    X = typeof(axes)
+    new{T,N,A,X}(blocks,blockids,axes,ptrs)
+  end
+end
+
+function _compute_ptrs(blockids,axes)
+  s = map(i->first(blocksize(i)),axes)
+  ptrs = zeros(Int,s)
+  for (i,c) in enumerate(blockids)
+    ptrs[c...] = i
+  end
+  ptrs
+end
+
+function BlockArrays.getblock(a::BlockArrayCoo{T,N,A} where {T,N},i::Integer...) where A
+  p = a.ptrs[i...]
+  if p>0
+    a.blocks[p]
+  else
+    s = map((k,l)->length(k[Block(l)]),a.axes,i)
+    block = A(undef,s)
+    fill!(block,zero(eltype(A)))
+    block
+  end
+end
+
+function BlockArrays.getblock!(block,a::BlockArrayCoo{T,N,A} where {T,N},i::Integer...) where A
+  p = a.ptrs[i...]
+  if p>0
+    block .= a.blocks[p]
+  else
+    fill!(block,zero(eltype(A)))
+    block
+  end
+end
+
+Base.size(a::BlockArrayCoo) = map(length,Base.axes(a))
+
+Base.axes(a::BlockArrayCoo) = a.axes
+
+function Base.getindex(a::BlockArrayCoo,i::Integer...)
+  s = map(findblockindex,a.axes,i)
+  ai = a[s...]
+  ai
+end
+
+struct VectorOfBlockArrayCoo{T,B,N,X} <: AbstractVector{T}
+  blocks::B
+  blockids::Vector{NTuple{N,Int}}
+  axes::X
+  ptrs::Array{Int,N}
+  function VectorOfBlockArrayCoo(
+    blocks::Tuple,
+    blockids::Vector{NTuple{N,Int}},
+    axes::AbstractArray{<:NTuple{N}},
+    ptrs::Array{Int,N} = _compute_ptrs(blockids,testitem(axes))) where N
+
+    msg = "Trying to build a VectorOfBlockArrayCoo with repeated blocks"
+    @assert _no_has_repeaded_blocks(blockids,ptrs) msg
+    @assert all(map(length,blocks) .==  length(first(blocks)) )
+
+    B = typeof(blocks)
+    X = typeof(axes)
+    blocks_i = collect(map(testitem,blocks))
+    axes_i = testitem(axes)
+    t = BlockArrayCoo(blocks_i,blockids,axes_i,ptrs)
+    T = typeof(t)
+    new{T,B,N,X}(blocks,blockids,axes,ptrs)
+  end
+end
+
+function _no_has_repeaded_blocks(blockids::Vector{NTuple{N,Int}},ptrs) where N
+  maxblocks = size(ptrs)
+  touched = zeros(Int,maxblocks)
+  for c in blockids
[…]
