### 2020-08-02 gridap/Gridap.jl#335 @Kevin-Mattheus-Moerman
@fverdugo could you paste the code generating the desired expression here? Also could you share a screenshot of the expression being rendered correctly (e.g. using a different Latex editor)? I'll have out development team look at it then.
Worst case an ugly but practical solution would be to represent the expression as an image in the paper instead.
