### 2020-08-03 bb6fc5a
More binary ops
