+++ test/ArraysTests/BlockArraysCooTests.jl
@@ -227,6 +227,16 @@ function apply_field_op(k::FieldOp,a::VectorOfBlockArrayCoo)
   VectorOfBlockArrayCoo(blocks,a.blockids,a.axes,a.ptrs)
 end
 
+function apply_field_op(k::FieldOp,a::VectorOfBlockArrayCoo,b::AbstractArray)
+  blocks = map(block->apply(k,block,b), a.blocks)
+  VectorOfBlockArrayCoo(blocks,a.blockids,a.axes,a.ptrs)
+end
+
+function apply_field_op(k::FieldOp,b::AbstractArray,a::VectorOfBlockArrayCoo)
+  blocks = map(block->apply(k,b,block), a.blocks)
+  VectorOfBlockArrayCoo(blocks,a.blockids,a.axes,a.ptrs)
+end
+
 function apply_field_op(k::FieldOp,a::VectorOfBlockArrayCoo{Ta,N} where Ta, b::VectorOfBlockArrayCoo{Tb,N} where Tb) where N
   @assert size(a.ptrs) == size(b.ptrs)
   blocks = []
@@ -291,30 +301,45 @@ function apply_field_op(k::FieldOp{typeof(-)},a::VectorOfBlockArrayCoo{Ta,N} whe
   VectorOfBlockArrayCoo(Tuple(blocks),blockids,a.axes)
 end
 
-#function apply_field_op(k::FieldOp,a::VectorOfBlockArrayCoo{Ta,N} where Ta, b::VectorOfBlockArrayCoo{Ta,N}) where N
-#  @assert size(a.ptrs) == size(b.ptrs)
-#  blocks = []
-#  blockids = []
-#  cis = CartesianIndices(a.ptrs)
-#  for ci in cis
-#    p1 = a.ptrs[ci]
-#    p2 = a.ptrs[ci]
-#    if p1 > 0 && p2 >0
-#      block = apply(k,a.blocks[p1],b.blocks[p2])
-#      push!(blocks,block)
-#      push!(blockids,Tuple(ci))
-#    elseif p2 > 0
-#      block = apply(k,b.blocks[p2])
-#      push!(blocks,block)
-#      push!(blockids,Tuple(ci))
-#    elseif p1 > 0
-#      block = a.blocks[p1]
-#      push!(blocks,block)
-#      push!(blockids,Tuple(ci))
-#    end
-#  end
-#  VectorOfBlockArrayCoo(Tuple(blocks),collect(blockids),a.axes)
-#end
+function apply_field_op(k::FieldOp,a::VectorOfBlockArrayCoo{Ta,2} where Ta, b::VectorOfBlockArrayCoo{Tb,3} where Tb)
+  axs = apply( (a1,a2) -> (a1[1],a1[2],a2[3]) ,a.axes,b.axes)
+  blocks = []
+  blockids = NTuple{3,Int}[]
+  nfield1 = size(a.ptrs,2)
+  nfield2 = size(b.ptrs,3)
+  for f1 in 1:nfield1
+    I1 = Block(1,f1)
+    for f2 in 1:nfield2
+      I2 = Block(1,1,f2)
+      if !is_zero_block(a,I1) && !is_zero_block(b,I2)
+        block = apply(k,a[I1],b[I2])
+        push!(blocks,block)
+        push!(blockids,(1,f1,f2))
+      end
+    end
+  end
+  VectorOfBlockArrayCoo(Tuple(blocks),blockids,axs)
+end
+
[…]
