### 2020-08-03 b78aa8a
Deleted apply_field_op
