+++ test/ArraysTests/BlockArraysCooTests.jl
@@ -7,6 +7,7 @@ import Gridap.Arrays: kernel_cache
 import Gridap.Arrays: apply_kernel!
 import Gridap.Arrays: array_cache
 import Gridap.Arrays: getindex!
+import Gridap.Arrays: apply
 
 Gridap.Arrays.array_caches() = ()
 Gridap.Arrays.getitems!(::Tuple{},::Tuple{},i) = ()
@@ -218,26 +219,22 @@ function is_zero_block(a::VectorOfBlockArrayCoo,i::Integer...)
   p < 0
 end
 
-function apply_field_op(k::FieldOp,args::AbstractArray...)
-  apply(k,args...)
-end
-
-function apply_field_op(k::FieldOp,a::VectorOfBlockArrayCoo)
+function apply(k::FieldOp,a::VectorOfBlockArrayCoo)
   blocks = map(b->apply(k,b), a.blocks)
   VectorOfBlockArrayCoo(blocks,a.blockids,a.axes,a.ptrs)
 end
 
-function apply_field_op(k::FieldOp,a::VectorOfBlockArrayCoo,b::AbstractArray)
+function apply(k::FieldOp,a::VectorOfBlockArrayCoo,b::AbstractArray)
   blocks = map(block->apply(k,block,b), a.blocks)
   VectorOfBlockArrayCoo(blocks,a.blockids,a.axes,a.ptrs)
 end
 
-function apply_field_op(k::FieldOp,b::AbstractArray,a::VectorOfBlockArrayCoo)
+function apply(k::FieldOp,b::AbstractArray,a::VectorOfBlockArrayCoo)
   blocks = map(block->apply(k,b,block), a.blocks)
   VectorOfBlockArrayCoo(blocks,a.blockids,a.axes,a.ptrs)
 end
 
-function apply_field_op(k::FieldOp,a::VectorOfBlockArrayCoo{Ta,N} where Ta, b::VectorOfBlockArrayCoo{Tb,N} where Tb) where N
+function apply(k::FieldOp,a::VectorOfBlockArrayCoo{Ta,N} where Ta, b::VectorOfBlockArrayCoo{Tb,N} where Tb) where N
   @assert size(a.ptrs) == size(b.ptrs)
   blocks = []
   blockids = NTuple{N,Int}[]
@@ -253,7 +250,7 @@ function apply_field_op(k::FieldOp,a::VectorOfBlockArrayCoo{Ta,N} where Ta, b::V
   VectorOfBlockArrayCoo(Tuple(blocks),blockids,a.axes)
 end
 
-function apply_field_op(k::FieldOp{typeof(+)},a::VectorOfBlockArrayCoo{Ta,N} where Ta, b::VectorOfBlockArrayCoo{Tb,N} where Tb) where N
+function apply(k::FieldOp{typeof(+)},a::VectorOfBlockArrayCoo{Ta,N} where Ta, b::VectorOfBlockArrayCoo{Tb,N} where Tb) where N
   @assert size(a.ptrs) == size(b.ptrs)
   blocks = []
   blockids = NTuple{N,Int}[]
@@ -277,7 +274,7 @@ function apply_field_op(k::FieldOp{typeof(+)},a::VectorOfBlockArrayCoo{Ta,N} whe
   VectorOfBlockArrayCoo(Tuple(blocks),blockids,a.axes)
 end
 
[…]
