### 2020-08-04 c01f4ef
Reverted re-definition of Base.getindex() for Table
Thus, now, one cannot longer modify the contents of a Table using Base.getindex()
