### 2020-08-04 bf66192
Updated NEWS.md
