+++ NEWS.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+  - Product binary operation (`*`) between a scalar and a field gradient. Since PR [#340](https://github.com/gridap/Gridap.jl/pull/340).
+
 ## [0.13.2] - 2020-7-31
 
 ### Added
