### 2020-08-04 bb67edd
Better definition of changes in NEWS.md
