### 2020-08-04 615aaea
Added degenerated cases for array_caches and getitems!
