### 2020-08-04 32ea2f5
Moved BlockArrayCoo to the library
