+++ test/ArraysTests/BlockArraysCooTests.jl
@@ -1,491 +1,33 @@
 module BlockArraysCooTests
 
-using BlockArrays
-using Gridap
-using Gridap.Arrays
-using Gridap.Fields
-using Gridap.Fields: IntKernel
-using Gridap.TensorValues
-using LinearAlgebra
-import Gridap.Arrays: kernel_cache
-import Gridap.Arrays: apply_kernel!
-import Gridap.Arrays: array_cache
-import Gridap.Arrays: getindex!
-import Gridap.Arrays: apply
-
-Gridap.Arrays.array_caches() = ()
-Gridap.Arrays.getitems!(::Tuple{},::Tuple{},i) = ()
-
-struct BlockArrayCoo{T,N,A,X} <: AbstractBlockArray{T,N}
-  blocks::Vector{A}
-  blockids::Vector{NTuple{N,Int}}
-  axes::X
-  ptrs::Array{Int,N}
-  zero_blocks::Vector{A}
-  function BlockArrayCoo(
-    blocks::Vector{A},
-    blockids::Vector{NTuple{N,Int}},
-    axes::NTuple{N},
-    ptrs::Array{Int,N}=_compute_ptrs(blockids,axes),
-    zero_blocks::Vector{A}=_compute_zero_blocks(A,ptrs,axes)) where {T,N,A<:AbstractArray{T,N}}
-
-    X = typeof(axes)
-    new{T,N,A,X}(blocks,blockids,axes,ptrs,zero_blocks)
-  end
-end
-
-function _compute_ptrs(blockids,axes)
-  s = map(i->first(blocksize(i)),axes)
-  ptrs = zeros(Int,s)
-  for (i,c) in enumerate(blockids)
-    ptrs[c...] = i
-  end
-  m = 1
-  for (k,p) in enumerate(ptrs)
-    if p ==0
-      ptrs[k] = -m
-      m += 1
-    end
-  end
-  ptrs
-end
-
-function _compute_zero_blocks(::Type{A},ptrs,axes) where A
-  cis = CartesianIndices(ptrs)
-  zero_blocks = A[]
-  for ci in cis
-    p = ptrs[ci]
-    if p<0
-      i = Tuple(ci)
-      s = map((k,l)->length(k[Block(l)]),axes,i)
-      block = A(undef,s)
-      fill!(block,zero(eltype(A)))
-      push!(zero_blocks,block)
-    end
-  end
-  zero_blocks
-end
-
-function BlockArrays.getblock(a::BlockArrayCoo,i::Integer...)
-  p = a.ptrs[i...]
-  if p>0
-    a.blocks[p]
-  else
-    a.zero_blocks[-p]
-  end
-end
-
-function is_zero_block(a,b::Block)
-  i = convert(Tuple,b)
-  is_zero_block(a,i...)
-end
-
-function is_zero_block(a,b::CartesianIndex)
-  i = Tuple(b)
-  is_zero_block(a,i...)
-end
-
-function is_zero_block(a::BlockArrayCoo,i::Integer...)
-  p = a.ptrs[i...]
-  @assert p != 0
-  p < 0
-end
-
-function BlockArrays.getblock!(block,a::BlockArrayCoo,i::Integer...)
-  p = a.ptrs[i...]
-  if p>0
-    block .= a.blocks[p]
-  else
-    block .= a.zero_blocks[-p]
-  end
-end
-
-Base.size(a::BlockArrayCoo) = map(length,Base.axes(a))
-
-Base.axes(a::BlockArrayCoo) = a.axes
-
[…]
