### 2020-08-04 88d3015
Moving code from tests to library
