+++ src/Arrays/BlockArraysCoo.jl
@@ -17,6 +17,9 @@ struct BlockArrayCoo{T,N,A,X} <: AbstractBlockArray{T,N}
   end
 end
 
+const BlockMatrixCoo = BlockArrayCoo{T,2} where T
+const BlockVectorCoo = BlockArrayCoo{T,1} where T
+
 function _compute_ptrs(blockids,axes)
   s = map(i->first(blocksize(i)),axes)
   ptrs = zeros(Int,s)
@@ -110,5 +113,151 @@ function Base.getindex(a::BlockArrayCoo,i::Integer...)
   s = map(findblockindex,a.axes,i)
   ai = a[s...]
   ai
+end
+
+function Base.:*(a::BlockArrayCoo,b::BlockArrayCoo)
+  c = _mul_block_result(a,b)
+  mul!(c,a,b)
+end
+
+function _mul_block_result(a::BlockMatrixCoo,b::BlockVectorCoo)
+  @assert blocksize(a,2) == blocksize(b,1)
+  T = promote_type(eltype(a),eltype(b))
+  blocks = Vector{T}[]
+  blockids = Tuple{Int}[]
+  for i in 1:blocksize(a,1)
+    for j in 1:blocksize(a,2)
+      if is_nonzero_block(a,Block(i,j)) && is_nonzero_block(b,Block(j))
+        block = zeros(T,size(a[Block(i,j)],1))
+        push!(blocks,block)
+        push!(blockids,(i,))
+        break
+      end
+    end
+  end
+  axs = (axes(a)[1],)
+  BlockArrayCoo(blocks,blockids,axs)
+end
+
+function _mul_block_result(a::BlockMatrixCoo,b::BlockMatrixCoo)
+  @assert blocksize(a,2) == blocksize(b,1)
+  T = promote_type(eltype(a),eltype(b))
+  blocks = Matrix{T}[]
+  blockids = Tuple{Int,Int}[]
+  for i in 1:blocksize(a,1)
+    for j in 1:blocksize(b,2)
+      for k in 1:blocksize(a,2)
+        if is_nonzero_block(a,Block(i,k)) && is_nonzero_block(b,Block(k,j))
+          block = zeros(T,size(a[Block(i,k)],1),size(b[Block(k,j)],2))
+          push!(blocks,block)
+          push!(blockids,(i,j))
+          break
+        end
+      end
+    end
+  end
+  axs = (axes(a)[1],axes(b)[2])
+  BlockArrayCoo(blocks,blockids,axs)
+end
+
+@static if VERSION >= v"1.3"
+
+  function LinearAlgebra.mul!(c::BlockVectorCoo,a::BlockMatrixCoo,b::BlockVectorCoo,α,β)
+    for I in 1:blocksize(a,1)
+      cI = c[Block(I)]
+      for i in eachindex(cI)
+        cI[i] = β*cI[i]
+      end
+      for J in 1:blocksize(a,2)
+        if is_nonzero_block(a,Block(I,J)) && is_nonzero_block(b,Block(J))
+          aIJ = a[Block(I,J)]
+          bJ = b[Block(J)]
+          mul!(cI,aIJ,bJ,α,1)
+        end
+      end
+    end
+    c
+  end
+
+  function LinearAlgebra.mul!(c::BlockMatrixCoo,a::BlockMatrixCoo,b::BlockMatrixCoo,α,β)
+    for I in 1:blocksize(a,1)
+      for J in 1:blocksize(b,2)
+        cIJ = c[Block(I,J)]
+        for ij in eachindex(cIJ)
+          cIJ[ij] = β*cIJ[ij]
+        end
+        for K in 1:blocksize(a,2)
+          if is_nonzero_block(a,Block(I,K)) && is_nonzero_block(b,Block(K,J))
[…]
