### 2020-08-04 edc19e4
Added MulKernel
