+++ src/Arrays/Arrays.jl
@@ -60,6 +60,7 @@ export test_kernel
 export bcast
 export elem
 export contract
+export MulKernel
 export kernel_return_type
 export kernel_return_types
 export kernel_testitem
