+++ src/Arrays/Kernels.jl
@@ -465,3 +465,28 @@ end
   c
 end
 
+struct MulKernel <: Kernel end
+
+function kernel_cache(k::MulKernel,a,b)
+  c = a*b
+  CachedArray(c)
+end
+
+@inline function apply_kernel!(cache,k::MulKernel,a::AbstractMatrix,b::AbstractVector)
+  m = axes(a,1)
+  setaxes!(cache,(m,))
+  c = cache.array
+  mul!(c,a,b)
+  c
+end
+
+@inline function apply_kernel!(cache,k::MulKernel,a::AbstractMatrix,b::AbstractMatrix)
+  m = axes(a,1)
+  n = axes(b,2)
+  setaxes!(cache,(m,n))
+  c = cache.array
+  mul!(c,a,b)
+  c
+end
+
+
