### 2020-08-05 25d5e95
Added a pair of flags to create_gridap_image.jl that allow the user to point to an existing source directory for Gridap and the tutorials, instead of cloning them
