### 2020-08-05 41867a5
Fixing some bugs
