### 2020-08-05 4eecb85
Fixing printing of BlockVectorCoo
