### 2020-08-05 363ee01
Added VectorOfBlockArray
