+++ src/Arrays/VectorsOfBlockArrayCoo.jl
@@ -1,0 +1,127 @@
+
+struct VectorOfBlockArrayCoo{T,N,B,X,Z} <: AbstractVector{T}
+  blocks::B
+  blockids::Vector{NTuple{N,Int}}
+  axes::X
+  ptrs::Array{Int,N}
+  zero_blocks::Z
+  function VectorOfBlockArrayCoo(
+    blocks::Tuple,
+    blockids::Vector{NTuple{N,Int}},
+    axes::AbstractArray{<:NTuple{N}},
+    ptrs::Array{Int,N} = _compute_ptrs(blockids,testitem(axes)),
+    zero_blocks::Tuple=_compute_zero_blocks_array(blocks,ptrs,axes)) where N
+
+    msg = "Trying to build a VectorOfBlockArrayCoo with repeated blocks"
+    @assert _no_has_repeaded_blocks(blockids,ptrs) msg
+    @assert all(map(length,blocks) .==  length(first(blocks)) )
+
+    B = typeof(blocks)
+    X = typeof(axes)
+    blocks_i = collect(map(testitem,blocks))
+    axes_i = testitem(axes)
+    zero_blocks_i = collect(map(testitem,zero_blocks))
+    t = BlockArrayCoo(blocks_i,blockids,axes_i,ptrs,zero_blocks_i)
+    T = typeof(t)
+    Z = typeof(zero_blocks)
+    new{T,N,B,X,Z}(blocks,blockids,axes,ptrs,zero_blocks)
+  end
+end
+
+const VectorOfBlockVectorCoo = VectorOfBlockArrayCoo{T,1} where T
+const VectorOfBlockMatrixCoo = VectorOfBlockArrayCoo{T,2} where T
+
+function _no_has_repeaded_blocks(blockids::Vector{NTuple{N,Int}},ptrs) where N
+  maxblocks = size(ptrs)
+  touched = zeros(Int,maxblocks)
+  for c in blockids
+    touched[c...] += 1
+  end
+  all( touched .<= 1 )
+end
+
+function _compute_zero_blocks_array(blocks,ptrs,axes)
+  A = eltype(first(blocks))
+  cis = CartesianIndices(ptrs)
+  zero_blocks = []
+  for ci in cis
+    p = ptrs[ci]
+    if p<0
+      block = apply(axes) do a
+        i = Tuple(ci)
+        s = map((k,l)->length(k[Block(l)]),a,i)
+        b = A(undef,s)
+        fill!(b,zero(eltype(A)))
+        b
+      end
+      push!(zero_blocks,block)
+    end
+  end
+  Tuple(zero_blocks)
+end
+
+BlockArrays.blocksize(a::VectorOfBlockArrayCoo) = blocksize(testitem(a))
+
+BlockArrays.blocksize(a::VectorOfBlockArrayCoo,i::Integer) = blocksize(testitem(a),i)
+
+function BlockArrays.eachblock(a::VectorOfBlockArrayCoo)
+  cis = CartesianIndices(blocksize(a))
+  blocks = map(ci->Block(Tuple(ci)),cis)
+  ( a[block] for block in blocks  )
+end
+
+Base.IndexStyle(::Type{<:VectorOfBlockArrayCoo}) = IndexLinear()
+
+Base.size(a::VectorOfBlockArrayCoo) = (length(first(a.blocks)),)
+
+function array_cache(a::VectorOfBlockArrayCoo)
+  blocks_i = collect(map(testitem,a.blocks))
+  zero_blocks_i = collect(eltype(blocks_i),map(testitem,a.zero_blocks))
+  ca = array_cache(a.axes)
[…]
