### 2020-08-05 #2520 @Kevin-Mattheus-Moerman
@PetrKryslUCSD, @TeroFrondelius, @KristofferC could you give a brief update on where this stands? It looks like @fverdugo resolved some of the issues pointed out. :point_up:, perhaps you can check those out.

@TeroFrondelius, @KristofferC can you start working on those checkboxes?

Thanks for all your help!
