### 2020-08-05 gridap/Gridap.jl#335 @fverdugo
This is a MWE with some code snippets containing the Greek operators.

Consider a file `main.tex` with these contents:
```tex
\documentclass{article}
\usepackage[utf8x]{inputenc}
\usepackage{textgreek}
\begin{document}
\begin{verbatim}
f(x) = -Δ(u)(x); g(x) = u(x)
a(u,v) = ∇(u)⋅∇(v); l(v) = v*f
trian_Ω = Triangulation(model)
\end{verbatim}
\end{document}
```

Compile it with pdflatex
```bash
pdflatex main.tex
```
Result:
[main.pdf](https://github.com/gridap/Gridap.jl/files/5028555/main.pdf)

As I have said, the trick is to include these lines in the preamble:
```tex
\usepackage[utf8x]{inputenc}
\usepackage{textgreek}
```
