### 2020-08-05 gridap/Gridap.jl#336 @fverdugo
Perhaps the problem is that the instructions on how to build the notebooks are not in the section called **How to start**.

https://gridap.github.io/Tutorials/stable/#How-to-start-1

I'll try to fix this.
