### 2020-08-05 gridap/Gridap.jl#336 @fverdugo
Hi @PetrKryslUCSD

I have rewritten the confusing part. I hope that now it is better:

https://gridap.github.io/Tutorials/v0.12/

If you are happy with the changes, please close the issue.

Thanks!
