### 2020-08-06 2a9b661
iSeveral fixes to work with blocks of blocks
