+++ src/Arrays/BlockArraysCoo.jl
@@ -71,10 +71,17 @@ end
 const BlockMatrixCoo = BlockArrayCoo{T,2} where T
 const BlockVectorCoo = BlockArrayCoo{T,1} where T
 
-function all_zero_blocks(a::BlockArrayCoo)
+function zeros_like(a::BlockArrayCoo)
   blocks = eltype(a.blocks)[]
   blockids = eltype(a.blockids)[]
   BlockArrayCoo(blocks,blockids,a.axes)
+end
+
+function zeros_like(a::AbstractArray)
+  T = eltype(a)
+  b = similar(a,T,size(a))
+  fill!(b,zero(T))
+  b
 end
 
 function _compute_ptrs(blockids,axes)
@@ -186,7 +193,8 @@ end
 function Base.:+(a::BlockArrayCoo{Ta,N},b::BlockArrayCoo{Tb,N}) where {Ta,Tb,N}
   @assert axes(a) == axes(b)
   @assert blockaxes(a) == blockaxes(b)
-  A = promote_type(eltype(a.blocks),eltype(b.blocks))
+  I1 = first(eachblockindex(a))
+  A = typeof(a[I1]+b[I1])
   blocks = A[]
   blockids = NTuple{N,Int}[]
   for (I,aI) in enumerateblocks(a)
@@ -203,7 +211,8 @@ end
 function Base.:-(a::BlockArrayCoo{Ta,N},b::BlockArrayCoo{Tb,N}) where {Ta,Tb,N}
   @assert axes(a) == axes(b)
   @assert blockaxes(a) == blockaxes(b)
-  A = promote_type(eltype(a.blocks),eltype(b.blocks))
+  I1 = first(eachblockindex(a))
+  A = typeof(a[I1]-b[I1])
   blocks = A[]
   blockids = NTuple{N,Int}[]
   for (I,aI) in enumerateblocks(a)
@@ -222,56 +231,50 @@ function Base.:*(a::BlockArrayCoo,b::Number)
 end
 
 function Base.:*(a::Number,b::BlockArrayCoo)
-  blocks = [a*block for block in b.blocks]
+  f(block)= a*block
+  blocks = f.(b.blocks)
   BlockArrayCoo(blocks,b.blockids,b.axes,b.ptrs,b.zero_blocks)
 end
 
-function Base.:*(a::BlockArrayCoo,b::BlockArrayCoo)
-  c = _mul_block_result(a,b)
-  mul!(c,a,b)
-end
-
-function _mul_block_result(a::BlockMatrixCoo,b::BlockVectorCoo)
+function Base.:*(a::BlockMatrixCoo,b::BlockVectorCoo)
   @assert blocksize(a,2) == blocksize(b,1)
-  T = promote_type(eltype(a),eltype(b))
-  blocks = Vector{T}[]
+  A = typeof(a[Block(1,1)]*b[Block(1)])
+  blocks = A[]
   blockids = Tuple{Int}[]
   for i in 1:blocksize(a,1)
+    block = zeros_like(a[Block(i,1)]*b[Block(1)])
     for j in 1:blocksize(a,2)
       if is_nonzero_block(a,Block(i,j)) && is_nonzero_block(b,Block(j))
-        block = zeros(T,size(a[Block(i,j)],1))
-        push!(blocks,block)
-        push!(blockids,(i,))
-        break
+        block += a[Block(i,j)]*b[Block(j)]
       end
     end
+    push!(blocks,block)
+    push!(blockids,(i,))
   end
   axs = (axes(a)[1],)
   BlockArrayCoo(blocks,blockids,axs)
 end
 
-function _mul_block_result(a::BlockMatrixCoo,b::BlockMatrixCoo)
+function Base.:*(a::BlockMatrixCoo,b::BlockMatrixCoo)
[…]
