### 2020-08-06 gridap/Gridap.jl#342 @TeroFrondelius
joss paper: OSI approved software license?
Hi,
I really like your idea to add Citing Gridap and Contributing to Gridap titles to the LICENSE.md file. However, I have to compare the text here: https://opensource.org/licenses/MIT and it's not matching. I think this might be the reason why Github is not automatically recognized your license to be MIT. I hope you can find a solution to keep the extra stuff in because I really see the benefit. Maybe you can link some discussion or something that states this is ok. Of course, the dull option is to remove the extra stuff and keep only the original MIT license text.

X-ref: https://github.com/openjournals/joss-reviews/issues/2520
