### 2020-08-06 gridap/Gridap.jl#343 @TeroFrondelius
joss paper: Contribution and authorship
I think that in Engineering we have too few authors in all papers. Anyway, I checked the contributions and the best of my understanding @victorsndvg has made some major contributions and could be listed as joss-paper author.

Xref: https://github.com/openjournals/joss-reviews/issues/2520
