### 2020-08-06 c909cb7
Stressing VectorsOfBlockArrayCoo for heterogeneous case
