### 2020-08-07 c4c19f6
More improvements for supporting blocks of blocks
