+++ src/Arrays/BlockArraysCoo.jl
@@ -30,7 +30,7 @@ function BlockArrayCoo(
   BlockArrayCoo(blocks,blockids,ax,ptrs,zero_blocks)
 end
 
-function BlockArrayCoo(allblocks::AbstractArray{A,N},mask::AbstractArray{Bool,N}=fill(true,size(allblocks))) where {A,N}
+function BlockArrayCoo(allblocks::AbstractArray{A,N},ax::NTuple{N},mask::AbstractArray{Bool,N}=fill(true,size(allblocks))) where {A,N}
   blocks = A[]
   zero_blocks = A[]
   ptrs = zeros(Int,size(allblocks))
@@ -48,25 +48,26 @@ function BlockArrayCoo(allblocks::AbstractArray{A,N},mask::AbstractArray{Bool,N}
       pn += 1
     end
   end
-  axs = _compute_axes_from_allblocks(allblocks)
-  BlockArrayCoo(blocks,axs,ptrs,zero_blocks)
+  BlockArrayCoo(blocks,ax,ptrs,zero_blocks)
 end
 
-function _compute_axes_from_allblocks(allblocks)
-  N = ndims(allblocks)
-  s = size(allblocks)
-  axtmp = [ zeros(Int,s[n]) for n in 1:N]
-  cis = CartesianIndices(allblocks)
-  for ci in cis
-    block = allblocks[ci]
-    sb = size(block)
-    for (n,cin) in enumerate(Tuple(ci))
-      axtmp[n][cin] = sb[n]
-    end
-  end
-  axs = map(blockedrange,Tuple(axtmp))
-  axs
-end
+#  axs = _compute_axes_from_allblocks(allblocks)
+#
+#function _compute_axes_from_allblocks(allblocks)
+#  N = ndims(allblocks)
+#  s = size(allblocks)
+#  axtmp = [ zeros(Int,s[n]) for n in 1:N]
+#  cis = CartesianIndices(allblocks)
+#  for ci in cis
+#    block = allblocks[ci]
+#    sb = size(block)
+#    for (n,cin) in enumerate(Tuple(ci))
+#      axtmp[n][cin] = sb[n]
+#    end
+#  end
+#  axs = map(blockedrange,Tuple(axtmp))
+#  axs
+#end
 
 const BlockMatrixCoo = BlockArrayCoo{T,2} where T
 const BlockVectorCoo = BlockArrayCoo{T,1} where T
@@ -96,6 +97,28 @@ function zeros_like(::Type{A},axs::Tuple) where A <: AbstractArray
   a = similar(A,axs)
   fill!(a,zero(eltype(a)))
   a
+end
+
+Base.similar(a::BlockArrayCoo) = similar(a,eltype(a),axes(a))
+
+Base.similar(a::BlockArrayCoo,axs::Tuple) = similar(a,eltype(a),axs)
+
+function Base.similar(a::BlockArrayCoo{S,N} where S,::Type{T}, axs::NTuple{N,<:BlockedUnitRange}) where {T,N}
+  _similar_block_array_coo(a,T,axs)
+end
+
+function _similar_block_array_coo(a,::Type{T},axs) where T
+  S = eltype(a)
+  @notimplementedif S != T
+  A = eltype(a.blocks)
+  blocks = A[]
+  for p in 1:length(a.blocks)
+    I = a.blockids[p]
+    laxs = map( local_range, axs, I)
+    block = similar(a.blocks[p],T,laxs)
+    push!(blocks,block)
+  end
+  BlockArrayCoo(blocks,a.blockids,axs,a.ptrs)
 end
 
 function local_range(a::BlockedUnitRange,k::Integer)
[…]
