### 2020-08-08 gridap/Gridap.jl#343 @santiagobadia
@TeroFrondelius thanks for the suggestion. In any case, we would like to keep the list of authors is it is. @fverdugo and myself were the ones that started the project, designed its kernel, and the only developers during the first stage of the project.
