### 2020-08-08 gridap/Gridap.jl#342 @santiagobadia
After some research, it seems there is no way for GitHub to automatically detect an _enriched_ license file. The only two possibilities are: 1) to remove the citation, certificate of origin, etc, and keep the plain MIT license or 2) to keep it without the Github detection.

We would like to keep it as it is, if possible, since we have not modified a line of the MIT license and I consider that our license fulfills the requirements of JOSS. In fact, it is very common to find respected open source Github projects, e.g., Julia itself, DifferentialEquations.jl, deal.ii, etc, that are in the same situation as Gridap (no Github license detection).
