### 2020-08-10 182fdff
Fixing VectorsOfBlockArrayCoo for block of blocks
