### 2020-08-10 gridap/Gridap.jl#342 @TeroFrondelius
I come from a corporate background and the legal department can be a big pain in the ass. Thus, I want you to avoid a similar situation than this famous JSLint case https://www.youtube.com/watch?v=-hCimLnIsDA
