### 2020-08-10 ec8390e
Some punctuation proposals for paper.md
I ran the text through Grammarly and here are some fixes.
