### 2020-08-10 6da2d79
Moved some code to the library
