+++ src/Fields/Fields.jl
@@ -20,6 +20,7 @@ using Gridap.Arrays: NumberOrArray
 using Gridap.Arrays: AppliedArray
 using Gridap.Arrays: Contracted
 using LinearAlgebra: ⋅
+using BlockArrays
 
 using Test
 using DocStringExtensions
@@ -77,6 +78,7 @@ import Gridap.Arrays: kernel_cache
 import Gridap.Arrays: apply_kernel!
 import Gridap.Arrays: kernel_return_type
 import Gridap.Arrays: kernel_testitem!
+import Gridap.Arrays: apply
 import Gridap.TensorValues: outer
 import Gridap.TensorValues: inner
 import Gridap.TensorValues: symmetric_part
