### 2020-08-10 gridap/Gridap.jl#348 @TeroFrondelius
joss paper: list of references
Currently, you don't have any general FEM or algorithms citations, also the number of citations seems a bit too low for me. Could you add some references here:

> The library provides a feature-rich set of discretization
techniques, including continuous and discontinuous FE methods with Lagrangian, RaviartThomas, or Nédélec interpolations

Also here, if you have a good reference(s):

>  Gridap adopts a novel software design that allows one to manipulate different types
of data associated with the cells of the computational mesh in a convenient way

And here:

> cells of the mesh are usually lazy,

Xref: https://github.com/openjournals/joss-reviews/issues/2520
