### 2020-08-10 gridap/Gridap.jl#343 @TeroFrondelius
Here: https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort, it defines 1000 LOC as a major contribution. Of course, I didn't look very detailed, but @victorsndvg has contributed 5903 LOC in 78 commits.
