### 2020-08-10 #2520 @KristofferC
Hello again and sorry for the late replies. I will give my comments so they are available here within 24 hours. I hope that is OK.
