### 2020-08-10 526d2e8
More optimizations for Block of Blocks
