+++ test/runtests.jl
@@ -30,6 +30,10 @@ using Test
 
 @time @testset "Visualization" begin include("VisualizationTests/runtests.jl") end
 
+@time @testset "Issues" begin
+    include("issue_349.jl")
+end
+
 include("GridapTests/runtests.jl")
 
 include("../bench/runbenchs.jl")
