+++ README.md
@@ -53,7 +53,7 @@ These are some popular PDEs solved with the Gridap library. Examples taken from 
 
 ## How to cite Gridap
 
-If you have used the Gridap library in a scientific publication, please cite the project as follows:
+In order to give credit to the `Gridap` contributors, we simply ask you to cite the refence below in any publication in which you have made use of `Gridap` packages:
 
 ```
 @software{gridap_project,
