### 2020-08-10 8f581b3
Delete LICENSE-old.md
