### 2020-08-10 gridap/Gridap.jl#342 @santiagobadia
Ok, we have now put a plain MIT license, which is detected by Github. I close the issue.
