+++ paper.bib
@@ -4,7 +4,6 @@ archivePrefix = {arXiv},
 arxivId = {1708.01773},
 author = {Badia, Santiago and Mart{\'{i}}n, Alberto F. and Principe, Javier},
 doi = {10.1007/s11831-017-9244-1},
-eprint = {1708.01773
 file = {:home/santiago/.local/share/data/Mendeley Ltd./Mendeley Desktop/Downloaded/Badia, Mart{\'{i}}n, Principe - 2018 - FEMPAR An Object-Oriented Parallel Finite Element Framework.pdf:pdf},
 issn = {18861784},
 journal = {Archives of Computational Methods in Engineering},
