### 2020-08-11 gridap/Gridap.jl#343 @santiagobadia
As far as I can see, this number is about the min scholarship effort of the project to be accepted. We consider that LOC alone do not give much information. E.g., it is not the same to create wrappers using clang than designing the kernel of Gridap.

Victor created wrappers to PETSc and re-implement in Julia the COO matrices and compression algorithms in our previous [project](https://github.com/fempar). This is not the core of the library, and as a result, we consider that the list of authors is appropriate and complete in its current version. On the other hand, the project has had about 200K of LOCs.

So, we consider that this the list of authors is appropriate and complete.
