### 2020-08-11 gridap/Gridap.jl#348 @santiagobadia
We have added a reference with respect to the finite element abstractions.

We do not have references for the other two sentences, but right after these sentences we explain what do they mean.  The software design is new, it is the first paper about Gridap, and since JOSS articles are not full-length articles, we cannot explain them with more detail here.
