### 2020-08-11 gridap/Gridap.jl#343 @TeroFrondelius
I accept the explanation, even though I don't fully agree. Engineering and mathematics should learn from medical and physics, where the list of authors is much longer and smaller contributions to the research are also recognized.
