### 2020-08-11 gridap/Gridap.jl#348 @TeroFrondelius
Still, the total number of References seems low to me, comparing this is a fully-featured FEM package. Nevertheless, it's impossible to add a lot of references to a short text.
