### 2020-08-11 gridap/Gridap.jl#353 @TeroFrondelius
joss paper: memory benchmark
In the paper, you have this statement:
>  Instead, the value for a specific cell is computed on-the-fly when needed, which certainly reduces memory requirements.

It would be really cool to benchmark this. We have had in JuliaFEM this benchmark case ([here is the paper](https://rakenteidenmekaniikka.journal.fi/article/view/64224/26397)):
![image](https://user-images.githubusercontent.com/12501534/89879688-f98f8d00-dbcb-11ea-9fb1-96a60dea87e4.png)

Unfortunately, we don't mention memory usage, probably because it's huge, but I guess we should run both models on the same computer anyway. What do you think?

Ping @KristofferC and @ahojukka5

Xref: https://github.com/openjournals/joss-reviews/issues/2520
