### 2020-08-11 #2520 @fverdugo
> That's cool. However, I don't have suitable rights to assign milestones to the issues.

Don't worry, I am doing it.
