+++ src/Visualization/VisualizationData.jl
@@ -1,0 +1,25 @@
+"""
+"""
+function visualization_data(trian::Triangulation; order=-1, nsubcells=-1, celldata=Dict(), cellfields=Dict())
+
+  if order == -1 && nsubcells == -1
+    # Use the given cells as visualization cells
+    f = (reffe) -> UnstructuredGrid(reffe)
+  elseif order != -1 && nsubcells == -1
+    # Use cells of given order as visualization cells
+    f = (reffe) -> UnstructuredGrid(LagrangianRefFE(Float64,get_polytope(reffe),order))
+  elseif order == -1 && nsubcells != -1
+    # Use use linear sub-cells with nsubcells per direction
+    f = (reffe) -> UnstructuredGrid(compute_reference_grid(reffe,nsubcells))
+  else
+    @unreachable "order and nsubcells kw-arguments can not be given at the same time"
+  end
+
+  ref_grids = map(f, get_reffes(trian))
+  visgrid = VisualizationGrid(trian,ref_grids)
+
+  cdata = _prepare_cdata(celldata,visgrid.sub_cell_to_cell)
+  pdata = _prepare_pdata(trian,cellfields,visgrid.cell_to_refpoints)
+
+  return (grid=visgrid, nodaldata=pdata, celldata=cdata)
+end
