+++ src/Visualization/Vtk.jl
@@ -394,34 +394,9 @@ end
 """
 """
 function createvtk(trian::Triangulation, filebase; order=-1, nsubcells=-1, celldata=Dict(), cellfields=Dict())
-
-  if order == -1 && nsubcells == -1
-    # Use the given cells as visualization cells
-
-    f = (reffe) -> UnstructuredGrid(reffe)
-
-  elseif order != -1 && nsubcells == -1
-    # Use cells of given order as visualization cells
-
-    f = (reffe) -> UnstructuredGrid(LagrangianRefFE(Float64,get_polytope(reffe),order))
-
-  elseif order == -1 && nsubcells != -1
-    # Use use linear sub-cells with nsubcells per direction
-
-    f = (reffe) -> UnstructuredGrid(compute_reference_grid(reffe,nsubcells))
-
-  else
-    @unreachable "order and nsubcells kw-arguments can not be given at the same time"
-  end
-
-  ref_grids = map(f, get_reffes(trian))
-  visgrid = VisualizationGrid(trian,ref_grids)
-
-  cdata = _prepare_cdata(celldata,visgrid.sub_cell_to_cell)
-  pdata = _prepare_pdata(trian,cellfields,visgrid.cell_to_refpoints)
-
-  create_vtk_file(visgrid,filebase,celldata=cdata,nodaldata=pdata)
-
+  visdata = visualization_data(trian; order=order,
+    nsubcells=nsubcells, celldata=celldata, cellfields=cellfields)
+  create_vtk_file(visdata.grid,filebase,celldata=visdata.celldata,nodaldata=visdata.nodaldata)
 end
 
 function _prepare_pdata(trian,cellfields,samplingpoints)
