### 2020-08-11 b7a98b9
polish VisualizationData tests
