### 2020-08-11 ba2c7fa
add VisualizationData struct
