+++ test/VisualizationTests/VisualizationDataTests.jl
@@ -30,5 +30,6 @@ nodes_per_cell = num_nodes(first(get_reffes(visdata.grid)))
 @test size(visdata.nodaldata["u"]) == (ncells * nodes_per_cell,)
 @test size(visdata.celldata["cell"]) == (ncells,)
 @test visdata.grid isa Gridap.Visualization.VisualizationGrid
+@test visdata isa Visualization.VisualizationData
 
 end#module
