### 2020-08-11 17ef105
Finishing to polish block of block case
