+++ test/FieldsTests/FieldOperationsTestsNew.jl
@@ -1,0 +1,281 @@
+module FieldOperationsTests
+
+using BlockArrays
+using Gridap.Arrays
+using Gridap.Fields
+using Gridap.Fields: IntKernel
+using Gridap.Fields: FieldOpKernel
+using Gridap.Fields: trialize_array_of_matrices
+using Gridap.TensorValues
+
+using Test
+using FillArrays
+using LinearAlgebra
+
+l = 10
+np = 3
+ndofs = 4
+
+al = [rand(np,ndofs) for k in 1:l]
+bl = [rand(np) for k in 1:l]
+cl = [rand(np,ndofs) for k in 1:l]
+
+f(a,b) = 2*a-b*a
+dl = apply(FieldOpKernel(f),bl,bl)
+test_array(dl,map((a,b)->f.(a,b),bl,bl))
+
+f(a,b) = 2*a-b*a
+dl = apply(FieldOpKernel(f),al,bl)
+test_array(dl,map((a,b)->f.(a,b),al,bl))
+
+f(a,b) = 2*a-b
+dl = apply(FieldOpKernel(f),al,cl)
+test_array(dl,map((a,b)->f.(a,b),al,cl))
+
+f(a,b,c) = b*(2*a-c)
+dl = apply(FieldOpKernel(f),al,bl,cl)
+test_array(dl,map((a,b,c)->f.(a,b,c),al,bl,cl))
+
+dl = trialize_array_of_matrices(al)
+test_array(dl,map(a->reshape(a,(size(a,1),1,size(a,2))),al))
+
+atl = trialize_array_of_matrices(al)
+ctl = trialize_array_of_matrices(cl)
+
+f(a,b) = 2*a*b
+dl = apply(FieldOpKernel(f),al,atl)
+test_array(dl,map((a,b)->f.(a,b),al,atl))
+
+f(a,c,at,ct) = 2*(a+c)*(2*at-ct)
+dl = apply(FieldOpKernel(f),al,cl,atl,ctl)
+test_array(dl,map((a,c,at,ct)->f.(a,c,at,ct),al,cl,atl,ctl))
+
+@test size(dl[1]) == (np,ndofs,ndofs)
+
+# Blocks
+
+blocks = (al,)
+blockids = [(1,1)]
+axs_i = (blockedrange([np]),blockedrange([ndofs,ndofs]))
+axs = Fill(axs_i,l)
+aBl = VectorOfBlockArrayCoo(blocks,blockids,axs)
+
+blocks = (cl,)
+blockids = [(1,2)]
+cBl = VectorOfBlockArrayCoo(blocks,blockids,axs)
+
+atBl = trialize_array_of_matrices(aBl)
+ctBl = trialize_array_of_matrices(cBl)
+@test isa(atBl,VectorOfBlockArrayCoo)
+
+f(a) = 2*a
+dl = apply(FieldOpKernel(f),aBl)
+@test isa(dl,VectorOfBlockArrayCoo)
+test_array(dl,map((a)->f.(a),aBl))
+
+f(a,b) = 2*a + a*b
+dl = apply(FieldOpKernel(f),aBl,bl)
+@test isa(dl,VectorOfBlockArrayCoo)
+test_array(dl,map((a,b)->f.(a,b),aBl,bl))
+
+f(b,a) = 2*a + a*b
+dl = apply(FieldOpKernel(f),bl,atBl)
+@test isa(dl,VectorOfBlockArrayCoo)
+test_array(dl,map((a,b)->f.(a,b),bl,atBl))
+
+f(b,a) = 2*a + b
+dl = apply(FieldOpKernel(f),atBl,ctBl)
[…]

[…]
