### 2020-08-11 gridap/Gridap.jl#353 @santiagobadia
Due to the fact that the paper must be very short, we would like to skip performance analyses here. As stated in the JOSS rules, a performance analysis is only needed to support a performance claims in the article. **If you consider that this is a performance claim, we can eliminate it.** In any case, the statement is true, since the storage of the cell-wise element matrices replicates the entries related to degrees of freedom on the skeleton of the mesh.

The memory efficiency of `Gridap`, and even more important, its time-to-solution, compared to other software packages like deal.ii, Fenics, or JuliaFEM is very interesting, and we would like to evaluate this in the near future. But it is out of the scope of this article. So, I would not relate the performance analysis with the current JOSS paper.
