### 2020-08-11 4d7287b
move vtk independent functions to VisualizationData.jl file
