+++ src/Visualization/Vtk.jl
@@ -399,165 +399,6 @@ function createvtk(trian::Triangulation, filebase; order=-1, nsubcells=-1, celld
   create_vtk_file(visdata.grid,filebase,celldata=visdata.celldata,nodaldata=visdata.nodaldata)
 end
 
-function _prepare_pdata(trian,cellfields,samplingpoints)
-  pdata = Dict()
-  for (k,v) in cellfields
-    _v = CellField(v,trian)
-    pdata[k], = _prepare_node_to_coords(evaluate(_v,samplingpoints))
-  end
-  pdata
-end
-
-function _prepare_cdata(celldata,sub_cell_to_cell)
-  cdata = Dict()
-  for (k,v) in celldata
-    cell_to_a = collect(v)
-    T = eltype(cell_to_a)
-    sub_cell_to_a = zeros(T,length(sub_cell_to_cell))
-    _fill_sub_cell_to_a!(sub_cell_to_a,cell_to_a,sub_cell_to_cell)
-    cdata[k] = sub_cell_to_a
-  end
-  k2 = "cell"
-  @assert ! haskey(cdata,k2) "cell is a reserved key"
-  cdata[k2] = sub_cell_to_cell
-  cdata
-end
-
-function _fill_sub_cell_to_a!(sub_cell_to_a,cell_to_a,sub_cell_to_cell)
-  for (sub_cell, cell) in enumerate(sub_cell_to_cell)
-    sub_cell_to_a[sub_cell] = cell_to_a[cell]
-  end
-end
-
-struct VisualizationGrid{Dc,Dp,T} <: Grid{Dc,Dp}
-  sub_grid::UnstructuredGrid{Dc,Dp,T,false}
-  sub_cell_to_cell::Vector{Int}
-  cell_to_refpoints::CompressedArray{Vector{Point{Dc,T}},1,Vector{Vector{Point{Dc,T}}},Vector{Int8}}
-end
-
-get_reffes(g::VisualizationGrid) = get_reffes(g.sub_grid)
-
-get_cell_type(g::VisualizationGrid) = get_cell_type(g.sub_grid)
-
-get_node_coordinates(g::VisualizationGrid) = get_node_coordinates(g.sub_grid)
-
-get_cell_nodes(g::VisualizationGrid) = get_cell_nodes(g.sub_grid)
-
-function VisualizationGrid(trian::Triangulation, ref_grids::Vector{<:UnstructuredGrid})
-
-  cell_to_ctype = collect1d(get_cell_type(trian))
-  ctype_to_refpoints = map(get_node_coordinates, ref_grids)
-  cell_to_refpoints = CompressedArray(ctype_to_refpoints,cell_to_ctype)
-  cell_map = get_cell_map(trian)
-  cell_to_points = evaluate(cell_map, cell_to_refpoints)
-
-  node_to_coords, cell_to_offset = _prepare_node_to_coords(cell_to_points)
-
-  ctype_to_scell_to_snodes = map(get_cell_nodes,ref_grids)
-
-  sub_cell_to_nodes, sub_cell_to_cell = _prepare_sub_cell_to_nodes(
-    cell_to_ctype,ctype_to_scell_to_snodes,cell_to_offset)
-
-  ctype_to_reffes = map(get_reffes,ref_grids)
-  ctype_to_scell_type = map(get_cell_type,ref_grids)
-  sctype_to_reffe, sub_cell_to_sctype = _prepare_sctype_to_reffe(
-    ctype_to_reffes,ctype_to_scell_type,cell_to_ctype)
-
-  sub_grid = UnstructuredGrid(
-    node_to_coords,
-    sub_cell_to_nodes,
-    sctype_to_reffe,
[…]
