+++ src/Visualization/VisualizationData.jl
@@ -1,7 +1,52 @@
+struct VisualizationGrid{Dc,Dp,T} <: Grid{Dc,Dp}
+  sub_grid::UnstructuredGrid{Dc,Dp,T,false}
+  sub_cell_to_cell::Vector{Int}
+  cell_to_refpoints::CompressedArray{Vector{Point{Dc,T}},1,Vector{Vector{Point{Dc,T}}},Vector{Int8}}
+end
+
+get_reffes(g::VisualizationGrid) = get_reffes(g.sub_grid)
+
+get_cell_type(g::VisualizationGrid) = get_cell_type(g.sub_grid)
+
+get_node_coordinates(g::VisualizationGrid) = get_node_coordinates(g.sub_grid)
+
+get_cell_nodes(g::VisualizationGrid) = get_cell_nodes(g.sub_grid)
+
+function VisualizationGrid(trian::Triangulation, ref_grids::Vector{<:UnstructuredGrid})
+
+  cell_to_ctype = collect1d(get_cell_type(trian))
+  ctype_to_refpoints = map(get_node_coordinates, ref_grids)
+  cell_to_refpoints = CompressedArray(ctype_to_refpoints,cell_to_ctype)
+  cell_map = get_cell_map(trian)
+  cell_to_points = evaluate(cell_map, cell_to_refpoints)
+
+  node_to_coords, cell_to_offset = _prepare_node_to_coords(cell_to_points)
+
+  ctype_to_scell_to_snodes = map(get_cell_nodes,ref_grids)
+
+  sub_cell_to_nodes, sub_cell_to_cell = _prepare_sub_cell_to_nodes(
+    cell_to_ctype,ctype_to_scell_to_snodes,cell_to_offset)
+
+  ctype_to_reffes = map(get_reffes,ref_grids)
+  ctype_to_scell_type = map(get_cell_type,ref_grids)
+  sctype_to_reffe, sub_cell_to_sctype = _prepare_sctype_to_reffe(
+    ctype_to_reffes,ctype_to_scell_type,cell_to_ctype)
+
+  sub_grid = UnstructuredGrid(
+    node_to_coords,
+    sub_cell_to_nodes,
+    sctype_to_reffe,
+    sub_cell_to_sctype,
+    Val{false}())
+
+  VisualizationGrid(sub_grid,sub_cell_to_cell,cell_to_refpoints)
+
+end
+
 struct VisualizationData
-    grid::VisualizationGrid
-    celldata
-    nodaldata
+  grid::VisualizationGrid
+  celldata
+  nodaldata
 end
 
 """
@@ -27,5 +72,120 @@ function visualization_data(trian::Triangulation; order=-1, nsubcells=-1, cellda
   cdata = _prepare_cdata(celldata,visgrid.sub_cell_to_cell)
   pdata = _prepare_pdata(trian,cellfields,visgrid.cell_to_refpoints)
 
-  return VisualizationData(visgrid, nodaldata, celldata)
+  return VisualizationData(visgrid, cdata, pdata)
+end
+
+
+function _prepare_node_to_coords(cell_to_points)
+  cell_to_offset = zeros(Int,length(cell_to_points))
+  P = eltype(eltype(cell_to_points))
+  node_to_coords = P[]
+  cache = array_cache(cell_to_points)
+  _prepare_node_to_coords_fill!(node_to_coords,cell_to_offset,cache,cell_to_points)
+  (node_to_coords, cell_to_offset)
+end
+
+function _prepare_node_to_coords_fill!(node_to_coords,cell_to_offset,cache,cell_to_points)
[…]
