### 2020-08-11 c8e62ea
fix VisualizationTests
