+++ test/VisualizationTests/runtests.jl
@@ -2,10 +2,10 @@ module VisualizationTests
 
 using Test
 
+@testset "visualization_data" begin include("VisualizationDataTests.jl") end
+
 @testset "PrintOpTrees" begin include("PrintOpTreesTests.jl") end
 
 @testset "Vtk" begin include("VtkTests.jl") end
-
-@testset "visualization_data" begin include("VisualizationData.jl") end
 
 end # module
