### 2020-08-11 #2520 @KristofferC
With the exception of the typesetting of the code part (which confused me for a while), I think this paper/software is well written and I have checked all the boxes (and thus recommend accepting this paper).
