### 2020-08-11 #2520 @fverdugo
@Kevin-Mattheus-Moerman Have you asked to the whedon technical staff if it is possible to hook these lines
```tex
\usepackage[utf8x]{inputenc}
\usepackage{textgreek}
```

Into your pandoc template so that the nabla operators are displayed correctly ?
