### 2020-08-11 7c6d8d7
Added operate_fields and operate_arrays_of_fields
