### 2020-08-11 1632b18
Update NEWS.md [skip ci]
