+++ src/Fields/Fields.jl
@@ -56,6 +56,7 @@ export evaluate_fields!
 export field_gradients
 export field_array_cache
 export evaluate_field_array
+export evaluate_field_arrays
 export field_array_gradient
 export gradient_type
 export curl
