### 2020-08-12 db3897b
Implemented Gradient rules for fields at local level
