### 2020-08-12 6642554
Implemented gradient rules at global level
