### 2020-08-12 b26e391
Gradient of VectorOfBlockBasisCoo
