### 2020-08-12 9f4b103
Moving VectorOfBlockBasisCoo to the library
