### 2020-08-12 9d79612
Some clean up
