### 2020-08-12 gridap/Gridap.jl#353 @TeroFrondelius
Naturally, my motivation is to get the benchmark in the paper, to get a citation to JuliaFEM paper.
