### 2020-08-12 e90d466
Deactivating code down stream and using the new field operations
