+++ test/FieldsTests/FieldOperationsTests.jl
@@ -1,124 +1,365 @@
 module FieldOperationsTests
 
+using BlockArrays
 using Gridap.Arrays
 using Gridap.Fields
-using Gridap.Fields: MockField, MockBasis
+using Gridap.Fields: IntKernel
+using Gridap.Fields: FieldOpKernel
+using Gridap.Fields: trialize_basis_value
 using Gridap.TensorValues
+using Gridap.Fields: MockField, MockBasis
+
+using Test
 using FillArrays
-
-p1 = Point(2,2)
-p2 = Point(4,2)
-p3 = Point(1,3)
-p4 = Point(5,2)
-x = [p1,p2,p3,p4]
-np = length(x)
-
-va = 3.0
-vb = 3.3
-d = 2
-fa = MockField{d}(va)
-fb = MockField{d}(vb)
-fax = evaluate(fa,x)
-fbx = evaluate(fb,x)
-∇fax = evaluate(gradient(fa),x)
-∇fbx = evaluate(gradient(fb),x)
-
-wa = 4.5
-wb = 4.5
-ndofa = 8
-ndofb = 5
-ba = MockBasis{d}(wa,ndofa)
-bb = MockBasis{d}(wb,ndofb)
-bax = evaluate(ba,x)
-bbx = evaluate(bb,x)
-∇bax = evaluate(gradient(ba),x)
-∇bbx = evaluate(gradient(bb),x)
+using LinearAlgebra
 
 l = 10
-ax = fill(x,l)
-afa = Fill(fa,l)
-afb = Fill(fb,l)
-aba = Fill(ba,l)
-abb = Fill(bb,l)
+np = 3
+ndofs = 4
 
-g = field_operation(inner,fa,fb)
-gx = [inner(ai,bj) for (ai,bj) in zip(fax,fbx)]
-test_field(g,x,gx)
+# Operations between fields and array of fields
 
-ag = field_array_operation(inner,afa,afb)
-agx = fill(gx,l)
-test_array_of_fields(ag,ax,agx)
+p = Point(1,2)
+x = fill(p,np)
+z = 2.0
 
-struct FieldPlaceHolder <: Field end
+v = VectorValue(3.0,1.5)
+w = VectorValue(3.4,3.5)
+test_basis = MockBasis{2}(v,ndofs)
+test_basis_2 = MockBasis{2}(w,ndofs)
 
-ag = field_array_operation(FieldPlaceHolder,inner,fill(fa,l),afb)
-test_array(evaluate_field_array(ag,ax),agx)
+t1x = evaluate(test_basis,x)
+t2x = evaluate(test_basis_2,x)
+∇t1x = evaluate(∇(test_basis),x)
+∇t2x = evaluate(∇(test_basis_2),x)
 
-g = field_operation(inner,ba,fb)
-gx = zeros(size(bax))
-for p in 1:np
-  for i in 1:ndofa
-    gx[p,i] = inner(bax[p,i],fbx[p])
-  end
-end
-test_field(g,x,gx)
+b = operate_fields(+,test_basis,test_basis_2)
+r = broadcast(+,t1x,t2x)
+∇r = broadcast(+,∇t1x,∇t2x)
+test_field(b,x,r,grad=∇r)
 
-ag = field_array_operation(inner,aba,afb)
-agx = fill(gx,l)
-test_array_of_fields(ag,ax,agx)
+b = operate_fields(-,test_basis,test_basis_2)
+r = broadcast(-,t1x,t2x)
[…]
