+++ test/runtests.jl
@@ -16,26 +16,26 @@ using Test
 
 @time @testset "Fields" begin include("FieldsTests/runtests.jl") end
 
-@time @testset "Polynomials" begin include("PolynomialsTests/runtests.jl") end
-
-@time @testset "Integration" begin include("IntegrationTests/runtests.jl") end
-
-@time @testset "ReferenceFEs" begin include("ReferenceFEsTests/runtests.jl") end
-
-@time @testset "Geometry" begin include("GeometryTests/runtests.jl") end
-
-@time @testset "FESpaces" begin include("FESpacesTests/runtests.jl") end
-
-@time @testset "MultiField" begin include("MultiFieldTests/runtests.jl") end
-
-@time @testset "Visualization" begin include("VisualizationTests/runtests.jl") end
-
-@time @testset "Issues" begin
-    include("issue_349.jl")
-end
-
-include("GridapTests/runtests.jl")
-
-include("../bench/runbenchs.jl")
+#@time @testset "Polynomials" begin include("PolynomialsTests/runtests.jl") end
+#
+#@time @testset "Integration" begin include("IntegrationTests/runtests.jl") end
+#
+#@time @testset "ReferenceFEs" begin include("ReferenceFEsTests/runtests.jl") end
+#
+#@time @testset "Geometry" begin include("GeometryTests/runtests.jl") end
+#
+#@time @testset "FESpaces" begin include("FESpacesTests/runtests.jl") end
+#
+#@time @testset "MultiField" begin include("MultiFieldTests/runtests.jl") end
+#
+#@time @testset "Visualization" begin include("VisualizationTests/runtests.jl") end
+#
+#@time @testset "Issues" begin
+#    include("issue_349.jl")
+#end
+#
+#include("GridapTests/runtests.jl")
+#
+#include("../bench/runbenchs.jl")
 
 end # module
