+++ src/Fields/FieldOperationsNew.jl
@@ -1,394 +1,0 @@
-
-# Operate fields at local level
-
-"""
-"""
-function operate_fields(op::Function,args...)
-  k = FieldOpKernel(op)
-  apply_kernel_to_field(k,args...)
-end
-
-# Operate fields at global level
-
-"""
-"""
-function operate_arrays_of_fields(op::Function,args...)
-  k = FieldOpKernel(op)
-  apply_to_field_array(k,args...)
-end
-
-function operate_arrays_of_fields(::Type{T},op::Function,args...) where T
-  k = FieldOpKernel(op)
-  apply_to_field_array(T,k,args...)
-end
-
-
-# Operations between field values
-
-struct FieldOpKernel{T} <: Kernel
-  op::T
-end
-
-# FieldOpKernel does broadcast by default. It will work always assuming that the trial bases have shape (np,1,ndof)
-# but perhaps inefficient for blocked matrices.
-# In any case, optimizations for block matrices will be done at the global level (for all cells)
-# instead of at the cell level in this kernel.
-# In other words, we can assume that this kernel receives standard non-blocked arrays in practice.
-
-function kernel_cache(k::FieldOpKernel,args...)
-  bk = bcast(k.op)
-  kernel_cache(bk,args...)
-end
-
-@inline function apply_kernel!(cache,k::FieldOpKernel,args...)
-  bk = bcast(k.op)
-  apply_kernel!(cache,bk,args...)
-end
-
-# Define gradients at local and global level
-
-function apply_kernel_gradient(k::FieldOpKernel,args...)
-  @notimplemented "The gradient of the result of operation $(k.op) is not yet implemented."
-end
-
-function apply_gradient(k::Valued{<:FieldOpKernel},args...)
-  @notimplemented "The gradient of the result of operation $(k.op) is not yet implemented."
-end
-
-for op in (:+,:-)
-  @eval begin
-
-    # Local level
-
-    function apply_kernel_gradient(k::FieldOpKernel{typeof($op)},a,b)
-      ga = field_gradient(a)
-      gb = field_gradient(b)
-      apply_kernel_to_field(k,ga,gb)
-    end
-
-    function apply_kernel_gradient(k::FieldOpKernel{typeof($op)},a)
-      ga = field_gradient(a)
-      apply_kernel_to_field(k,ga)
-    end
-
-    function apply_kernel_gradient(k::FieldOpKernel{typeof($op)},a::Number,b)
-      gb = field_gradient(b)
-      apply_kernel_to_field(k,gb)
-    end
-
-    function apply_kernel_gradient(k::FieldOpKernel{typeof($op)},b,a::Number)
-      gb = field_gradient(b)
-      apply_kernel_to_field(k,gb)
-    end
-
-    # Global level
-
-    function apply_gradient(k::Valued{FieldOpKernel{typeof($op)}},a,b)
-      ga = field_array_gradient(a)
-      gb = field_array_gradient(b)
-      apply(k,ga,gb)
-    end
-
-    function apply_gradient(k::Valued{FieldOpKernel{typeof($op)}},a)
-      ga = field_array_gradient(a)
-      apply(k,ga)
-    end
-
[…]
