+++ src/Gridap.jl
@@ -42,20 +42,20 @@ include("TensorValues/TensorValues.jl")
 
 include("Fields/Fields.jl")
 
-include("Polynomials/Polynomials.jl")
-
-include("Integration/Integration.jl")
-
-include("ReferenceFEs/ReferenceFEs.jl")
-
-include("Geometry/Geometry.jl")
-
-include("FESpaces/FESpaces.jl")
-
-include("MultiField/MultiField.jl")
-
-include("Visualization/Visualization.jl")
-
-include("Exports.jl")
+#include("Polynomials/Polynomials.jl")
+#
+#include("Integration/Integration.jl")
+#
+#include("ReferenceFEs/ReferenceFEs.jl")
+#
+#include("Geometry/Geometry.jl")
+#
+#include("FESpaces/FESpaces.jl")
+#
+#include("MultiField/MultiField.jl")
+#
+#include("Visualization/Visualization.jl")
+#
+#include("Exports.jl")
 
 end # module
