### 2020-08-12 2f5458b
Some changes in VectorOfBlockBasisCoo
