### 2020-08-12 24dd836
Some changes in diff ops for arrays of fields
