+++ paper.bib
@@ -17,6 +17,17 @@ title = {{FEMPAR: An Object-Oriented Parallel Finite Element Framework}},
 volume = {25},
 year = {2018}
 }
+@article{frondelius2017juliafem,
+  title={Julia{FEM} - open source solver for both industrial and academia usage},
+  volume={50}, 
+  url={https://rakenteidenmekaniikka.journal.fi/article/view/64224},
+  DOI={10.23998/rm.64224},
+  number={3},
+  journal={Rakenteiden Mekaniikka},
+  author={Frondelius, Tero and Aho, Jukka},
+  year={2017},
+  pages={229-233}
+}
 @article{Innes2018,
 author = {Innes, Mike},
 doi = {10.21105/joss.00602},
