+++ paper.md
@@ -27,7 +27,7 @@ Gridap is a new Finite Element (FE) framework, exclusively written in the Julia 
 
 The main motivation behind Gridap is to find an improved balance between computational performance, user-experience, and work-flow productivity when working with FE libraries. Previous FE frameworks, e.g., FEniCS [@Alnaes2015] or Deal.II [@Bangerth2007] usually provides a high-level user front-end to facilitate the use of the library and a computational back-end to achieve performance. The user front-end is usually programmable in an interpreted language like Python, whereas the computational back-end is usually coded in a compiled language like C/C++ or Fortran. Users can benefit from the high-level front-end (i.e., for rapid prototyping) and simultaneously enjoy the performance of the compiled back-end. This approach reaches a compromise between performance and productivity when the back-end provides all the functionality required by the user. However, it does not satisfactorily address the needs of researchers on numerical methods willing to extend the library with new techniques or features. These extensions usually need to be done at the level of the computational back-end for performance reasons. Thus, the researcher is forced to develop a new code in a compiled language like C/C++ instead of benefiting from the productivity of scripting languages like Python, incurring serious productivity losses. In order to overcome this limitation, Gridap is fully implemented in the Julia programming language [@Bezanson2017]. Julia combines the performance of compiled languages with the productivity of interpreted ones by using type inference and just-in-time compilation to generate fast code. As a result, there is no need to use two different languages to write low-level performance code and high-level user interfaces. In addition, writing a FE library in Julia also allows one to leverage the feature-rich ecosystem of Julia libraries and exploit its excellent package manager. It permits a seamless coupling of Gridap with application-specific libraries, like optimization [@Dunning2017], an approximation of ordinary differential equations [@Rackauckas2017], or data science [@Innes2018], which can certainly boost the capabilities of a FE solver.
 
-Another major feature of Gridap is that it is not a simple Julia translation of a standard object-oriented FE code. There are other FE libraries written in Julia that have been inspired by standard FE frameworks, see, e.g., JuliaFEM [frondelius2017juliafem] or JuAFEM [@Carlsson], whose interface resembles Deal.II.  In contrast,  Gridap adopts a novel software design that allows one to manipulate different types of data associated with the cells of the computational mesh in a convenient way. For instance, one can build an object representing the elemental matrices for all cells in the mesh using high-level API calls, without explicitly writing any for-loop. These objects representing data for all cells of the mesh are usually _lazy_, meaning that the underlying data is never stored for all cells in the mesh simultaneously. Instead, the value for a specific cell is computed on-the-fly when needed, which certainly reduces memory requirements.  This software design allows the library developers to hide assembly loops and other core computations from the user-code, leading to a very compact, user-friendly, syntax, while providing a high degree of flexibility for users to define their own weak forms.  A Poisson or Stokes problem can be solved with Gridap in 10-20 lines of code, as this example for the Poisson equation shows:
[…]
