+++ paper.bib
@@ -117,6 +117,11 @@ author = {Carlsson, Kristoffer},
 title = {{JuaFEM git repository}},
 url = {https://github.com/KristofferC/JuAFEM.jl}
 }
+@misc{Krysl,
+author = {Krysl, Petr},
+title = {{FinEtools git repository}},
+url = {https://github.com/PetrKryslUCSD/FinEtools.jl}
+}
 @book{Johnson2009,
 author = {Johnson, Claes},
 isbn = {048646900X},
