### 2020-08-12 gridap/Gridap.jl#353 @santiagobadia
Now we have added a citation to your paper.

As commented above, we are not making performance claims against your software or any other of the software projects we have cited, so we are not going to include performance analyses.

Please, make a decision based on this, and inform the editor @Kevin-Mattheus-Moerman about your final decision.
