### 2020-08-13 71c06b1
reactivating Polynomials, Integration, ReferenceFEs
