+++ test/runtests.jl
@@ -16,12 +16,12 @@ using Test
 
 @time @testset "Fields" begin include("FieldsTests/runtests.jl") end
 
-#@time @testset "Polynomials" begin include("PolynomialsTests/runtests.jl") end
-#
-#@time @testset "Integration" begin include("IntegrationTests/runtests.jl") end
-#
-#@time @testset "ReferenceFEs" begin include("ReferenceFEsTests/runtests.jl") end
-#
+@time @testset "Polynomials" begin include("PolynomialsTests/runtests.jl") end
+
+@time @testset "Integration" begin include("IntegrationTests/runtests.jl") end
+
+@time @testset "ReferenceFEs" begin include("ReferenceFEsTests/runtests.jl") end
+
 #@time @testset "Geometry" begin include("GeometryTests/runtests.jl") end
 #
 #@time @testset "FESpaces" begin include("FESpacesTests/runtests.jl") end
