### 2020-08-13 gridap/Gridap.jl#353 @TeroFrondelius
Thanks. I close this issue because it's joss paper related. We will need to open new issues both in Gridap and JuliaFEM repos or discourse topic about FEM benchmarking. Maybe the benchmark could be a separate repo, where everyone could upload their model/results. Could we even run the benchmark on Travis CI or similar?
