### 2020-08-13 #2520 @TeroFrondelius
This is a well-written software/documentation/paper. I have raised a few minor issues which all have been addressed promptly by the authors. Therefore, I recommend accepting this paper. At the end, I want to thank you for the interesting experience of open/transparent peer-review.
