### 2020-08-13 3be6d58
better keyword argument checks for FESpace
