### 2020-08-13 0772161
add docstrings for residual and friends
