### 2020-08-13 a279d0d
Refactored CellFields
