+++ src/Gridap.jl
@@ -48,7 +48,7 @@ include("Integration/Integration.jl")
 
 include("ReferenceFEs/ReferenceFEs.jl")
 
-#include("Geometry/Geometry.jl")
+include("Geometry/Geometry.jl")
 
 #include("FESpaces/FESpaces.jl")
 #
