+++ src/Geometry/CellFields.jl
@@ -1,29 +1,33 @@
 
 """
-    abstract type CellFieldLike <: GridapType end
+    abstract type CellField <: GridapType end
+
+This is a plain array of fields plus some metadata.
+It is interpreted as a `Field` object on each cell of a computational
+mesh.
 """
-abstract type CellFieldLike <: GridapType end
+abstract type CellField <: GridapType end
 
 """
-    get_array(cf::CellFieldLike)
+    get_array(cf::CellField)
 """
-function get_array(cf::CellFieldLike)
+function get_array(cf::CellField)
   @abstractmethod
 end
 
 """
-    get_cell_map(cf::CellFieldLike)
+    get_cell_map(cf::CellField)
 """
-function get_cell_map(cf::CellFieldLike)
+function get_cell_map(cf::CellField)
   @abstractmethod
 end
 
 """
-This trait returns `Val{true}()` when the `CellFieldLike` is defined in a
+This trait returns `Val{true}()` when the `CellField` is defined in a
 reference finite element space, and `Val{false}()` when it is defined in the
 physical space
 """
-RefStyle(::Type{<:CellFieldLike}) = @notimplemented
+RefStyle(::Type{<:CellField}) = @notimplemented
 
 # We use duck typing here for all types marked with the RefStyle
 RefStyle(::Type) = @notimplemented
@@ -33,84 +37,17 @@ is_in_ref_space(::T) where T = is_in_ref_space(T)
 is_in_physical_space(::Type{T}) where T = !is_in_ref_space(T)
 is_in_physical_space(a::T) where T = !is_in_ref_space(T)
 
-to_ref_space(a::CellFieldLike) = _to_ref_space(a,RefStyle(a))
-_to_ref_space(a,::Val{true}) = a
-function _to_ref_space(a,::Val{false})
-  cell_map = get_cell_map(a)
-  array = compose(  get_array(a), cell_map  )
-  no = similar_object(a,array)
-  change_ref_style(no)
-end
-
-to_physical_space(a::CellFieldLike) = _to_physical_space(a,RefStyle(a))
-_to_physical_space(a,::Val{true}) = @notimplemented # and probably not doable in some cases
-_to_physical_space(a,::Val{false}) = a
-
-# Assumption : x ALWAIS defined in the reference space
-# In the future we can also add the RefStyle to x
-"""
-    evaluate(cf::CellFieldLike,x)
-"""
-function evaluate(cf::CellFieldLike,x::AbstractArray)
-  _evaluate(cf,x,RefStyle(cf))
-end
-
-function _evaluate(cf::CellFieldLike,x::AbstractArray,::Val{true})
-  evaluate_field_array(get_array(cf),x)
-end
-
-function _evaluate(cf::CellFieldLike,x::AbstractArray,::Val{false})
-  cm = get_cell_map(cf)
-  _x = evaluate(cm,x)
-  evaluate_field_array(get_array(cf),_x)
-end
+# Tester
 
 """
-    similar_object(cf::CellFieldLike,array::AbstractArray)
-"""
-function similar_object(cf::CellFieldLike,array::AbstractArray)
-  @abstractmethod
-end
-
-"""
-    similar_object(cf1::CellFieldLike,cf2::CellFieldLike,array::AbstractArray)
-"""
-function similar_object(cf1::CellFieldLike,cf2::CellFieldLike,array::AbstractArray)
-  @abstractmethod
-end
-
-"""
-   change_ref_style(cf::CellFieldLike)
-
[…]
