+++ src/Fields/Fields.jl
@@ -124,4 +124,6 @@ include("VectorsOfBlockBasisCoo.jl")
 
 include("DiffOperators.jl")
 
+include("UnimplementedFields.jl")
+
 end # module
