+++ src/MultiField/MultiFieldFESpaces.jl
@@ -309,11 +309,26 @@ function compute_field_offsets(f::MultiFieldFESpace)
   offsets
 end
 
+@deprecate(
+  interpolate(fs::MultiFieldFESpace, object),
+  interpolate(object, fs::MultiFieldFESpace)
+)
+
+@deprecate(
+  interpolate_everywhere(fs::MultiFieldFESpace, object),
+  interpolate_everywhere(object, fs::MultiFieldFESpace)
+)
+
+@deprecate(
+  interpolate_dirichlet(fs::MultiFieldFESpace, object),
+  interpolate_dirichlet(object, fs::MultiFieldFESpace)
+)
+
 """
 The resulting MultiFieldFEFunction is in the space (in particular it fulfills Dirichlet BCs
 even in the case that the given cell field does not fulfill them)
 """
-function interpolate(fe::MultiFieldFESpace,objects)
+function interpolate(objects, fe::MultiFieldFESpace)
   free_values = zero_free_values(fe)
   blocks = SingleFieldFEFunction[]
   for (field, (U,object)) in enumerate(zip(fe.spaces,objects))
@@ -328,7 +343,7 @@ end
 like interpolate, but also compute new degrees of freedom for the dirichlet component.
 The resulting MultiFieldFEFunction does not necessary belongs to the underlying space
 """
-function interpolate_everywhere(fe::MultiFieldFESpace,objects)
+function interpolate_everywhere(objects, fe::MultiFieldFESpace)
   free_values = zero_free_values(fe)
   blocks = SingleFieldFEFunction[]
   for (field, (U,object)) in enumerate(zip(fe.spaces,objects))
@@ -342,7 +357,7 @@ end
 
 """
 """
-function interpolate_dirichlet(fe::MultiFieldFESpace,objects)
+function interpolate_dirichlet(objects, fe::MultiFieldFESpace)
   free_values = zero_free_values(fe)
   blocks = SingleFieldFEFunction[]
   for (field, (U,object)) in enumerate(zip(fe.spaces,objects))
