### 2020-08-13 053e08c
More work in VectorOfBlockBasisCoo
