+++ test/FieldsTests/VectorsOfBlockBasisCooTests.jl
@@ -23,18 +23,25 @@ a = MockBasis{2}(v,ndofs2)
 b = MockBasis{2}(w,ndofs1)
 
 al = Fill(a,l)
+atl = trialize_array_of_bases(al)
 bl = Fill(b,l)
 xl = Fill(x,l)
 
 blocks = (al,)
 blockids = [(1,2)]
-ranges_i = blockedrange([ndofs1,ndofs2])
+ranges_i = (blockedrange([ndofs1,ndofs2]),)
 ranges = Fill(ranges_i,l)
 aBl = VectorOfBlockBasisCoo(blocks,blockids,ranges)
 
 blocks = (bl,)
 blockids = [(1,1)]
 bBl = VectorOfBlockBasisCoo(blocks,blockids,ranges)
+
+blocks = (atl,)
+blockids = [(1,1,2)]
+ranges_i = (blockedrange([1]),blockedrange([ndofs1,ndofs2]))
+ranges = Fill(ranges_i,l)
+atBl = VectorOfBlockBasisCoo(blocks,blockids,ranges)
 
 aBl_x = evaluate(aBl,xl)
 @test isa(aBl_x,VectorOfBlockArrayCoo)
@@ -43,6 +50,11 @@ aBl_x = evaluate(aBl,xl)
 
 bBl_x = evaluate(bBl,xl)
 @test isa(bBl_x,VectorOfBlockArrayCoo)
+
+atBl_x = evaluate(atBl,xl)
+@test isa(atBl_x,VectorOfBlockArrayCoo)
+@test isa(atBl_x.axes,Fill)
+@test atBl_x.axes[1] == (blockedrange([np]),blockedrange([1]),blockedrange([ndofs1,ndofs2]))
 
 atBl = trialize_array_of_bases(aBl)
 atBl_x = evaluate(atBl,xl)
@@ -71,6 +83,23 @@ cl_x = evaluate(cl,xl)
 @test is_nonzero_block(cl_x,Block(1,1,2))
 @test is_zero_block(cl_x,Block(1,2,1))
 @test is_nonzero_block(cl_x,Block(1,2,2))
+
+ids = [2,3,5,1,2]
+cj = reindex(aBl,ids)
+xj = reindex(xl,ids)
+@test isa(cj,VectorOfBlockBasisCoo)
+cj_x = evaluate(cj,xj)
+@test isa(cj_x.axes,Fill)
+@test cj_x.axes[1] == (blockedrange([np]),blockedrange([ndofs1,ndofs2]))
+@test length(cj_x) == length(ids)
+@test length(cj) == length(ids)
+
+v = VectorValue(1.0,2.0,3.0)
+f = MockField{2}(v)
+fl = Fill(f,l)
+
+cl = compose_field_arrays(aBl,fl)
+@test isa(cl,VectorOfBlockBasisCoo)
 
 #cache = array_cache(cl_x)
 #using BenchmarkTools
