### 2020-08-13 #2520 @PetrKryslUCSD
I am satisfied that the software and the paper meet the requirements for publication (with the exception of the symbols in the code, which are still not typeset correctly). I recommend publication, provided the typesetting errors can be resolved.
