### 2020-08-13 9ee3399
added insert_array_of_bases_in_block
