+++ src/Fields/VectorsOfBlockBasisCoo.jl
@@ -13,7 +13,7 @@ end
 struct VectorOfBlockBasisCoo <: AbstractVector{BlockBasisCoo}
   blocks::Tuple
   blockids
-  ranges
+  axes
 end
 
 Base.size(v::VectorOfBlockBasisCoo) = (length(first(v.blocks)) ,)
@@ -21,8 +21,52 @@ Base.getindex(v::VectorOfBlockBasisCoo,i::Integer) = BlockBasisCoo()
 
 function evaluate_field_array(v::VectorOfBlockBasisCoo,x::AbstractArray)
   blocks = evaluate_field_arrays(v.blocks,x)
-  axs = apply(_new_axes,x,v.ranges)
-  VectorOfBlockArrayCoo(blocks,v.blockids,axs)
+  axs = apply(_new_axes,x,v.axes)
+  blockids = map(i->(1,i...),v.blockids)
+  VectorOfBlockArrayCoo(blocks,blockids,axs)
+end
+
+function field_array_gradient(v::VectorOfBlockBasisCoo)
+  blocks = map(field_array_gradient,v.blocks)
+  VectorOfBlockBasisCoo(blocks,v.blockids,v.axes)
+end
+
+function reindex(v::VectorOfBlockBasisCoo,j_to_i::AbstractArray)
+  blocks = map(b->reindex(b,j_to_i),v.blocks)
+  axs = reindex(v.axes,j_to_i)
+  VectorOfBlockBasisCoo(blocks,v.blockids,axs)
+end
+
+function compose_field_arrays(v::VectorOfBlockBasisCoo,f)
+  blocks = map(b->compose_field_arrays(b,f),v.blocks)
+  VectorOfBlockBasisCoo(blocks,v.blockids,v.axes)
+end
+
+function insert_array_of_bases_in_block(a,ax1,ax2,i)
+  blocks = (a,)
+  blockids = _compute_blockids(eltype(ax1),i)
+  axs = apply(_cat_axes,ax1,ax2)
+
+  VectorOfBlockBasisCoo(blocks,blockids,axs)
+end
+
+function _compute_blockids(::Type{<:NTuple{1}},i)
+  [(i,)]
+end
+
+function _compute_blockids(::Type{<:NTuple{2}},i)
+  [(1,i)]
+end
+
+function _cat_axes(a::NTuple{1},b::NTuple{1})
+  (_blockedrange([a[1],b[1]]),)
+end
+
+function _cat_axes(a::NTuple{2},b::NTuple{2})
+  @assert length(a[1]) == 1
+  @assert length(b[1]) == 1
+  ran = (_blockedrange([a[2],b[2]]),)
+  _add_singleton_block(ran)
 end
 
 function _new_axes(x,ran::NTuple{N,<:BlockedUnitRange} where N)
@@ -36,19 +80,21 @@ function _new_axes(x,ran::NTuple{N,<:TwoLevelBlockedUnitRange} where N)
   (blockedrange([r]),ran...)
 end
 
-function field_array_gradient(v::VectorOfBlockBasisCoo)
-  blocks = map(field_array_gradient,v.blocks)
-  VectorOfBlockBasisCoo(blocks,v.blockids,v.ranges)
+function _add_singleton_block(ran::NTuple{N,<:BlockedUnitRange} where N)
+  (blockedrange([1]),ran...)
 end
 
-function reindex(v::VectorOfBlockBasisCoo,j_to_i::AbstractArray)
-  blocks = map(b->reindex(b,j_to_i),v.blocks)
-  ranges = reindex(v.ranges,j_to_i)
-  VectorOfBlockBasisCoo(blocks,v.blockids,ranges)
[…]
