### 2020-08-13 b549e15
First attempt to add code as figure
