+++ build_code_snipped.sh
@@ -1,0 +1,2 @@
+pdflatex -shell-escape code.tex
+pdfcrop code.pdf code.pdf
