### 2020-08-13 #2520 @fverdugo
@Kevin-Mattheus-Moerman I have added the code sniped as a figure. Now the code symbols are visualized correctly.

I would say that this fixes all issues.
