### 2020-08-13 gridap/Gridap.jl#335 @fverdugo
@Kevin-Mattheus-Moerman
@PetrKryslUCSD

I have added the code sniped as a figure. Now the code symbols are visualized correctly.

https://raw.githubusercontent.com/openjournals/joss-papers/joss.02520/joss.02520/10.21105.joss.02520.pdf

You can close the issue, please.
