+++ src/MultiField/MultiFieldFESpaces.jl
@@ -333,7 +333,7 @@ function interpolate(objects, fe::MultiFieldFESpace)
   blocks = SingleFieldFEFunction[]
   for (field, (U,object)) in enumerate(zip(fe.spaces,objects))
     free_values_i = restrict_to_field(fe,free_values,field)
-    uhi = interpolate!(free_values_i,U,object)
+    uhi = interpolate!(object, free_values_i,U)
     push!(blocks,uhi)
   end
   MultiFieldFEFunction(free_values,fe,blocks)
@@ -349,7 +349,7 @@ function interpolate_everywhere(objects, fe::MultiFieldFESpace)
   for (field, (U,object)) in enumerate(zip(fe.spaces,objects))
     free_values_i = restrict_to_field(fe,free_values,field)
     dirichlet_values_i = zero_dirichlet_values(U)
-    uhi = interpolate_everywhere!(free_values_i,dirichlet_values_i,U,object)
+    uhi = interpolate_everywhere!(object, free_values_i,dirichlet_values_i,U)
     push!(blocks,uhi)
   end
   MultiFieldFEFunction(free_values,fe,blocks)
@@ -363,7 +363,7 @@ function interpolate_dirichlet(objects, fe::MultiFieldFESpace)
   for (field, (U,object)) in enumerate(zip(fe.spaces,objects))
     free_values_i = restrict_to_field(fe,free_values,field)
     dirichlet_values_i = zero_dirichlet_values(U)
-    uhi = interpolate_dirichlet!(free_values_i,dirichlet_values_i,U,object)
+    uhi = interpolate_dirichlet!(object, free_values_i,dirichlet_values_i,U)
     push!(blocks,uhi)
   end
   MultiFieldFEFunction(free_values,fe,blocks)
