+++ test/FESpacesTests/CDLagrangianFESpacesTests.jl
@@ -23,7 +23,7 @@ test_single_field_fe_space(V)
 u(x) = x
 
 U = TrialFESpace(V,u)
-uh = interpolate(U,u)
+uh = interpolate(u,U)
 e = u - uh
 trian = Triangulation(model)
 quad = CellQuadrature(trian,order)
@@ -33,7 +33,7 @@ el2 = sqrt(sum(integrate(inner(e,e),trian,quad)))
 reffe = LagrangianRefFE(T,QUAD,2)
 V = FESpace(model=model,reffe=reffe,conformity=CDConformity((CONT,DISC)))
 U = TrialFESpace(V,u)
-uh = interpolate(U,u)
+uh = interpolate(u,U)
 e = u - uh
 el2 = sqrt(sum(integrate(inner(e,e),trian,quad)))
 @test el2 < 1.0e-10
@@ -41,7 +41,7 @@ el2 = sqrt(sum(integrate(inner(e,e),trian,quad)))
 reffe = LagrangianRefFE(T,QUAD,(2,1))
 V = FESpace(model=model,reffe=reffe,conformity=CDConformity((DISC,CONT)))
 U = TrialFESpace(V,u)
-uh = interpolate(U,u)
+uh = interpolate(u,U)
 e = u - uh
 el2 = sqrt(sum(integrate(inner(e,e),trian,quad)))
 @test el2 < 1.0e-10
@@ -50,7 +50,7 @@ reffe = LagrangianRefFE(T,QUAD,(2,0))
 V = FESpace(model=model,reffe=reffe,conformity=CDConformity((CONT,DISC)))
 u(x) = VectorValue(x[1],0.0)
 U = TrialFESpace(V,u)
-uh = interpolate(U,u)
+uh = interpolate(u,U)
 e = u - uh
 el2 = sqrt(sum(integrate(inner(e,e),trian,quad)))
 @test el2 < 1.0e-10
@@ -59,7 +59,7 @@ reffe = LagrangianRefFE(T,QUAD,(2,0))
 V = FESpace(model=model,reffe=reffe,conformity=CDConformity((DISC,DISC)))
 u(x) = VectorValue(x[1],0.0)
 U = TrialFESpace(V,u)
-uh = interpolate(U,u)
+uh = interpolate(u,U)
 e = u - uh
 el2 = sqrt(sum(integrate(inner(e,e),trian,quad)))
 @test el2 < 1.0e-10
