+++ test/FESpacesTests/DivConformingFESpacesTests.jl
@@ -25,7 +25,7 @@ test_single_field_fe_space(V)
 
 U = TrialFESpace(V,u)
 
-uh = interpolate(U,u)
+uh = interpolate(u,U)
 
 e = u - uh
 
@@ -42,7 +42,7 @@ test_single_field_fe_space(V)
 
 U = TrialFESpace(V,u)
 
-uh = interpolate(U,u)
+uh = interpolate(u,U)
 
 el2 = sqrt(sum(integrate(inner(e,e),trian,quad)))
 @test el2 < 1.0e-10
