### 2020-08-14 112d3de
add interpolation note to NEWS.md
