+++ src/Geometry/CellFields.jl
@@ -23,6 +23,13 @@ function get_cell_map(cf::CellField)
 end
 
 """
+    get_cell_axes(cf::CellField)
+"""
+function get_cell_axes(cf::CellField)
+  @abstractmethod
+end
+
+"""
 This trait returns `Val{true}()` when the `CellField` is defined in a
 reference finite element space, and `Val{false}()` when it is defined in the
 physical space
@@ -62,6 +69,7 @@ function test_cell_field(cf::CellField,x::AbstractArray,b::AbstractArray,pred=(=
   @test get_array(_cf) === get_array(cf)
   @test is_in_ref_space(cf) == !is_in_ref_space(_cf)
   @test is_in_physical_space(cf) == !is_in_physical_space(_cf)
+  @test isa(get_cell_axes(cf),AbstractArray{<:Tuple})
 end
 
 # Concrete implementation
@@ -71,16 +79,22 @@ struct GenericCellField{R} <: CellField
       array::AbstractArray
       cell_map::AbstractArray
       ref_trait::Val{R}
+      cell_axes::AbstractArray
     end
 """
 struct GenericCellField{R} <: CellField
   array::AbstractArray
   cell_map::AbstractArray
   ref_trait::Val{R}
+  cell_axes::AbstractArray
 end
 
 function GenericCellField(array::AbstractArray,cell_map::AbstractArray)
-  GenericCellField(array,cell_map,Val{true}())
+  GenericCellField(array,cell_map,Val{true}(),Fill((),length(cell_map)))
+end
+
+function GenericCellField(array::AbstractArray,cell_map::AbstractArray,ref_trait::Val)
+  GenericCellField(array,cell_map,ref_trait,Fill((),length(cell_map)))
 end
 
 function get_array(cf::GenericCellField)
@@ -89,6 +103,10 @@ end
 
 function get_cell_map(cf::GenericCellField)
   cf.cell_map
+end
+
+function get_cell_axes(cf::GenericCellField)
+  cf.cell_axes
 end
 
 function RefStyle(::Type{<:GenericCellField{R}}) where {R}
@@ -100,11 +118,11 @@ end
 # Preserving and manipulating the metadata
 
 """
-    similar_object(cf::CellField,array::AbstractArray)
+    similar_object(cf::CellField,array::AbstractArray,cell_axes::AbstractArray)
 """
-function similar_object(cf::CellField,array::AbstractArray)
+function similar_object(cf::CellField,array::AbstractArray,cell_axes::AbstractArray)
   cm = get_cell_map(cf)
-  GenericCellField(array,cm,RefStyle(cf))
+  GenericCellField(array,cm,RefStyle(cf),cell_axes)
 end
 
 """
@@ -121,16 +139,6 @@ function change_ref_style(cf::CellField)
   GenericCellField(ar,cm,new_sty)
 end
 
-"""
-    have_same_metadata(a::CellField,b::CellField)
-
-Check if `a` and `b` have the same metadata, or at least, if 
-`similar_object(a,get_array(b))` is a `CellField` object that makes sense.
-"""
-function have_same_metadata(a::CellField,b::CellField)
-  RefStyle(a) == RefStyle(b)
-end
-
 # Move between ref and phys spaces by using the underling cell_map
 
 to_ref_space(a::CellField) = _to_ref_space(a,RefStyle(a))
@@ -138,7 +146,7 @@ _to_ref_space(a,::Val{true}) = a
 function _to_ref_space(a,::Val{false})
   cell_map = get_cell_map(a)
   array = compose(  get_array(a), cell_map  )
-  no = similar_object(a,array)
[…]
