### 2020-08-14 0c43a67
Added some functions to merge on the skeleton
