+++ src/Geometry/Geometry.jl
@@ -104,6 +104,12 @@ export to_physical_space
 export is_in_physical_space
 export is_in_ref_space
 export cell_field_from_function
+export is_basis
+export is_test
+export is_trial
+export trialize_cell_basis
+export merge_skeleton_dof_ids
+export merge_skeleton_cell_fields
 
 export GridTopology
 export num_cells
