### 2020-08-14 a359cd0
Re implemented restrict for CellField
