### 2020-08-14 9ab494b
Added the concept of metasize for CellField objects
