+++ src/Geometry/CellFields.jl
@@ -44,6 +44,29 @@ is_in_ref_space(::T) where T = is_in_ref_space(T)
 is_in_physical_space(::Type{T}) where T = !is_in_ref_space(T)
 is_in_physical_space(a::T) where T = !is_in_ref_space(T)
 
+"""
+This trait provides information about the field size at a single 
+evaluation point.
+
+For physical fields `MetaSizeStyle(T) == Val( () )`
+For test bases `MetaSizeStyle(T) == Val( (:,) )`
+For trial bases `MetaSizeStyle(T) = Val( (1,:) )`
+For fields representing elemental matrices `MetaSizeStyle(T) = Val( (:,:) )`
+"""
+MetaSizeStyle(::Type{<:CellField}) = @notimplemented
+
+# We use duck typing here for all types marked with the MetaSizeStyle trait
+MetaSizeStyle(::Type) = @notimplemented
+MetaSizeStyle(::T) where T = MetaSizeStyle(T)
+get_metasize(::Type{T}) where T = get_val_parameter(MetaSizeStyle(T))
+get_metasize(::T) where T = get_val_parameter(MetaSizeStyle(T))
+is_test(::Type{T}) where T = get_metasize(T) == (:,)
+is_test(::T) where T = is_test(T)
+is_trial(::Type{T}) where T = get_metasize(T) == (1,:)
+is_trial(::T) where T = is_trial(T)
+is_basis(::Type{T}) where T =  is_test(T) || is_trial(T)
+is_basis(::T) where T = is_basis(T)
+
 # Tester
 
 """
@@ -69,7 +92,11 @@ function test_cell_field(cf::CellField,x::AbstractArray,b::AbstractArray,pred=(=
   @test get_array(_cf) === get_array(cf)
   @test is_in_ref_space(cf) == !is_in_ref_space(_cf)
   @test is_in_physical_space(cf) == !is_in_physical_space(_cf)
+  @test get_metasize(_cf) == get_metasize(cf)
+  @test get_cell_axes(_cf) == get_cell_axes(cf)
   @test isa(get_cell_axes(cf),AbstractArray{<:Tuple})
+  @test isa(get_metasize(cf),Tuple)
+  @test isa(MetaSizeStyle(cf),Val)
 end
 
 # Concrete implementation
@@ -82,11 +109,12 @@ struct GenericCellField{R} <: CellField
       cell_axes::AbstractArray
     end
 """
-struct GenericCellField{R} <: CellField
+struct GenericCellField{R,S} <: CellField
   array::AbstractArray
   cell_map::AbstractArray
   ref_trait::Val{R}
   cell_axes::AbstractArray
+  metasize::Val{S}
 end
 
 function GenericCellField(array::AbstractArray,cell_map::AbstractArray)
@@ -95,6 +123,11 @@ end
 
 function GenericCellField(array::AbstractArray,cell_map::AbstractArray,ref_trait::Val)
   GenericCellField(array,cell_map,ref_trait,Fill((),length(cell_map)))
+end
+
+function GenericCellField(
+  array::AbstractArray,cell_map::AbstractArray,ref_trait::Val,cell_axes::AbstractArray)
+  GenericCellField(array,cell_map,ref_trait,cell_axes,Val(()))
 end
 
 function get_array(cf::GenericCellField)
@@ -113,40 +146,20 @@ function RefStyle(::Type{<:GenericCellField{R}}) where {R}
   Val{R}()
 end
 
+function MetaSizeStyle(::Type{<:GenericCellField{R,S}}) where {R,S}
+  Val{S}()
+end
+
 # The rest of the file is some default API using the CellField interface
-
[…]
