### 2020-08-15 86edfe7
doc typos
