### 2020-08-15 96a1b14
doc typos
