### 2020-08-16 40776a0
Starting to equip GenericCellField with memoization
