### 2020-08-17 407a0a9
Fixing some tests
