+++ src/Gridap.jl
@@ -50,8 +50,8 @@ include("ReferenceFEs/ReferenceFEs.jl")
 
 include("Geometry/Geometry.jl")
 
-#include("FESpaces/FESpaces.jl")
-#
+include("FESpaces/FESpaces.jl")
+
 #include("MultiField/MultiField.jl")
 #
 #include("Visualization/Visualization.jl")
