+++ test/runtests.jl
@@ -22,8 +22,10 @@ using Test
 
 @time @testset "ReferenceFEs" begin include("ReferenceFEsTests/runtests.jl") end
 
-@time @testset "Geometry" begin include("GeometryTests/runtests.jl") end
+@time @testset "CellData" begin include("CellDataTests/runtests.jl") end
 
+#@time @testset "Geometry" begin include("GeometryTests/runtests.jl") end
+#
 #@time @testset "FESpaces" begin include("FESpacesTests/runtests.jl") end
 #
 #@time @testset "MultiField" begin include("MultiFieldTests/runtests.jl") end
