+++ test/CellDataTests/runtests.jl
@@ -1,0 +1,7 @@
+module CellDataTests
+
+using Test
+
+@testset "CellFields" begin include("CellFieldsTests.jl") end
+
+end # module
