+++ test/CellDataTests/runtests.jl
@@ -4,4 +4,6 @@ using Test
 
 @testset "CellFields" begin include("CellFieldsTests.jl") end
 
+@testset "CellQuadratures" begin include("CellQuadraturesTests.jl") end
+
 end # module
