### 2020-08-17 48ed74b
Added CellData.QPointCellField
