+++ src/Geometry/CellFields.jl
@@ -1,477 +1,0 @@
-
-"""
-    abstract type CellField <: GridapType end
-
-This is a plain array of fields plus some metadata.
-It is interpreted as a `Field` object on each cell of a computational
-mesh.
-"""
-abstract type CellField <: GridapType end
-
-"""
-    get_array(cf::CellField)
-"""
-function get_array(cf::CellField)
-  @abstractmethod
-end
-
-"""
-    get_cell_map(cf::CellField)
-"""
-function get_cell_map(cf::CellField)
-  @abstractmethod
-end
-
-"""
-    get_cell_axes(cf::CellField)
-"""
-function get_cell_axes(cf::CellField)
-  @abstractmethod
-end
-
-"""
-This trait returns `Val{true}()` when the `CellField` is defined in a
-reference finite element space, and `Val{false}()` when it is defined in the
-physical space
-"""
-RefStyle(::Type{<:CellField}) = @notimplemented
-
-# We use duck typing here for all types marked with the RefStyle
-RefStyle(::Type) = @notimplemented
-RefStyle(a) = RefStyle(typeof(a))
-is_in_ref_space(::Type{T}) where T = get_val_parameter(RefStyle(T))
-is_in_ref_space(::T) where T = is_in_ref_space(T)
-is_in_physical_space(::Type{T}) where T = !is_in_ref_space(T)
-is_in_physical_space(a::T) where T = !is_in_ref_space(T)
-
-"""
-This trait provides information about the field size at a single 
-evaluation point.
-
-For physical fields `MetaSizeStyle(T) == Val( () )`
-For test bases `MetaSizeStyle(T) == Val( (:,) )`
-For trial bases `MetaSizeStyle(T) = Val( (1,:) )`
-For fields representing elemental matrices `MetaSizeStyle(T) = Val( (:,:) )`
-"""
-MetaSizeStyle(::Type{<:CellField}) = @notimplemented
-
-# We use duck typing here for all types marked with the MetaSizeStyle trait
-MetaSizeStyle(::Type) = @notimplemented
-MetaSizeStyle(::T) where T = MetaSizeStyle(T)
-get_metasize(::Type{T}) where T = get_val_parameter(MetaSizeStyle(T))
-get_metasize(::T) where T = get_val_parameter(MetaSizeStyle(T))
-is_test(::Type{T}) where T = get_metasize(T) == (:,)
-is_test(::T) where T = is_test(T)
-is_trial(::Type{T}) where T = get_metasize(T) == (1,:)
-is_trial(::T) where T = is_trial(T)
-is_basis(::Type{T}) where T =  is_test(T) || is_trial(T)
-is_basis(::T) where T = is_basis(T)
-
-# Tester
-
-"""
-    test_cell_field(
-      cf::CellField,
-      x::AbstractArray,
-      b::AbstractArray,
-      pred=(==);
-      grad=nothing)
-"""
-function test_cell_field(cf::CellField,x::AbstractArray,b::AbstractArray,pred=(==);grad=nothing)
-  cell_map = get_cell_map(cf)
-  @test isa(cell_map,AbstractArray)
-  a = evaluate(cf,x)
-  test_array(a,b,pred)
-  if grad != nothing
-    g = evaluate(gradient(cf),x)
-    test_array(g,grad,pred)
-  end
-  rs = RefStyle(cf)
-  @test isa(get_val_parameter(rs),Bool)
-  _cf = change_ref_style(cf)
-  @test get_array(_cf) === get_array(cf)
-  @test is_in_ref_space(cf) == !is_in_ref_space(_cf)
[…]
