### 2020-08-17 784d118
Refactred CellData + Geometry
