+++ test/runtests.jl
@@ -24,8 +24,8 @@ using Test
 
 @time @testset "CellData" begin include("CellDataTests/runtests.jl") end
 
-#@time @testset "Geometry" begin include("GeometryTests/runtests.jl") end
-#
+@time @testset "Geometry" begin include("GeometryTests/runtests.jl") end
+
 #@time @testset "FESpaces" begin include("FESpacesTests/runtests.jl") end
 #
 #@time @testset "MultiField" begin include("MultiFieldTests/runtests.jl") end
