+++ test/GeometryTests/AppendedTriangulationsTests.jl
@@ -7,6 +7,7 @@ using Gridap.Geometry
 using Gridap.Fields
 using Gridap.Integration
 using LinearAlgebra: ⋅
+using Gridap.CellData
 
 domain = (0,1,0,1)
 partition = (10,10)
