+++ test/FESpacesTests/runtests.jl
@@ -70,4 +70,6 @@ using Test
 
 @testset "FETermsWithAutodiff" begin include("FETermsWithAutodiffTests.jl") end
 
+@testset "AppendedTriangulations" begin include("AppendedTriangulationsTests.jl") end
+
 end # module
