### 2020-08-17 a6415ca
Added attach_dirichlet and attach_constraints_rows  and _cols
