### 2020-08-17 ceaece9
Resolving constraints at skeleton
