+++ test/CellDataTests/AttachConstraintsTests.jl
@@ -1,7 +1,16 @@
 module AttachConstraintsTests
 
+using Test
+using FillArrays
+using Gridap.Helpers
+using Gridap.Fields
+using Gridap.ReferenceFEs
 using Gridap.Arrays
+using Gridap.Integration
 using Gridap.CellData
+using Gridap.TensorValues
+
+using Gridap.Fields: MockField, MockBasis, OtherMockBasis
 
 ncells = 10
 ndofs = 3
@@ -34,5 +43,86 @@ test_array(a,r)
 #cache = array_cache(a)
 #using BenchmarkTools
 #@btime getindex!($cache,$a,2)
+
+ndofs = 4
+
+z = 2.0
+v = VectorValue(3.0,1.5)
+w = VectorValue(3.4,3.5)
+a = MockBasis{2}(v,ndofs)
+b = MockBasis{2}(w,ndofs)
+c = fill(1.0,ndofs)
+f = OtherMockBasis{2}(ndofs)
+g = MockField{2}(v)
+
+l = 10
+zl = [ z for  i in 1:l]
+cl = fill(c,l)
+fl = Fill(f,l)
+ϕl = lincomb(fl,cl)
+gl = fill(g,l)
+al = Fill(a,l)
+bl = fill(b,l)
+
+gf = GenericCellField(gl,ϕl,Val(true))
+af = GenericCellField(al,ϕl,Val(true),Fill((Base.OneTo(ndofs),),l),Val((:,)))
+bf = GenericCellField(bl,ϕl,Val(true),Fill((Base.OneTo(ndofs),),l),Val((:,)))
+zf = convert_to_cell_field(zl,ϕl)
+df = af*zf
+dft = trialize_cell_basis(df)
+
+degree = 3
+quad = CellQuadrature(degree,[QUAD,],ones(Int,l))
+
+ndofs_c = 5
+cellconstr = [ rand(ndofs_c,ndofs) for i in 1:l]
+cellvals = [ rand(ndofs) for i in 1:l]
+cellvec = integrate(bf⋅v,ϕl,quad)
+cellmat = integrate(∇(bf)⊙∇(dft),ϕl,quad)
+cellmatvec = pair_arrays(cellmat,cellvec)
+cellmatvec = attach_dirichlet(cellmatvec,cellvals)
+cellmatvec = attach_constraints_rows(cellmatvec,cellconstr)
+cellmatvec = attach_constraints_cols(cellmatvec,cellconstr)
+
+@test size(cellmatvec[1][1]) == (ndofs_c,ndofs_c)
+@test size(cellmatvec[1][2]) == (ndofs_c,)
+
+# Test at skeleton
+aS = merge_cell_fields_at_skeleton(af,af)
+dSt = merge_cell_fields_at_skeleton(dft,dft)
+
+axesL = Fill((Base.OneTo(ndofs),),l)
+axesR = axesL
+cellvalsL = [ rand(ndofs) for i in 1:l]
+cellvalsR = cellvalsL
+cellvals = merge_cell_dofs_at_skeleton(cellvalsL,cellvalsR,axesL,axesR)
+cellconstrL = [ rand(ndofs_c,ndofs) for i in 1:l]
+cellconstrR = cellconstrL
+axesL_rows = Fill((Base.OneTo(ndofs_c),),l)
+axesR_rows = axesL_rows
+axesL_cols = axesL
+axesR_cols = axesR
+cellconstr = merge_cell_constraints_at_skeleton(cellconstrL,cellconstrR,axesL_rows,axesR_rows,axesL_cols,axesR_cols)
+
+cellvec = integrate(jump(aS⋅v),ϕl,quad)
[…]
