### 2020-08-17 9ecbc07
Adding BlockArrays to compat section
