### 2020-08-17 9c42ea6
Fixing some test for Julia 1.3
