### 2020-08-18 c4c2813
fix some docstrings
