### 2020-08-18 7101eff
Moved @law and update_state_variables! to CellData module
