### 2020-08-18 c2c2003
Starting the refactoring in Gridap.FESpaces [skip ci]
