+++ src/FESpaces/FESpacesInterfaces.jl
@@ -1,21 +1,4 @@
 
-# The object returned by get_cell_basis has to implement the following trait
-
-FECellBasisStyle(::Type{T}) where T = Val{false}()
-
-FECellBasisStyle(cell_basis) = FECellBasisStyle(typeof(cell_basis))
-
-"""
-"""
-function is_a_fe_cell_basis(cell_basis)
-  v = FECellBasisStyle(cell_basis)
-  get_val_parameter(v)
-end
-
-function is_a_fe_cell_basis(::Type)
-  @unreachable "is_a_fe_cell_basis cannot be called on types"
-end
-
 """
 """
 abstract type FESpace <: GridapType end
@@ -55,97 +38,11 @@ end
 
 """
 """
-function get_cell_basis(f::FESpace)
-  @abstractmethod
-end
-
-"""
-"""
 function constraint_style(::Type{<:FESpace})
   @abstractmethod
 end
 
 constraint_style(f::T) where T<:FESpace = constraint_style(T)
-
-function get_cell_isconstrained(f::FESpace)
-  @abstractmethod
-end
-
-function get_cell_constraints(f::FESpace)
-  @abstractmethod
-end
-
-"""
-"""
-function get_constraint_kernel_matrix_cols(f::FESpace)
-  _default_constraint_kernel(f,constraint_style(f))
-end
-
-"""
-"""
-function get_constraint_kernel_matrix_rows(f::FESpace)
-  _default_constraint_kernel(f,constraint_style(f))
-end
-
-"""
-"""
-function get_constraint_kernel_vector(f::FESpace)
-  _default_constraint_kernel(f,constraint_style(f))
-end
-
-function _default_constraint_kernel(f,::Val{false})
-  function default_kernel(m,isconstr,constr)
-    m
-  end
-end
-
-function _default_constraint_kernel(f,::Val{true})
-  @abstractmethod
-end
-
-"""
-"""
-function test_fe_space(f::FESpace)
-  free_values = zero_free_values(f)
-  fe_function = FEFunction(f,free_values)
-  test_fe_function(fe_function)
-  fe_basis = get_cell_basis(f)
-  @test isa(has_constraints(f),Bool)
-  @test isa(has_constraints(typeof(f)),Bool)
-end
-
-function test_fe_space(f::FESpace,matvecdata,matdata,vecdata)
-  test_fe_space(f)
-
-  cellmat, cellidsrows, cellidscols = matdata
-  cm = apply_constraints_matrix_cols(f,cellmat,cellidscols)
-  if ! has_constraints(f)
-    @test cm === cellmat
-  end
-  cm = apply_constraints_matrix_rows(f,cellmat,cellidsrows)
-  if ! has_constraints(f)
-    @test cm === cellmat
-  end
-
-  cellvec, cellidsrows = vecdata
-  cv = apply_constraints_vector(f,cellvec,cellidsrows)
-  if ! has_constraints(f)
-    @test cv === cellvec
-  end
-
-  cellmatvec, cellidsrows, cellidscols = matvecdata
-  cmv = apply_constraints_matrix_and_vector_cols(f,cellmatvec,cellidscols)
-  if ! has_constraints(f)
-    @test cmv === cellmatvec
-  end
-  cmv = apply_constraints_matrix_and_vector_rows(f,cellmatvec,cellidsrows)
-  if ! has_constraints(f)
-    @test cmv === cellmatvec
-  end
-
-end
-
-# API
 
 """
 """
@@ -158,140 +55,166 @@ has_constraints(f::T) where T<:FESpace = has_constraints(T)
 
 """
 """
[…]
