+++ src/Gridap.jl
@@ -53,8 +53,8 @@ include("CellData/CellData.jl")
 
 include("Geometry/Geometry.jl")
 
-#include("FESpaces/FESpaces.jl")
-#
+include("FESpaces/FESpaces.jl")
+
 #include("MultiField/MultiField.jl")
 #
 #include("Visualization/Visualization.jl")

[…]
