### 2020-08-19 c62a732
Some test restructuring
