+++ test/runtests.jl
@@ -26,8 +26,8 @@ using Test
 
 @time @testset "Geometry" begin include("GeometryTests/runtests.jl") end
 
-#@time @testset "FESpaces" begin include("FESpacesTests/runtests.jl") end
-#
+@time @testset "FESpaces" begin include("FESpacesTests/runtests.jl") end
+
 #@time @testset "MultiField" begin include("MultiFieldTests/runtests.jl") end
 #
 #@time @testset "Visualization" begin include("VisualizationTests/runtests.jl") end
