### 2020-08-19 66bd050
Refactored Assember (broken tests at the skeleton)
