+++ src/FESpaces/SparseMatrixAssemblers.jl
@@ -1,5 +1,5 @@
 
-struct SingleFieldSparseMatrixAssembler{M,V} <: SparseMatrixAssembler
+struct GenericSparseMatrixAssembler{M,V} <: SparseMatrixAssembler
   matrix_type::Type{M}
   vector_type::Type{V}
   trial::SingleFieldFESpace
@@ -9,17 +9,17 @@ end
 
 function SparseMatrixAssembler(
   mat::Type,vec::Type,trial::SingleFieldFESpace,test::SingleFieldFESpace,strategy::AssemblyStrategy)
-  SingleFieldSparseMatrixAssembler(mat,vec,trial,test,strategy)
+  GenericSparseMatrixAssembler(mat,vec,trial,test,strategy)
 end
 
 function SparseMatrixAssembler(mat::Type,vec::Type,trial::SingleFieldFESpace,test::SingleFieldFESpace)
   strategy = DefaultAssemblyStrategy()
-  SingleFieldSparseMatrixAssembler(mat,vec,trial,test,strategy)
+  GenericSparseMatrixAssembler(mat,vec,trial,test,strategy)
 end
 
 function SparseMatrixAssembler(mat::Type,trial::SingleFieldFESpace,test::SingleFieldFESpace)
   strategy = DefaultAssemblyStrategy()
-  SingleFieldSparseMatrixAssembler(mat,Vector{Float64},trial,test,strategy)
+  GenericSparseMatrixAssembler(mat,Vector{Float64},trial,test,strategy)
 end
 
 """
@@ -28,24 +28,23 @@ function SparseMatrixAssembler(trial::SingleFieldFESpace,test::SingleFieldFESpac
   matrix_type = SparseMatrixCSC{Float64,Int}
   vector_type = Vector{Float64}
   strategy = DefaultAssemblyStrategy()
-  SingleFieldSparseMatrixAssembler(matrix_type,vector_type,trial,test,strategy)
+  GenericSparseMatrixAssembler(matrix_type,vector_type,trial,test,strategy)
 end
 
-get_test(a::SingleFieldSparseMatrixAssembler) = a.test
+get_test(a::GenericSparseMatrixAssembler) = a.test
 
-get_trial(a::SingleFieldSparseMatrixAssembler) = a.trial
+get_trial(a::GenericSparseMatrixAssembler) = a.trial
 
-get_matrix_type(a::SingleFieldSparseMatrixAssembler) = a.matrix_type
+get_matrix_type(a::GenericSparseMatrixAssembler) = a.matrix_type
 
-get_vector_type(a::SingleFieldSparseMatrixAssembler) = a.vector_type
+get_vector_type(a::GenericSparseMatrixAssembler) = a.vector_type
 
-get_assembly_strategy(a::SingleFieldSparseMatrixAssembler) = a.strategy
+get_assembly_strategy(a::GenericSparseMatrixAssembler) = a.strategy
 
-function assemble_vector_add!(b,a::SingleFieldSparseMatrixAssembler,vecdata)
-  celldofs = get_cell_dofs(a.test)
+function assemble_vector_add!(b,a::GenericSparseMatrixAssembler,vecdata)
   for (cellvec, cellids) in zip(vecdata...)
-    rows = reindex(celldofs,cellids)
-    vals = apply_constraints_vector(a.test,cellvec,cellids)
+    rows = get_cell_dofs(a.test,cellids)
+    vals = attach_constraints_rows(a.test,cellvec,cellids)
     rows_cache = array_cache(rows)
     vals_cache = array_cache(vals)
     _assemble_vector!(b,vals_cache,rows_cache,vals,rows,a.strategy)
[…]
