+++ test/FESpacesTests/TrialFESpacesTests.jl
@@ -62,6 +62,14 @@ quad = CellQuadrature(trian,order)
 el2 = sqrt(sum(integrate(inner(e,e),trian,quad)))
 @test el2 < 1.0e-10
 
+@test_broken begin
+  uh = zero(U)
+  cellidsL = [4,2,1,3]
+  cellidsR = [2,4,3,1]
+  cellidsS = SkeletonPair(cellidsL,cellidsR)
+  cell_vals = get_cell_values(uh,cellidsS)
+end
+
 #trian = get_triangulation(model)
 #
 #using Gridap.Visualization
