### 2020-08-20 6331ad4
added filtered arrays
