+++ src/Arrays/Arrays.jl
@@ -41,6 +41,7 @@ export setsize!
 export CompressedArray
 export LocalToGlobalArray
 export LocalToGlobalPosNegArray
+export FilteredCellArray
 
 export kernel_cache
 export kernel_caches
@@ -117,6 +118,8 @@ include("Tables.jl")
 include("LocalToGlobalArrays.jl")
 
 include("LocalToGlobalPosNegArrays.jl")
+
+include("FilteredArrays.jl")
 
 include("Reindex.jl")
 
