### 2020-08-20 bdb3aed
Update NEWS.md
