### 2020-08-20 24988e4
Update NEWS.md
