### 2020-08-20 3f715f6
Assembler working on skeleton
