### 2020-08-20 4e36193
ttaching dirichlet bcs without a vector
