+++ src/CellData/AttachDirichlet.jl
@@ -23,3 +23,26 @@ end
     matvec
   end
 end
+
+function Arrays.kernel_cache(k::AttachDirichletKernel,mat::AbstractMatrix,vals,mask)
+  cm = kernel_cache(MulKernel(),mat,vals)
+  cv = CachedArray(mat*vals)
+  fill!(cv.array,zero(eltype(cv)))
+  (cm,cv)
+end
+
+@inline function Arrays.apply_kernel!(cache,k::AttachDirichletKernel,mat::AbstractMatrix,vals,mask)
+  cm, cv = cache
+  if mask
+    vec_with_bcs = apply_kernel!(cm,MulKernel(),mat,vals)
+    scale_entries!(vec_with_bcs,-1)
+    (mat, vec_with_bcs)
+  else
+    if size(mat,1) != size(cv,1)
+      m = axes(mat,1)
+      setaxes!(cv,(m,))
+      fill!(cv.array,zero(eltype(cv)))
+    end
+    (mat, cv.array)
+  end
+end
