### 2020-08-20 24c76e6
Refactored FETerms
