+++ test/FESpacesTests/FETermsTests.jl
@@ -9,6 +9,7 @@ using Gridap.Geometry
 using Gridap.Integration
 using Gridap.Fields
 using Gridap.FESpaces
+using Gridap.CellData
 using LinearAlgebra
 
 domain =(0,1,0,1)
