+++ src/FESpaces/FETerms.jl
@@ -46,11 +46,6 @@ function _setup_jac_and_res(celljac,cellres)
   (celljacres, nothing, nothing)
 end
 
-function _setup_jac_and_res(celljac::SkeletonCellMatrix,cellres::SkeletonCellVector)
-  #TODO for the moment do not pair on the skeleton
-  (nothing, celljac, cellres)
-end
-
 function _setup_jac_and_res(celljac::Nothing,cellres)
   (nothing, nothing, cellres)
 end
@@ -79,20 +74,19 @@ end
 
 """
 Returns an object (e.g. a CellVector) representing the contribution to the
-system rhs of the given term. Returns nothing if the term has not
+system rhs of the given term (with Dirichlet bcs included). Returns nothing if the term has not
 contribution (typically for linear terms)
 """
 function get_cell_vector(t::AffineFETerm,uhd,v)
   @abstractmethod
 end
 
+"""
+Returns an object (e.g. a CellVector) representing the contribution to the
+system rhs of the given term (without Dirichlet bcs). Returns nothing if the term has not
+contribution (typically for linear terms)
+"""
 function get_cell_vector(t::AffineFETerm,v)
-  @abstractmethod
-end
-
-"""
-"""
-function get_cell_values(t::FETerm,uhd)
   @abstractmethod
 end
 
@@ -110,34 +104,19 @@ function get_cell_matrix_and_vector(t::AffineFETerm,uhd,u,v)
   @assert is_a_fe_cell_basis(v)
   @assert is_a_fe_cell_basis(u)
   cellmat = get_cell_matrix(t,u,v)
-  cellvec = _get_cell_vector_tmp_hack(cellmat,t,v,uhd) #TODO
-  cellvals = get_cell_values(t,uhd)
-  _setup_cell_matrix_and_vector(cellmat,cellvec,cellvals)
-end
-
-function _get_cell_vector_tmp_hack(cellmat,t,v,uhd) #TODO
   cellvec = get_cell_vector(t,v)
-  cellvec
-end
-
-function _get_cell_vector_tmp_hack(cellmat::SkeletonCellMatrix,t,v,uhd) #TODO
-  cellvec = get_cell_vector(t,uhd,v)
-  cellvec
+  cellvals = get_cell_values(uhd,get_cell_id(t))
+  _setup_cell_matrix_and_vector(cellmat,cellvec,cellvals)
 end
 
 function  _setup_cell_matrix_and_vector(cellmat,cellvec,cellvals)
   cellmatvec = pair_arrays(cellmat,cellvec)
-  cellmatvec_with_diri = attach_dirichlet_bcs(cellmatvec,cellvals)
+  cellmatvec_with_diri = attach_dirichlet(cellmatvec,cellvals)
   (cellmatvec_with_diri, nothing, nothing)
 end
 
-function  _setup_cell_matrix_and_vector(cellmat::SkeletonCellMatrix,cellvec::SkeletonCellVector,cellvals)
-  # TODO for the moment do not pair quantities on the skeleton and assume that cellvec has dirichlet bcs
-  (nothing, cellmat, cellvec)
-end
-
 function _setup_cell_matrix_and_vector(cellmat,cellvec::Nothing,cellvals)
-  cellmatvec_with_diri = attach_dirichlet_bcs(cellmat,cellvals)
+  cellmatvec_with_diri = attach_dirichlet(cellmat,cellvals)
   (cellmatvec_with_diri, nothing, nothing)
 end
 
[…]
