### 2020-08-20 8be8351
Re-activated FEOperators and Co and removed CellKernels
