+++ test/FESpacesTests/FETermsTests.jl
@@ -128,27 +128,10 @@ jac = ∇(ϕ)
 jac_q = evaluate(jac,q)
 x_q = evaluate(ϕ,q)
 
-function poisson_matvec_kernel!(mat,vec,∇v,∇u,v,j,w,x)
-  Q = length(w)
-  M,N = size(mat)
-  for q in 1:Q
-    dV = det(j[q])*w[q]
-    f_q = f(x[q])
-    for n in 1:N
-      for m in 1:M
-        mat[m,n] += ∇v[q,m]⊙∇u[q,n]*dV
-      end
-    end
-    for m in 1:M
-      vec[m] += v[q,m]*f_q*dV
-    end
-  end
-end
-
 function matvecfun(u,v)
-  v_q = evaluate(v,q)
-  ∇v_q = evaluate(∇(v),q)
-  apply_cellmatvec(poisson_matvec_kernel!, ∇v_q, ∇v_q, v_q, jac_q, w_q, x_q)
+  cellmat = integrate(∇(u)⋅∇(v),trian,quad)
+  cellvec = integrate(v*f,trian,quad)
+  pair_arrays(cellmat,cellvec)
 end
 
 t_matvec_Ω = AffineFETermFromCellMatVec(matvecfun,trian)
@@ -166,30 +149,10 @@ assemble_matrix_and_vector!(A,b,assem,data)
 x = A \ b
 @test x ≈ get_free_values(uh)
 
-function poisson_jacres_kernel!(jac,res,∇v,∇du,v,∇uh,j,w,x)
-  Q = length(w)
-  M,N = size(jac)
-  for q in 1:Q
-    dV = det(j[q])*w[q]
-    f_q = f(x[q])
-    for m in 1:M
-      for n in 1:N
-        jac[m,n] += ∇v[q,m]⊙∇du[q,n]*dV
-      end
-      res[m] += ∇v[q,m]⊙∇uh[q]*dV
-    end
-    for m in 1:M
-      res[m] -= v[q,m]*f_q*dV
-    end
-  end
-end
-
 function jacresfun(uh,du,v)
-  v_q = evaluate(v,q)
-  ∇v_q = evaluate(∇(v),q)
-  ∇du_q = ∇v_q
-  ∇uh_q = evaluate(∇(uh),q)
-  apply_cellmatvec(poisson_jacres_kernel!, ∇v_q, ∇du_q, v_q, ∇uh_q, jac_q, w_q, x_q)
+  cellmat = integrate(∇(du)⋅∇(v),trian,quad)
+  cellvec = integrate(∇(uh)⋅∇(v)-v*f,trian,quad)
+  pair_arrays(cellmat,cellvec)
 end
 
 t_jacres_Ω = FETermFromCellJacRes(jacresfun,trian)
