### 2020-08-20 d477051
Reactivating other part of FESpaces module
