+++ test/FESpacesTests/FEAutodiffTests.jl
@@ -6,6 +6,8 @@ using Gridap.Arrays
 using Gridap.Fields
 using Gridap.Geometry
 using Gridap.TensorValues 
+using Gridap.CellData
+
 domain = (0,1,0,1)
 partition = (2,2)
 model = CartesianDiscreteModel(domain,partition)
