### 2020-08-20 e654324
Finishing clean-up in FESpaces
