+++ test/FESpacesTests/FESpaceFactoriesTests.jl
@@ -7,6 +7,7 @@ using Gridap.Arrays
 using Gridap.TensorValues
 using Gridap.ReferenceFEs
 using Gridap.FESpaces
+using Gridap.CellData
 
 domain = (0,1,0,1)
 partition = (3,3)
