+++ test/FESpacesTests/CellBasesTests.jl
@@ -1,224 +1,0 @@
-module CellBasesTests
-
-using Test
-using Gridap.Arrays
-using Gridap.TensorValues
-using Gridap.ReferenceFEs
-using Gridap.Geometry
-using Gridap.Integration
-using Gridap.Fields
-using Gridap.FESpaces
-using Gridap.FESpaces: SkeletonCellBasis
-using Gridap.FESpaces: ReducedSkeletonCellBasis
-using Gridap.FESpaces: SkeletonCellVector
-using Gridap.FESpaces: SkeletonCellMatrixField
-using Gridap.FESpaces: SkeletonCellMatrix
-
-using Gridap.Fields: MockField, MockBasis
-
-domain =(0,1,0,1)
-partition = (3,3)
-model = CartesianDiscreteModel(domain,partition)
-
-order = 2
-grid_topology = get_grid_topology(model)
-polytopes = get_polytopes(grid_topology)
-reffes = [LagrangianRefFE(Float64,p,order) for p in polytopes]
-
-dirichlet_tags = ["tag_1","tag_6"]
-V0 = GradConformingFESpace(reffes,model,dirichlet_tags)
-
-trian = get_triangulation(model)
-
-degree = 2
-quad = CellQuadrature(trian,degree)
-q = get_coordinates(quad)
-
-cell_basis = get_cell_basis(V0)
-
-cbx = collect(evaluate(cell_basis,q))
-
-f(x) = sin(4*pi*(x[1]-x[2]^2))+1
-
-uh = interpolate_everywhere(f,V0)
-uhx = collect(evaluate(uh,q))
-
-v = GenericCellBasis(Val{false}(),get_array(cell_basis),get_cell_map(cell_basis),Val{true}())
-u = GenericCellBasis(Val{true}(),get_array(cell_basis),get_cell_map(cell_basis),Val{true}())
-
-dv = v
-du = u
-
-a = rand(length(v))
-
-r = operate(+,uh,2*uh,a)
-@test isa(r,CellField)
-rq = collect(evaluate(r,q))
-
-r = operate(+,du,uh,a)
-@test isa(r,CellBasis)
-@test is_trial(r)
-rq = collect(evaluate(r,q))
-
-r = operate(+,uh,du,a)
-@test isa(r,CellBasis)
-@test is_trial(r)
-rq = collect(evaluate(r,q))
-
-r = operate(+,du,du,uh,a)
-@test isa(r,CellBasis)
-@test is_trial(r)
-rq = collect(evaluate(r,q))
-
-r = u*2
-test_cell_basis(r,q,2*cbx)
-@test is_trial(r)
-@test ! is_test(r)
-
-r = 2*u
-test_cell_basis(r,q,2*cbx)
-@test is_trial(r)
-
-r = u + uh
-rr = [ ai .+ bi for (ai,bi) in  zip(cbx,uhx)]
-test_cell_basis(r,q,rr)
-@test is_trial(r)
-
-r = v*2
-test_cell_basis(r,q,2*cbx)
-@test is_test(r)
-@test ! is_trial(r)
-
-r = 2*v
-test_cell_basis(r,q,2*cbx)
-@test is_test(r)
-
-r = v + uh
-rr = [ ai .+ bi for (ai,bi) in  zip(cbx,uhx)]
-test_cell_basis(r,q,rr)
-@test is_test(r)
-
-w = inner(u,v)
-@test isa(w,CellMatrixField)
-
-w = u * v
-@test isa(w,CellMatrixField)
-wx = collect(evaluate(w,q))
-
-s = 2*w
-test_cell_matrix_field(s,q,2*wx,≈)
-
-s = w*2
[…]
