+++ src/FESpaces/Assemblers.jl
@@ -166,141 +166,3 @@ function test_assembler(a::Assembler,matdata,vecdata,data)
   @test num_free_dofs(test_fesp) == length(b)
 end
 
-# This is an extended interface that only makes sense for assemblers that build (sequential) sparse matrices
-# (e.g. not for matrix free assemblers or for distributed assemblers)
-
-"""
-"""
-abstract type SparseMatrixAssembler <: Assembler end
-
-"""
-"""
-function get_matrix_type(a::SparseMatrixAssembler)
-  @abstractmethod
-end
-
-"""
-"""
-function get_vector_type(a::SparseMatrixAssembler)
-  @abstractmethod
-end
-
-function allocate_vector(a::SparseMatrixAssembler,vecdata)
-  n = num_free_dofs(get_test(a))
-  allocate_vector(get_vector_type(a),n)
-end
-
-function assemble_vector!(b,a::SparseMatrixAssembler,vecdata)
-  fill_entries!(b,zero(eltype(b)))
-  assemble_vector_add!(b,a,vecdata)
-end
-
-"""
-"""
-function count_matrix_nnz_coo(a::SparseMatrixAssembler,matdata)
-  @abstractmethod
-end
-
-"""
-"""
-function count_matrix_and_vector_nnz_coo(a::SparseMatrixAssembler,data)
-  @abstractmethod
-end
-
-"""
-"""
-function fill_matrix_coo_symbolic!(I,J,a::SparseMatrixAssembler,matdata,n=0)
-  @abstractmethod
-end
-
-function fill_matrix_and_vector_coo_symbolic!(I,J,a::SparseMatrixAssembler,data,n=0)
-  @abstractmethod
-end
-
-function allocate_matrix(a::SparseMatrixAssembler,matdata)
-  n = count_matrix_nnz_coo(a,matdata)
-  I,J,V = allocate_coo_vectors(get_matrix_type(a),n)
-  fill_matrix_coo_symbolic!(I,J,a,matdata)
-  m = num_free_dofs(get_test(a))
-  n = num_free_dofs(get_trial(a))
-  finalize_coo!(get_matrix_type(a),I,J,V,m,n)
-  sparse_from_coo(get_matrix_type(a),I,J,V,m,n)
-end
-
-function assemble_matrix!(mat,a::SparseMatrixAssembler,matdata)
-  z = zero(eltype(mat))
-  fill_entries!(mat,z)
-  assemble_matrix_add!(mat,a,matdata)
-end
-
-"""
-"""
-function fill_matrix_coo_numeric!(I,J,V,a::SparseMatrixAssembler,matdata,n=0)
-  @abstractmethod
-end
-
-function assemble_matrix(a::SparseMatrixAssembler,matdata)
-
-  n = count_matrix_nnz_coo(a,matdata)
-  I,J,V = allocate_coo_vectors(get_matrix_type(a),n)
-
-  fill_matrix_coo_numeric!(I,J,V,a,matdata)
-
-  m = num_free_dofs(get_test(a))
-  n = num_free_dofs(get_trial(a))
-  finalize_coo!(get_matrix_type(a),I,J,V,m,n)
-  sparse_from_coo(get_matrix_type(a),I,J,V,m,n)
-end
-
-
-function allocate_matrix_and_vector(a::SparseMatrixAssembler,data)
-
-  n = count_matrix_and_vector_nnz_coo(a,data)
-
-  I,J,V = allocate_coo_vectors(get_matrix_type(a),n)
-  fill_matrix_and_vector_coo_symbolic!(I,J,a,data)
-  m = num_free_dofs(get_test(a))
[…]
