+++ test/FieldsTests/VectorsOfBlockBasisCooTests.jl
@@ -29,33 +29,47 @@ xl = Fill(x,l)
 
 ax1 = Fill((Base.OneTo(ndofs1),),l)
 ax2 = Fill((Base.OneTo(ndofs2),),l)
-aBl = insert_array_of_bases_in_block(al,ax1,ax2,2)
+aBl = insert_array_of_bases_in_block(2,al,ax1,ax2)
 aBl_x = evaluate(aBl,xl)
 @test isa(aBl,VectorOfBlockBasisCoo)
 @test isa(aBl_x,VectorOfBlockArrayCoo)
 
-bBl = insert_array_of_bases_in_block(bl,ax1,ax2,1)
+ax1 = Fill((Base.OneTo(ndofs1),),l)
+ax2 = Fill((Base.OneTo(ndofs2),),l)
+aBl = insert_array_of_bases_in_block(3,al,ax1,ax2,ax2)
+aBl_x = evaluate(aBl,xl)
+@test isa(aBl,VectorOfBlockBasisCoo)
+@test isa(aBl_x,VectorOfBlockArrayCoo)
+
+bBl = insert_array_of_bases_in_block(1,bl,ax1,ax2)
 bBl_x = evaluate(bBl,xl)
 @test isa(bBl,VectorOfBlockBasisCoo)
 @test isa(bBl_x,VectorOfBlockArrayCoo)
 
 ax1 = Fill((Base.OneTo(1),Base.OneTo(ndofs1)),l)
 ax2 = Fill((Base.OneTo(1),Base.OneTo(ndofs2)),l)
-atBl = insert_array_of_bases_in_block(atl,ax1,ax2,2)
+atBl = insert_array_of_bases_in_block(2,atl,ax1,ax2)
+atBl_x = evaluate(atBl,xl)
+@test isa(atBl,VectorOfBlockBasisCoo)
+@test isa(atBl_x,VectorOfBlockArrayCoo)
+
+ax1 = Fill((Base.OneTo(1),Base.OneTo(ndofs1)),l)
+ax2 = Fill((Base.OneTo(1),Base.OneTo(ndofs2)),l)
+atBl = insert_array_of_bases_in_block(3,atl,ax1,ax2,ax2)
 atBl_x = evaluate(atBl,xl)
 @test isa(atBl,VectorOfBlockBasisCoo)
 @test isa(atBl_x,VectorOfBlockArrayCoo)
 
 ax1 = aBl.axes
 ax2 = aBl.axes
-aSl = insert_array_of_bases_in_block(aBl,ax1,ax2,1)
+aSl = insert_array_of_bases_in_block(2,aBl,ax1,ax2)
 @test isa(aSl,VectorOfBlockBasisCoo)
 aSl_x = evaluate(aSl,xl)
 @test isa(aSl_x,VectorOfBlockArrayCoo)
 
 ax1 = atBl.axes
 ax2 = atBl.axes
-atSl = insert_array_of_bases_in_block(atBl,ax1,ax2,1)
+atSl = insert_array_of_bases_in_block(1,atBl,ax1,ax2)
 @test isa(atSl,VectorOfBlockBasisCoo)
 atSl_x = evaluate(atSl,xl)
 @test isa(atSl_x,VectorOfBlockArrayCoo)
