### 2020-08-21 dc1f3af
Refactored the code part of MultiField
