+++ src/MultiField/MultiFieldFESpaces.jl
@@ -49,16 +49,12 @@ end
 
 function FESpaces.get_cell_axes(f::MultiFieldFESpace)
   all_axes = map(get_cell_axes,f.spaces)
-  _multifield_cell_axes(all_axes)
+  create_array_of_blocked_axes(all_axes...)
 end
 
 function FESpaces.get_cell_axes_with_constraints(f::MultiFieldFESpace)
   all_axes = map(get_cell_axes_with_constraints,f.spaces)
-  _multifield_cell_axes(all_axes)
-end
-
-function _multifield_cell_axes(all_axes)
-  apply(Fields._cat_axes,all_axes...)
+  create_array_of_blocked_axes(all_axes...)
 end
 
 function FESpaces.get_cell_basis(f::MultiFieldFESpace)
@@ -74,318 +70,202 @@ function _insert_cell_basis_in_block(i::Int,cb,all_axes)
   similar_object(cb,array,array.axes,MetaSizeStyle(cb))
 end
 
+function FESpaces.FEFunction(fe::MultiFieldFESpace, free_values)
+  blocks = SingleFieldFEFunction[]
+  for (field, U) in enumerate(fe.spaces)
+    free_values_i = restrict_to_field(fe,free_values,field)
+    uhi = FEFunction(U,free_values_i)
+    push!(blocks,uhi)
+  end
+  MultiFieldFEFunction(free_values,fe,blocks)
+end
+
+function FESpaces.EvaluationFunction(fe::MultiFieldFESpace, free_values)
+  blocks = SingleFieldFEFunction[]
+  for (field, U) in enumerate(fe.spaces)
+    free_values_i = restrict_to_field(fe,free_values,field)
+    uhi = EvaluationFunction(U,free_values_i)
+    push!(blocks,uhi)
+  end
+  MultiFieldFEFunction(free_values,fe,blocks)
+end
+
+function CellData.CellField(fe::MultiFieldFESpace,cell_values)
+  @notimplemented
+end
+
+function CellData.CellField(fe::MultiFieldFESpace,cell_values::VectorOfBlockArrayCoo)
+  single_fields = CellField[]
+  for field in 1:num_fields(fe)
+    cf = CellField(fe.spaces[field],cell_values[Block(field)])
+    push!(single_fields,cf)
+  end
+  MultiFieldCellField(single_fields)
+end
+
+"""
+    restrict_to_field(f::MultiFieldFESpace,free_values::AbstractVector,field::Integer)
+"""
+function restrict_to_field(f::MultiFieldFESpace,free_values::AbstractVector,field::Integer)
+  _restrict_to_field(f,MultiFieldStyle(f),free_values,field)
+end
+
+function  _restrict_to_field(f,::MultiFieldStyle,free_values,field)
+  @notimplemented
+end
+
+function  _restrict_to_field(f,::ConsecutiveMultiFieldStyle,free_values,field)
+  offsets = compute_field_offsets(f)
+  U = f.spaces
+  pini = offsets[field] + 1
+  pend = offsets[field] + num_free_dofs(U[field])
+  SubVector(free_values,pini,pend)
+end
+
+"""
+    compute_field_offsets(f::MultiFieldFESpace)
+"""
+function compute_field_offsets(f::MultiFieldFESpace)
+  @assert MultiFieldStyle(f) == ConsecutiveMultiFieldStyle()
+  U = f.spaces
+  n = length(U)
+  offsets = zeros(Int,n)
+  for i in 1:(n-1)
+    Ui = U[i]
+    offsets[i+1] = offsets[i] + num_free_dofs(Ui)
+  end
+  offsets
+end
+
[…]
