+++ test/runtests.jl
@@ -28,7 +28,7 @@ using Test
 
 @time @testset "FESpaces" begin include("FESpacesTests/runtests.jl") end
 
-#@time @testset "MultiField" begin include("MultiFieldTests/runtests.jl") end
+@time @testset "MultiField" begin include("MultiFieldTests/runtests.jl") end
 
 #@time @testset "Visualization" begin include("VisualizationTests/runtests.jl") end
 #
