+++ src/FESpaces/FESpaces.jl
@@ -149,8 +149,11 @@ export compute_dirichlet_values_for_tags
 export compute_dirichlet_values_for_tags!
 export test_single_field_fe_space
 export interpolate
+export interpolate!
 export interpolate_everywhere
+export interpolate_everywhere!
 export interpolate_dirichlet
+export interpolate_dirichlet!
 export get_cell_dof_basis
 
 export SingleFieldFEFunction
