+++ src/Gridap.jl
@@ -55,7 +55,7 @@ include("Geometry/Geometry.jl")
 
 include("FESpaces/FESpaces.jl")
 
-#include("MultiField/MultiField.jl")
+include("MultiField/MultiField.jl")
 
 #include("Visualization/Visualization.jl")
 #
