### 2020-08-21 f809d2b
Deleting some files not needed any more in MultiField
