+++ test/MultiFieldTests/runtests.jl
@@ -8,7 +8,7 @@ using Test
 
 @testset "MultiFieldFEFunctions" begin include("MultiFieldFEFunctionsTests.jl") end
 
-#@testset "MultiFieldSparseMatrixAssemblers" begin include("MultiFieldSparseMatrixAssemblersTests.jl") end
+@testset "MultiFieldSparseMatrixAssemblers" begin include("MultiFieldSparseMatrixAssemblersTests.jl") end
 #
 #@testset "MultiFieldFEOperators" begin include("MultiFieldFEOperatorsTests.jl") end
 #
