### 2020-08-21 8050411
Splitting FESpaces tests in two tests sets to avoid problems in travis
