+++ test/runtests.jl
@@ -32,8 +32,8 @@ using Test
 
 @time @testset "MultiField" begin include("MultiFieldTests/runtests.jl") end
 
-#@time @testset "Visualization" begin include("VisualizationTests/runtests.jl") end
-#
+@time @testset "Visualization" begin include("VisualizationTests/runtests.jl") end
+
 #@time @testset "Issues" begin
 #    include("issue_349.jl")
 #    include("issue_368.jl")
