+++ src/Gridap.jl
@@ -57,7 +57,7 @@ include("FESpaces/FESpaces.jl")
 
 include("MultiField/MultiField.jl")
 
-#include("Visualization/Visualization.jl")
+include("Visualization/Visualization.jl")
 #
 #include("Exports.jl")
 
