+++ src/Gridap.jl
@@ -58,7 +58,7 @@ include("FESpaces/FESpaces.jl")
 include("MultiField/MultiField.jl")
 
 include("Visualization/Visualization.jl")
-#
-#include("Exports.jl")
+
+include("Exports.jl")
 
 end # module
