### 2020-08-21 e46640c
Updating IsotropicDamageTests
