### 2020-08-21 ee5bf7e
Fixing docs/make.jl
