### 2020-08-21 ed8d3bf
Fixed un-initialization related bug in mul! for block matrices
