+++ NEWS.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## [0.13.4] - Unreleased
+## [0.13.4] - 2020-08-24
 
 ### Added 
   - New `FilteredCellArray` since PR [#372](https://github.com/gridap/Gridap.jl/pull/372/).
