### 2020-08-23 670dabd
Update NEWS.md
