### 2020-08-24 5fffc8a
Update NEWS.md [skip ci]
