### 2020-08-24 2ee70a6
Removed bench and reducing test length
