+++ test/runtests.jl
@@ -35,12 +35,9 @@ using Test
 @time @testset "Visualization" begin include("VisualizationTests/runtests.jl") end
 
 @time @testset "Issues" begin
-    include("issue_349.jl")
     include("issue_368.jl")
 end
 
 include("GridapTests/runtests.jl")
-
-include("../bench/runbenchs.jl")
 
 end # module
