### 2020-08-25 d84a8d3
Update NEWS.md
