### 2020-08-25 3cef3ee
Allowing a non-blocked matrix on skeleton and multi-field
