+++ test/FESpacesTests/SparseMatrixAssemblersTests.jl
@@ -151,4 +151,10 @@ test_sparse_matrix_assembler(assem,matdata,vecdata,data)
 A = assemble_matrix(assem,matdata)
 @test A == zeros(num_free_dofs(V),num_free_dofs(U))
 
+# Allowing a non-blocked matrix in a context where you would expect a blocked one.
+scellmat = map(i->zeros(size(i)),scellmat)
+matdata = ([scellmat],[scellids],[scellids])
+A = assemble_matrix(assem,matdata)
+@test A == zeros(num_free_dofs(V),num_free_dofs(U))
+
 end # module
