### 2020-08-25 e93b76a
Update NEWS.md
