### 2020-08-25 5d8ba28
Update NEWS.md
