### 2020-08-25 c7e2bf0
Implementing operations with several arguments for multi-field
