### 2020-08-25 12fcdd1
Functions Kernel
