+++ src/NewKernels/FunctionKernels.jl
@@ -1,0 +1,10 @@
+struct FunctionKernel{F} <: NewKernel
+  f::F
+end
+
+evaluate!(cache,k::FunctionKernel,x...) = k.f(x...)
+
+function return_type(k::FunctionKernel,x...)
+Ts = map(typeof,x)
+return_type(k.f,Ts...)
+end
