### 2020-08-25 de07e31
No duck-typing npw
