### 2020-08-25 e6654bf
More structs
