+++ src/NewKernels/Interface.jl
@@ -1,0 +1,185 @@
+
+"""
+Abstract type representing the operations to be used in the [`apply`](@ref) function.
+
+Derived types must implement the following method:
+
+- [`apply_kernel!(cache,k,x...)`](@ref)
+
+and optionally these ones:
+
+- [`kernel_cache(k,x...)`](@ref)
+- [`kernel_return_type(k,x...)`](@ref)
+
+The kernel interface can be tested with the [`test_kernel`](@ref) function.
+
+Note that most of the functionality implemented in terms of this interface
+relies in duck typing. That is, it is not strictly needed to work with types
+that inherit from `Kernel`. This is specially useful in order to accommodate
+existing types into this framework without the need to implement a wrapper type
+that inherits from `Kernel`. For instance, a default implementation is available
+for `Function` objects.  However, we recommend that new types inherit from `Kernel`.
+
+"""
+abstract type NewKernel <: GridapType end
+
+"""
+    kernel_return_type(f,x...)
+
+Returns the type of the result of calling kernel `f` with
+arguments of the types of the objects `x`.
+
+It defaults to `typeof(kernel_testitem(f,x...))`
+"""
+return_type(f::NewKernel,x) = typeof(testitem(f,x...))
+
+"""
+    kernel_cache(f,x...)
+
+Returns the `cache` needed to apply kernel `f` with arguments
+of the same type as the objects in `x`.
+This function returns `nothing` by default.
+"""
+cache(f::NewKernel,x) = nothing
+
+"""
+    apply_kernel!(cache,f,x...)
+
+Applies the kernel `f` at the arguments `x...` using
+the scratch data provided in the given `cache` object. The `cache` object
+is built with the [`kernel_cache`](@ref) function using arguments of the same type as in `x`.
+In general, the returned value `y` can share some part of its state with the `cache` object.
+If the result of two or more invocations of this function need to be accessed simultaneously
+(e.g., in multi-threading), create and use various `cache` objects (e.g., one cache
+per thread).
+"""
+evaluate!(cache,f::NewKernel,x...) = @abstractmethod
+
+
+# Testing the interface
+
+"""
+    test_kernel(f,x::Tuple,y,cmp=(==))
+
+Function used to test if the kernel `f` has been
+implemented correctly. `f` is a kernel object, `x` is a tuple containing the arguments
+of the kernel, and `y` is the expected result. Function `cmp` is used to compare
+the computed result with the expected one. The checks are done with the `@test`
+macro.
+"""
+function test_kernel(f,x::Tuple,y,cmp=(==))
+  z = evaluate(f,x...)
+  @test cmp(z,y)
+  @test typeof(z) == return_type(f,x...)
+  cache = cache(f,x...)
+  z = evaluate!(cache,f,x...)
+  @test cmp(z,y)
+  z = evaluate!(cache,f,x...)
+  @test cmp(z,y)
+  z = testitem!(cache,f,x...)
+  @test cmp(typeof(z),typeof(y))
+end
+
+
+# Functions on kernel objects
+
+"""
+    apply_kernel(f,x...)
+
+apply the kernel `f` at the arguments in `x` by creating a temporary cache
+internally. This functions is equivalent to
+```jl
+cache = kernel_cache(f,x...)
+apply_kernel!(cache,f,x...)
+```
+"""
[…]
