### 2020-08-25 534bc69
Added a new type extension of SingleFieldFESpace, named FESpaceWithDofPotentiallyRemoved. FESpaceWithDofPotentiallyRemoved is though to be a future replacement for FESpaceWithLastDofRemoved, as it is based on the same concept, but it is more general as it lets the user to activate/deactivate whether to remove the dof at hand, and select the particular dof to be removed
