+++ src/FESpaces/FESpaces.jl
@@ -205,6 +205,8 @@ export collect_cell_jacobian_and_residual
 export collect_cell_residual
 
 export FESpaceWithLastDofRemoved
+export FESpaceWithDofPotentiallyRemoved
+
 export ZeroMeanFESpace
 export CLagrangianFESpace
 export ConformingFESpace
@@ -253,6 +255,8 @@ include("FEOperatorsFromTerms.jl")
 include("FESolvers.jl")
 
 include("FESpacesWithLastDofRemoved.jl")
+
+include("FESpacesWithDofPotentiallyRemoved.jl")
 
 include("ZeroMeanFESpaces.jl")
 
