### 2020-08-25 #2520 @Kevin-Mattheus-Moerman
@fverdugo some final steps:

- [ ] I've reviewed your paper and confirm it looks good. However can you remove this last sentence:
`The source code for Gridap has been archived to Zenodo with the linked DOI 10.5281/zenodo.3934468.`, I ask this since the paper will have a link (top left) to the latest and more accurate archived version anyway.

- [ ] This work is about to be processed for acceptance. Can you check the paper yourself once more and also check that the author names, affiliations, and the acknowledgements are accurate?

- [ ] Can archive the latest version of the reviewed work on Zenodo and report back here with the archive DOI? Can you make sue the archive meta-data like the title and authors match that of the paper (you may need to edit information)?

- [ ] The version is currently listed as `v0.12.0`, can you check if this is still accurate or if it needs to be altered?
