### 2020-08-25 a5e00a1
Modified ZeroMeanFESpace such that it no longer depends on FESpaceWithLastDofRemoved, but on FESpaceWithDofPotentiallyRemoved,
