### 2020-08-25 0319cdd
added new kernels
