+++ src/NewKernels/FunctionKernels.jl
@@ -4,7 +4,11 @@ end
 
 evaluate!(cache,k::FunctionKernel,x...) = k.f(x...)
 
+# function return_type(k::FunctionKernel,x...)
+  # Ts = map(typeof,x)
+  # return_type(k.f,Ts...)
+# end
+
 function return_type(k::FunctionKernel,x...)
-Ts = map(typeof,x)
-return_type(k.f,Ts...)
+  typeof(evaluate(k,x...))
 end
