+++ src/NewKernels/NewKernels.jl
@@ -1,5 +1,24 @@
 module NewKernels
 
+using Gridap.Helpers
+using Gridap.Inference
+
+using Test
+
+export NewKernel
+
+export FunctionKernel
+export BroadcastKernel
+export CompositionKernel
+
+export return_type
+export cache
+export evaluate!
+export evaluate
+export test_kernel
+export caches
+export return_types
+
 include("Interface.jl")
 
 include("BroadcastKernels.jl")
