### 2020-08-25 f8e8942
Implementing missing case for apply and ExtendedVector
