### 2020-08-25 f032e83
Revert "Implementing missing case for apply and ExtendedVector"
This reverts commit f8e8942c30ffa8d6756e9b68386c19d16914a3dc.
