### 2020-08-25 54a0bf5
Fulfilling editor's requests.
