### 2020-08-25 #2520 @fverdugo
>  I've reviewed your paper and confirm it looks good. However can you remove this last sentence:
The source code for Gridap has been archived to Zenodo with the linked DOI 10.5281/zenodo.3934468., I ask this since the paper will have a link (top left) to the latest and more accurate archived version anyway.

I have removed the line (see the proof). **Note however, that the link to the archive (top left in the proof) is not working.**

 > This work is about to be processed for acceptance. Can you check the paper yourself once more and also check that the author names, affiliations, and the acknowledgements are accurate?

It is ok from my side (except the broken link, which has to be fixed). @santiagobadia can you take your final look?

> Can you archive the latest version of the reviewed work on Zenodo and report back here with the archive DOI? Can you make sure the archive meta-data like the title and authors match that of the paper (you may need to edit information)?

I have archived the latest stable version of Gridap (v0.13.4)

  [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.3999839.svg)](https://doi.org/10.5281/zenodo.3999839)

The meta-data matches the one of the paper.

 > The version is currently listed as v0.12.0, can you check if this is still accurate or if it needs to be altered? (it should match the Zenodo archive version also)

v0.12.0 is out-of-date. The latest archived one is v0.13.4.
