### 2020-08-25 #2520 @Kevin-Mattheus-Moerman
@fverdugo the link should work in the accepted paper form. You asked @santiagobadia to review the paper as well so I'll wait until I hear back from yourself or @santiagobadia if we can proceed with acceptance.
