### 2020-08-26 4a1a427
Tests passing for NewKernels so far
