+++ src/NewKernels/FunctionKernels.jl
@@ -4,11 +4,12 @@ end
 
 evaluate!(cache,k::FunctionKernel,x...) = k.f(x...)
 
-# function return_type(k::FunctionKernel,x...)
-  # Ts = map(typeof,x)
-  # return_type(k.f,Ts...)
-# end
-
 function return_type(k::FunctionKernel,x...)
-  typeof(evaluate(k,x...))
+  Ts = map(typeof,x)
+  return_type(k.f,Ts...)
 end
+
+# @santiagobadia : Why not just this?
+#function return_type(k::FunctionKernel,x...)
+#  typeof(evaluate(k,x...))
+#end
