+++ src/NewKernels/Tests/NewKernelsTests.jl
@@ -1,38 +1,23 @@
 module NewKernelsTests
 
 using Test
-using Gridap.Arrays
+using Gridap.Arrays: CachedArray
 using Gridap.NewKernels
 using Gridap.TensorValues
-using LinearAlgebra
+# using LinearAlgebra
+# using Gridap.Inference
 
-NewKernels.test_kernel(FunctionKernel(+),(3,2),5)
+test_kernel(FunctionKernel(+),(3,2),5)
 @test NewKernels.return_types((FunctionKernel(+),FunctionKernel(/)),1,1) == (Int,Float64)
 
 f = BroadcastKernel(+)
 a = rand(3,2)
 b = 3
 c = a .+ b
-test_kernel(f,(a,b),c)
-@test ==(z,y)
-@test typeof(z) == return_type(f,x...)
-isa(FunctionKernel,NewKernel)
+NewKernels.test_kernel(f,(a,b),c)
 
-@test kernel_return_types((+,/),1,1) == (Int,Float64)
 
-f = bcast(+)
-a = rand(3,2)
-b = 3
-c = a .+ b
-test_kernel(f,(a,b),c)
-
-#test_kernel(1,(),1)
-#test_kernel(1,(1,),1)
-#
-#test_kernel([1,2],(),[1,2])
-#test_kernel([1,2],(1,),[1,2])
-
-k = elem(-)
+k = BroadcastKernel(-)
 test_kernel(k,(1,),-1)
 test_kernel(k,([1,2],),[-1,-2])
 test_kernel(k,(1,2),-1)
@@ -42,10 +27,7 @@ test_kernel(k,([1,2],2),[-1,0])
 test_kernel(k,(2,[1,2]),[1,0])
 test_kernel(k,([3,4],[1,2]),[2,2])
 
-k = contract(-)
-test_kernel(k,([3,4],[1,2]),3-1+4-2)
-
-f = bcast(⋅)
+f = BroadcastKernel(⋅)
 a = fill(TensorValue(2,0,0,0,2,0,0,0,2),2)
 b = VectorValue(1,2,3)
 c = zeros(VectorValue{3,Int},2)
