+++ src/NewKernels/NewKernels.jl
@@ -3,6 +3,8 @@ module NewKernels
 using Gridap.Helpers
 using Gridap.Inference
 
+using Gridap.Arrays: CachedArray
+
 using Test
 
 export NewKernel
@@ -12,12 +14,14 @@ export BroadcastKernel
 export CompositionKernel
 
 export return_type
-export cache
+export return_cache
 export evaluate!
 export evaluate
 export test_kernel
-export caches
+export return_caches
 export return_types
+
+import Gridap.Inference: return_type
 
 include("Interface.jl")
 
