### 2020-08-26 bccd17a
Kernel gradients
