### 2020-08-26 984a102
Minor fixes for AD + BlockArrayCoo
