### 2020-08-26 #2520 @Kevin-Mattheus-Moerman
@fverdugo @santiagobadia congratulations on getting this nice work published! It was a pleasure to edit.

P.S. I have also personally started to explore this interesting software and am impressed by its speed/simplicity. It would be great if future work included the use of Julia based visualization, e.g. by using Makie for all meshing, boundary condition specification, and result visualizations.
