### 2020-08-26 #2520 @santiagobadia
Thanks @Kevin-Mattheus-Moerman for your work

> P.S. I have also personally started to explore this interesting software and am impressed by its speed/simplicity. It would be great if future work included the use of Julia based visualization, e.g. by using Makie for all meshing, boundary condition specification, and result visualizations.

Excellent!!!

Take a look at this repos wrt Makie + Gridap

https://github.com/gridap/GridapMakie.jl

@jw3126 is a new collaborator of the Gridap project that is doing an excellent job leading this dev
